Crate esp32p4

Peripheral access API for ESP32-P4 microcontrollers (generated using svd2rust v0.31.5 ( ))

You can find an overview of the generated API here.

API features to be included in the next svd2rust release can be generated by cloning the svd2rust repository, checking out the above commit, and running cargo doc --open.

Re-exports

pub use self::i2c0 as i2c1;

pub use self::i2s0 as i2s1;

pub use self::i2s0 as i2s2;

pub use self::mcpwm0 as mcpwm1;

pub use self::timg0 as timg1;

pub use self::trace0 as trace1;

pub use self::twai0 as twai1;

pub use self::twai0 as twai2;

pub use self::uart0 as uart1;

pub use self::uart0 as uart2;

pub use self::uart0 as uart3;

pub use self::uart0 as uart4;

Modules

adc

ADC (Analog to Digital Converter)

aes

AES (Advanced Encryption Standard) Accelerator

ahb_dma

AHB_DMA Peripheral

assist_debug

Debug Assist

axi_dma

AXI_DMA Peripheral

axi_icm

AXI_ICM Peripheral

bitscrambler

BITSCRAMBLER Peripheral

cache

CACHE Peripheral

dma

DMA (Direct Memory Access) Controller

ds

Digital Signature

ecc

ECC (ECC Hardware Accelerator)

ecdsa

ECDSA (Elliptic Curve Digital Signature Algorithm) Accelerator

efuse

eFuse Controller

generic

Common register and bit access and modify traits

gpio

General Purpose Input/Output

gpio_sd

Sigma-Delta Modulation

h264

H264 Encoder (Core)

h264_dma

H264 Encoder (DMA)

hmac

HMAC (Hash-based Message Authentication Code) Accelerator

hp_sys

High-Power System

hp_sys_clkrst

HP_SYS_CLKRST Peripheral

i2c0

I2C (Inter-Integrated Circuit) Controller 0

i2s0

I2S (Inter-IC Sound) Controller 0

i3c_mst

I3C Controller (Master)

i3c_mst_mem

I3C_MST_MEM Peripheral

i3c_slv

I3C Controller (Slave)

interrupt_core0

Interrupt Controller (Core 0)

interrupt_core1

Interrupt Controller (Core 1)

io_mux

Input/Output Multiplexer

isp

ISP Peripheral

jpeg

JPEG Codec

lcd_cam

Camera/LCD Controller

ledc

LED Control PWM (Pulse Width Modulation)

lp_adc

Low-power Analog to Digital Converter

lp_ana_peri

LP_ANA_PERI Peripheral

lp_aon_clkrst

LP_AON_CLKRST Peripheral

lp_gpio

Low-power General Purpose Input/Output

lp_huk

LP_HUK Peripheral

lp_i2c0

Low-power I2C (Inter-Integrated Circuit) Controller 0

lp_i2c_ana_mst

LP_I2C_ANA_MST Peripheral

lp_i2s0

Low-power I2S (Inter-IC Sound) Controller 0

lp_intr

Low-power Interrupt Controller

lp_io_mux

Low-power Input/Output Multiplexer

lp_peri

LP_PERI Peripheral

lp_sys

LP_SYS Peripheral

lp_timer

Low-power Timer

lp_touch

LP_TOUCH Peripheral

lp_tsens

Low-power Temperature Sensor

lp_uart

Low-power UART (Universal Asynchronous Receiver-Transmitter) Controller

lp_wdt

Low-power Watchdog Timer

mcpwm0

Motor Control Pulse-Width Modulation 0

mipi_csi_bridge

MIPI Camera Interface Bridge

mipi_csi_host

MIPI Camera Interface Host

mipi_dsi_bridge

MIPI Camera Interface Bridge

mipi_dsi_host

MIPI Display Interface Host

parl_io

Parallel IO Controller

pau

PAU Peripheral

pcnt

Pulse Count Controller

pmu

PMU Peripheral

ppa

PPA Peripheral

pvt

PVT Peripheral

rmt

Remote Control

rsa

RSA (Rivest Shamir Adleman) Accelerator

sdhost

SD/MMC Host Controller

sha

SHA (Secure Hash Algorithm) Accelerator

soc_etm

Event Task Matrix

spi0

SPI (Serial Peripheral Interface) Controller 0

spi1

SPI (Serial Peripheral Interface) Controller 1

spi2

SPI (Serial Peripheral Interface) Controller 2

spi3

SPI (Serial Peripheral Interface) Controller 3

systimer

System Timer

timg0

Timer Group 0

trace0

TRACE0 Peripheral

twai0

Two-Wire Automotive Interface

uart0

UART (Universal Asynchronous Receiver-Transmitter) Controller 0

uhci0

Universal Host Controller Interface 0

usb_device

Full-speed USB Serial/JTAG Controller

usb_wrap

USB_WRAP Peripheral

Macros

interrupt

Assigns a handler to an interrupt

Structs

ADC

ADC (Analog to Digital Converter)

AES

AES (Advanced Encryption Standard) Accelerator

AHB_DMA

AHB_DMA Peripheral

ASSIST_DEBUG

Debug Assist

AXI_DMA

AXI_DMA Peripheral

AXI_ICM

AXI_ICM Peripheral

BITSCRAMBLER

BITSCRAMBLER Peripheral

CACHE

CACHE Peripheral

DMA

DMA (Direct Memory Access) Controller

DS

Digital Signature

ECC

ECC (ECC Hardware Accelerator)

ECDSA

ECDSA (Elliptic Curve Digital Signature Algorithm) Accelerator

EFUSE

eFuse Controller

GPIO

General Purpose Input/Output

GPIO_SD

Sigma-Delta Modulation

H264

H264 Encoder (Core)

H264_DMA

H264 Encoder (DMA)

HMAC

HMAC (Hash-based Message Authentication Code) Accelerator

HP_SYS

High-Power System

HP_SYS_CLKRST

HP_SYS_CLKRST Peripheral

I2C0

I2C (Inter-Integrated Circuit) Controller 0

I2C1

I2C (Inter-Integrated Circuit) Controller 1

I2S0

I2S (Inter-IC Sound) Controller 0

I2S1

I2S (Inter-IC Sound) Controller 1

I2S2

I2S (Inter-IC Sound) Controller 2

I3C_MST

I3C Controller (Master)

I3C_MST_MEM

I3C_MST_MEM Peripheral

I3C_SLV

I3C Controller (Slave)

INTERRUPT_CORE0

Interrupt Controller (Core 0)

INTERRUPT_CORE1

Interrupt Controller (Core 1)

IO_MUX

Input/Output Multiplexer

ISP

ISP Peripheral

JPEG

JPEG Codec

LCD_CAM

Camera/LCD Controller

LEDC

LED Control PWM (Pulse Width Modulation)

LP_ADC

Low-power Analog to Digital Converter

LP_ANA_PERI

LP_ANA_PERI Peripheral

LP_AON_CLKRST

LP_AON_CLKRST Peripheral

LP_GPIO

Low-power General Purpose Input/Output

LP_HUK

LP_HUK Peripheral

LP_I2C0

Low-power I2C (Inter-Integrated Circuit) Controller 0

LP_I2C_ANA_MST

LP_I2C_ANA_MST Peripheral

LP_I2S0

Low-power I2S (Inter-IC Sound) Controller 0

LP_INTR

Low-power Interrupt Controller

LP_IO_MUX

Low-power Input/Output Multiplexer

LP_PERI

LP_PERI Peripheral

LP_SYS

LP_SYS Peripheral

LP_TIMER

Low-power Timer

LP_TOUCH

LP_TOUCH Peripheral

LP_TSENS

Low-power Temperature Sensor

LP_UART

Low-power UART (Universal Asynchronous Receiver-Transmitter) Controller

LP_WDT

Low-power Watchdog Timer

MCPWM0

Motor Control Pulse-Width Modulation 0

MCPWM1

Motor Control Pulse-Width Modulation 1

MIPI_CSI_BRIDGE

MIPI Camera Interface Bridge

MIPI_CSI_HOST

MIPI Camera Interface Host

MIPI_DSI_BRIDGE

MIPI Camera Interface Bridge

MIPI_DSI_HOST

MIPI Display Interface Host

PARL_IO

Parallel IO Controller

PAU

PAU Peripheral

PCNT

Pulse Count Controller

PMU

PMU Peripheral

PPA

PPA Peripheral

PVT

PVT Peripheral

Peripherals

All the peripherals.

RMT

Remote Control

RSA

RSA (Rivest Shamir Adleman) Accelerator

SDHOST

SD/MMC Host Controller

SHA

SHA (Secure Hash Algorithm) Accelerator

SOC_ETM

Event Task Matrix

SPI0

SPI (Serial Peripheral Interface) Controller 0

SPI1

SPI (Serial Peripheral Interface) Controller 1

SPI2

SPI (Serial Peripheral Interface) Controller 2

SPI3

SPI (Serial Peripheral Interface) Controller 3

SYSTIMER

System Timer

TIMG0

Timer Group 0

TIMG1

Timer Group 1

TRACE0

TRACE0 Peripheral

TRACE1

TRACE1 Peripheral

TWAI0

Two-Wire Automotive Interface

TWAI1

Two-Wire Automotive Interface

TWAI2

Two-Wire Automotive Interface

UART0

UART (Universal Asynchronous Receiver-Transmitter) Controller 0

UART1

UART (Universal Asynchronous Receiver-Transmitter) Controller 1

UART2

UART (Universal Asynchronous Receiver-Transmitter) Controller 2

UART3

UART (Universal Asynchronous Receiver-Transmitter) Controller 3

UART4

UART (Universal Asynchronous Receiver-Transmitter) Controller 4

UHCI0

Universal Host Controller Interface 0

USB_DEVICE

Full-speed USB Serial/JTAG Controller

USB_WRAP

USB_WRAP Peripheral

Enums

Interrupt

Enumeration of all the interrupts.

Constants

NVIC_PRIO_BITS

Number available in the NVIC for configuring priority