Module esp32p4::ahb_dma

AHB_DMA Peripheral

Modules

ahb_test

reserved

arb_timeout_rx

This retister is used to config arbiter time slice for rx dir

arb_timeout_tx

This retister is used to config arbiter time slice for tx dir

date

Version control register

in_conf0_ch

Configure 0 register of Rx channel 0

in_conf1_ch

Configure 1 register of Rx channel 0

in_crc_clear_ch

This register is used to clear ch0 crc result

in_crc_final_result_ch

This register is used to store ch0 crc result

in_crc_init_data_ch

This register is used to config ch0 crc initial data(max 32 bit)

in_dscr_bf0_ch

The last inlink descriptor address of Rx channel 0

in_dscr_bf1_ch

The second-to-last inlink descriptor address of Rx channel 0

in_dscr_ch

Current inlink descriptor address of Rx channel 0

in_err_eof_des_addr_ch

Inlink descriptor address when errors occur of Rx channel 0

in_int_clr_ch

Interrupt clear bits of channel 0

in_int_ena_ch

Interrupt enable bits of channel 0

in_int_raw_ch

Raw status interrupt of channel 0

in_int_st_ch

Masked interrupt of channel 0

in_link_addr_ch

Link descriptor configure of Rx channel 0

in_link_ch

Link descriptor configure and control register of Rx channel 0

in_peri_sel_ch

Peripheral selection of Rx channel 0

in_pop_ch

Pop control register of Rx channel 0

in_pri_ch

Priority register of Rx channel 0

in_state_ch

Receive status of Rx channel 0

in_suc_eof_des_addr_ch

Inlink descriptor address when EOF occurs of Rx channel 0

infifo_status_ch

Receive FIFO status of Rx channel 0

intr_mem_end_addr

The end address of accessible address space. The access address beyond this range would lead to descriptor error.

intr_mem_start_addr

The start address of accessible address space.

misc_conf

MISC register

out_conf0_ch

Configure 0 register of Tx channel 1

out_conf0_ch0

Configure 0 register of Tx channel 0

out_conf1_ch

Configure 1 register of Tx channel 0

out_crc_clear_ch

This register is used to clear ch0 crc result

out_crc_final_result_ch

This register is used to store ch0 crc result

out_crc_init_data_ch

This register is used to config ch0 crc initial data(max 32 bit)

out_dscr_bf0_ch

The last inlink descriptor address of Tx channel 0

out_dscr_bf1_ch

The second-to-last inlink descriptor address of Tx channel 0

out_dscr_ch

Current inlink descriptor address of Tx channel 0

out_eof_bfr_des_addr_ch

The last outlink descriptor address when EOF occurs of Tx channel 0

out_eof_des_addr_ch

Outlink descriptor address when EOF occurs of Tx channel 0

out_int_clr_ch

Interrupt clear bits of channel 0

out_int_ena_ch

Interrupt enable bits of channel 0

out_int_raw_ch

Raw status interrupt of channel 0

out_int_st_ch

Masked interrupt of channel 0

out_link_addr_ch

Link descriptor configure of Tx channel 0

out_link_ch

Link descriptor configure and control register of Tx channel 0

out_peri_sel_ch

Peripheral selection of Tx channel 0

out_pri_ch

Priority register of Tx channel 0.

out_push_ch

Push control register of Rx channel 0

out_state_ch

Transmit status of Tx channel 0

outfifo_status_ch

Transmit FIFO status of Tx channel 0

rx_arb_weigh_opt_dir_ch

This register is used to config off or on weigh optimization

rx_ch_arb_weigh_ch

This register is used to config ch0 arbiter weigh

rx_crc_data_en_addr_ch

This register is used to config addr of crc data_8bit en

rx_crc_data_en_wr_data_ch

This register is used to config crc data_8bit en

rx_crc_en_addr_ch

This register is used to config ch0 crc en addr

rx_crc_en_wr_data_ch

This resister is used to config ch0 crc en for every bit

rx_crc_width_ch

This register is used to confiig rx ch0 crc result width,2’b00 mean crc_width <=8bit,2’b01 8<crc_width<=16 ,2’b10 mean 16<crc_width <=24,2’b11 mean 24<crc_width<=32

tx_arb_weigh_opt_dir_ch

This register is used to config off or on weigh optimization

tx_ch_arb_weigh_ch

This register is used to config ch0 arbiter weigh

tx_crc_data_en_addr_ch

This register is used to config addr of crc data_8bit en

tx_crc_data_en_wr_data_ch

This register is used to config crc data_8bit en

tx_crc_en_addr_ch

This register is used to config ch0 crc en addr

tx_crc_en_wr_data_ch

This resister is used to config ch0 crc en for every bit

tx_crc_width_ch

This register is used to confiig tx ch0 crc result width,2’b00 mean crc_width <=8bit,2’b01 8<crc_width<=16 ,2’b10 mean 16<crc_width <=24,2’b11 mean 24<crc_width<=32

weight_en_rx

This register is used to config arbiter weigh function to on or off for rx dir

weight_en_tx

This register is used to config arbiter weigh function to on or off for tx dir

Structs

RegisterBlock

Register block

Type Aliases

AHB_TEST

AHB_TEST (rw) register accessor: reserved

ARB_TIMEOUT_RX

ARB_TIMEOUT_RX (rw) register accessor: This retister is used to config arbiter time slice for rx dir

ARB_TIMEOUT_TX

ARB_TIMEOUT_TX (rw) register accessor: This retister is used to config arbiter time slice for tx dir

DATE

DATE (rw) register accessor: Version control register

INFIFO_STATUS_CH

INFIFO_STATUS_CH (r) register accessor: Receive FIFO status of Rx channel 0

INTR_MEM_END_ADDR

INTR_MEM_END_ADDR (rw) register accessor: The end address of accessible address space. The access address beyond this range would lead to descriptor error.

INTR_MEM_START_ADDR

INTR_MEM_START_ADDR (rw) register accessor: The start address of accessible address space.

IN_CONF0_CH

IN_CONF0_CH (rw) register accessor: Configure 0 register of Rx channel 0

IN_CONF1_CH

IN_CONF1_CH (rw) register accessor: Configure 1 register of Rx channel 0

IN_CRC_CLEAR_CH

IN_CRC_CLEAR_CH (rw) register accessor: This register is used to clear ch0 crc result

IN_CRC_FINAL_RESULT_CH

IN_CRC_FINAL_RESULT_CH (r) register accessor: This register is used to store ch0 crc result

IN_CRC_INIT_DATA_CH

IN_CRC_INIT_DATA_CH (rw) register accessor: This register is used to config ch0 crc initial data(max 32 bit)

IN_DSCR_BF0_CH

IN_DSCR_BF0_CH (r) register accessor: The last inlink descriptor address of Rx channel 0

IN_DSCR_BF1_CH

IN_DSCR_BF1_CH (r) register accessor: The second-to-last inlink descriptor address of Rx channel 0

IN_DSCR_CH

IN_DSCR_CH (r) register accessor: Current inlink descriptor address of Rx channel 0

IN_ERR_EOF_DES_ADDR_CH

IN_ERR_EOF_DES_ADDR_CH (r) register accessor: Inlink descriptor address when errors occur of Rx channel 0

IN_INT_CLR_CH

IN_INT_CLR_CH (w) register accessor: Interrupt clear bits of channel 0

IN_INT_ENA_CH

IN_INT_ENA_CH (rw) register accessor: Interrupt enable bits of channel 0

IN_INT_RAW_CH

IN_INT_RAW_CH (rw) register accessor: Raw status interrupt of channel 0

IN_INT_ST_CH

IN_INT_ST_CH (r) register accessor: Masked interrupt of channel 0

IN_LINK_ADDR_CH

IN_LINK_ADDR_CH (rw) register accessor: Link descriptor configure of Rx channel 0

IN_LINK_CH

IN_LINK_CH (rw) register accessor: Link descriptor configure and control register of Rx channel 0

IN_PERI_SEL_CH

IN_PERI_SEL_CH (rw) register accessor: Peripheral selection of Rx channel 0

IN_POP_CH

IN_POP_CH (rw) register accessor: Pop control register of Rx channel 0

IN_PRI_CH

IN_PRI_CH (rw) register accessor: Priority register of Rx channel 0

IN_STATE_CH

IN_STATE_CH (r) register accessor: Receive status of Rx channel 0

IN_SUC_EOF_DES_ADDR_CH

IN_SUC_EOF_DES_ADDR_CH (r) register accessor: Inlink descriptor address when EOF occurs of Rx channel 0

MISC_CONF

MISC_CONF (rw) register accessor: MISC register

OUTFIFO_STATUS_CH

OUTFIFO_STATUS_CH (r) register accessor: Transmit FIFO status of Tx channel 0

OUT_CONF0_CH

OUT_CONF0_CH (rw) register accessor: Configure 0 register of Tx channel 1

OUT_CONF0_CH0

OUT_CONF0_CH0 (rw) register accessor: Configure 0 register of Tx channel 0

OUT_CONF1_CH

OUT_CONF1_CH (rw) register accessor: Configure 1 register of Tx channel 0

OUT_CRC_CLEAR_CH

OUT_CRC_CLEAR_CH (rw) register accessor: This register is used to clear ch0 crc result

OUT_CRC_FINAL_RESULT_CH

OUT_CRC_FINAL_RESULT_CH (r) register accessor: This register is used to store ch0 crc result

OUT_CRC_INIT_DATA_CH

OUT_CRC_INIT_DATA_CH (rw) register accessor: This register is used to config ch0 crc initial data(max 32 bit)

OUT_DSCR_BF0_CH

OUT_DSCR_BF0_CH (r) register accessor: The last inlink descriptor address of Tx channel 0

OUT_DSCR_BF1_CH

OUT_DSCR_BF1_CH (r) register accessor: The second-to-last inlink descriptor address of Tx channel 0

OUT_DSCR_CH

OUT_DSCR_CH (r) register accessor: Current inlink descriptor address of Tx channel 0

OUT_EOF_BFR_DES_ADDR_CH

OUT_EOF_BFR_DES_ADDR_CH (r) register accessor: The last outlink descriptor address when EOF occurs of Tx channel 0

OUT_EOF_DES_ADDR_CH

OUT_EOF_DES_ADDR_CH (r) register accessor: Outlink descriptor address when EOF occurs of Tx channel 0

OUT_INT_CLR_CH

OUT_INT_CLR_CH (w) register accessor: Interrupt clear bits of channel 0

OUT_INT_ENA_CH

OUT_INT_ENA_CH (rw) register accessor: Interrupt enable bits of channel 0

OUT_INT_RAW_CH

OUT_INT_RAW_CH (rw) register accessor: Raw status interrupt of channel 0

OUT_INT_ST_CH

OUT_INT_ST_CH (r) register accessor: Masked interrupt of channel 0

OUT_LINK_ADDR_CH

OUT_LINK_ADDR_CH (rw) register accessor: Link descriptor configure of Tx channel 0

OUT_LINK_CH

OUT_LINK_CH (rw) register accessor: Link descriptor configure and control register of Tx channel 0

OUT_PERI_SEL_CH

OUT_PERI_SEL_CH (rw) register accessor: Peripheral selection of Tx channel 0

OUT_PRI_CH

OUT_PRI_CH (rw) register accessor: Priority register of Tx channel 0.

OUT_PUSH_CH

OUT_PUSH_CH (rw) register accessor: Push control register of Rx channel 0

OUT_STATE_CH

OUT_STATE_CH (r) register accessor: Transmit status of Tx channel 0

RX_ARB_WEIGH_OPT_DIR_CH

RX_ARB_WEIGH_OPT_DIR_CH (rw) register accessor: This register is used to config off or on weigh optimization

RX_CH_ARB_WEIGH_CH

RX_CH_ARB_WEIGH_CH (rw) register accessor: This register is used to config ch0 arbiter weigh

RX_CRC_DATA_EN_ADDR_CH

RX_CRC_DATA_EN_ADDR_CH (rw) register accessor: This register is used to config addr of crc data_8bit en

RX_CRC_DATA_EN_WR_DATA_CH

RX_CRC_DATA_EN_WR_DATA_CH (rw) register accessor: This register is used to config crc data_8bit en

RX_CRC_EN_ADDR_CH

RX_CRC_EN_ADDR_CH (rw) register accessor: This register is used to config ch0 crc en addr

RX_CRC_EN_WR_DATA_CH

RX_CRC_EN_WR_DATA_CH (rw) register accessor: This resister is used to config ch0 crc en for every bit

RX_CRC_WIDTH_CH

RX_CRC_WIDTH_CH (rw) register accessor: This register is used to confiig rx ch0 crc result width,2’b00 mean crc_width <=8bit,2’b01 8<crc_width<=16 ,2’b10 mean 16<crc_width <=24,2’b11 mean 24<crc_width<=32

TX_ARB_WEIGH_OPT_DIR_CH

TX_ARB_WEIGH_OPT_DIR_CH (rw) register accessor: This register is used to config off or on weigh optimization

TX_CH_ARB_WEIGH_CH

TX_CH_ARB_WEIGH_CH (rw) register accessor: This register is used to config ch0 arbiter weigh

TX_CRC_DATA_EN_ADDR_CH

TX_CRC_DATA_EN_ADDR_CH (rw) register accessor: This register is used to config addr of crc data_8bit en

TX_CRC_DATA_EN_WR_DATA_CH

TX_CRC_DATA_EN_WR_DATA_CH (rw) register accessor: This register is used to config crc data_8bit en

TX_CRC_EN_ADDR_CH

TX_CRC_EN_ADDR_CH (rw) register accessor: This register is used to config ch0 crc en addr

TX_CRC_EN_WR_DATA_CH

TX_CRC_EN_WR_DATA_CH (rw) register accessor: This resister is used to config ch0 crc en for every bit

TX_CRC_WIDTH_CH

TX_CRC_WIDTH_CH (rw) register accessor: This register is used to confiig tx ch0 crc result width,2’b00 mean crc_width <=8bit,2’b01 8<crc_width<=16 ,2’b10 mean 16<crc_width <=24,2’b11 mean 24<crc_width<=32

WEIGHT_EN_RX

WEIGHT_EN_RX (rw) register accessor: This register is used to config arbiter weigh function to on or off for rx dir

WEIGHT_EN_TX

WEIGHT_EN_TX (rw) register accessor: This register is used to config arbiter weigh function to on or off for tx dir