Type Alias esp32p4::axi_dma::out_conf0_ch1::W

pub type W = W<OUT_CONF0_CH1_SPEC>;

Register OUT_CONF0_CH1 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn out_rst_ch1(&mut self) -> OUT_RST_CH1_W<'_, OUT_CONF0_CH1_SPEC>

Bit 0 - This bit is used to reset AXI_DMA channel1 Tx FSM and Tx FIFO pointer.

pub fn out_loop_test_ch1( &mut self ) -> OUT_LOOP_TEST_CH1_W<'_, OUT_CONF0_CH1_SPEC>

Bit 1 - reserved

pub fn out_auto_wrback_ch1( &mut self ) -> OUT_AUTO_WRBACK_CH1_W<'_, OUT_CONF0_CH1_SPEC>

Bit 2 - Set this bit to enable automatic outlink-writeback when all the data in tx buffer has been transmitted.

pub fn out_eof_mode_ch1(&mut self) -> OUT_EOF_MODE_CH1_W<'_, OUT_CONF0_CH1_SPEC>

Bit 3 - EOF flag generation mode when transmitting data. 1: EOF flag for Tx channel1 is generated when data need to transmit has been popped from FIFO in AXI_DMA

pub fn out_etm_en_ch1(&mut self) -> OUT_ETM_EN_CH1_W<'_, OUT_CONF0_CH1_SPEC>

Bit 4 - Set this bit to 1 to enable etm control mode, dma Tx channel1 is triggered by etm task.

pub fn out_burst_size_sel_ch1( &mut self ) -> OUT_BURST_SIZE_SEL_CH1_W<'_, OUT_CONF0_CH1_SPEC>

Bits 5:7 - 3’b000-3’b100:burst length 8byte~128byte

pub fn out_cmd_disable_ch1( &mut self ) -> OUT_CMD_DISABLE_CH1_W<'_, OUT_CONF0_CH1_SPEC>

Bit 8 - 1:mean disable cmd of this ch1

pub fn out_ecc_aec_en_ch1( &mut self ) -> OUT_ECC_AEC_EN_CH1_W<'_, OUT_CONF0_CH1_SPEC>

Bit 9 - 1: mean access ecc or aes domain,0: mean not

pub fn outdscr_burst_en_ch1( &mut self ) -> OUTDSCR_BURST_EN_CH1_W<'_, OUT_CONF0_CH1_SPEC>

Bit 10 - Set this bit to 1 to enable INCR burst transfer for Tx channel1 reading link descriptor when accessing internal SRAM.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual