Module esp32p4::bitscrambler::tx_state

Status registers

Structs

TX_STATE_SPEC

Status registers

Type Aliases

R

Register TX_STATE reader

TX_EOF_GET_CNT_R

Field TX_EOF_GET_CNT reader - represents the bytes numbers of bitscrambler tx core when get EOF

TX_EOF_OVERLOAD_R

Field TX_EOF_OVERLOAD reader - represents the some EOFs will be lost for bitscrambler tx core

TX_EOF_TRACE_CLR_W

Field TX_EOF_TRACE_CLR writer - write this bit to clear reg_bitscrambler_tx_eof_overload and reg_bitscrambler_tx_eof_get_cnt registers

TX_FIFO_EMPTY_R

Field TX_FIFO_EMPTY reader - represents the bitscrambler tx fifo in empty state

TX_IN_IDLE_R

Field TX_IN_IDLE reader - represents the bitscrambler tx core in halt mode

TX_IN_PAUSE_R

Field TX_IN_PAUSE reader - represents the bitscrambler tx core in pause mode

TX_IN_RUN_R

Field TX_IN_RUN reader - represents the bitscrambler tx core in run mode

TX_IN_WAIT_R

Field TX_IN_WAIT reader - represents the bitscrambler tx core in wait mode to wait write back done

W

Register TX_STATE writer