Type Alias esp32p4::bitscrambler::tx_state::R

pub type R = R<TX_STATE_SPEC>;

Register TX_STATE reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn tx_in_idle(&self) -> TX_IN_IDLE_R

Bit 0 - represents the bitscrambler tx core in halt mode

pub fn tx_in_run(&self) -> TX_IN_RUN_R

Bit 1 - represents the bitscrambler tx core in run mode

pub fn tx_in_wait(&self) -> TX_IN_WAIT_R

Bit 2 - represents the bitscrambler tx core in wait mode to wait write back done

pub fn tx_in_pause(&self) -> TX_IN_PAUSE_R

Bit 3 - represents the bitscrambler tx core in pause mode

pub fn tx_fifo_empty(&self) -> TX_FIFO_EMPTY_R

Bit 4 - represents the bitscrambler tx fifo in empty state

pub fn tx_eof_get_cnt(&self) -> TX_EOF_GET_CNT_R

Bits 16:29 - represents the bytes numbers of bitscrambler tx core when get EOF

pub fn tx_eof_overload(&self) -> TX_EOF_OVERLOAD_R

Bit 30 - represents the some EOFs will be lost for bitscrambler tx core