Module esp32p4::efuse::rd_repeat_data0

BLOCK0 data register 1.

Structs

RD_REPEAT_DATA0_SPEC

BLOCK0 data register 1.

Type Aliases

DIS_DOWNLOAD_MANUAL_ENCRYPT_R

Field DIS_DOWNLOAD_MANUAL_ENCRYPT reader - Represents whether flash encrypt function is disabled or enabled(except in SPI boot mode). 1: disabled. 0: enabled.

DIS_FORCE_DOWNLOAD_R

Field DIS_FORCE_DOWNLOAD reader - Represents whether the function that forces chip into download mode is disabled or enabled. 1: disabled. 0: enabled.

DIS_PAD_JTAG_R

Field DIS_PAD_JTAG reader - Represents whether JTAG is disabled in the hard way(permanently). 1: disabled. 0: enabled.

DIS_TWAI_R

Field DIS_TWAI reader - Represents whether TWAI function is disabled or enabled. 1: disabled. 0: enabled.

DIS_USB_JTAG_R

Field DIS_USB_JTAG reader - Represents whether the function of usb switch to jtag is disabled or enabled. 1: disabled. 0: enabled.

DIS_USB_SERIAL_JTAG_R

Field DIS_USB_SERIAL_JTAG reader - Represents whether USB-Serial-JTAG is disabled or enabled. 1: disabled. 0: enabled.

JTAG_SEL_ENABLE_R

Field JTAG_SEL_ENABLE reader - Represents whether the selection between usb_to_jtag and pad_to_jtag through strapping gpio15 when both EFUSE_DIS_PAD_JTAG and EFUSE_DIS_USB_JTAG are equal to 0 is enabled or disabled. 1: enabled. 0: disabled.

KM_HUK_GEN_STATE_LOW_R

Field KM_HUK_GEN_STATE_LOW reader - Set this bit to control validation of HUK generate mode. Odd of 1 is invalid, even of 1 is valid.

POWERGLITCH_EN_R

Field POWERGLITCH_EN reader - Represents whether power glitch function is enabled. 1: enabled. 0: disabled.

R

Register RD_REPEAT_DATA0 reader

RD_DIS_R

Field RD_DIS reader - Represents whether reading of individual eFuse block(block4~block10) is disabled or enabled. 1: disabled. 0: enabled.

SOFT_DIS_JTAG_R

Field SOFT_DIS_JTAG reader - Represents whether JTAG is disabled in soft way. Odd number: disabled. Even number: enabled.

SPI_DOWNLOAD_MSPI_DIS_R

Field SPI_DOWNLOAD_MSPI_DIS reader - Set this bit to disable accessing MSPI flash/MSPI ram by SYS AXI matrix during boot_mode_download.

USB_DEVICE_DREFH_R

Field USB_DEVICE_DREFH reader - USB intphy of usb device signle-end input high threshold, 1.76V to 2V. Step by 80mV

USB_DEVICE_EXCHG_PINS_R

Field USB_DEVICE_EXCHG_PINS reader - Enable usb device exchange pins of D+ and D-.

USB_OTG11_DREFH_R

Field USB_OTG11_DREFH reader - USB intphy of usb otg11 signle-end input high threshold, 1.76V to 2V. Step by 80mV

USB_OTG11_EXCHG_PINS_R

Field USB_OTG11_EXCHG_PINS reader - Enable usb otg11 exchange pins of D+ and D-.

USB_PHY_SEL_R

Field USB_PHY_SEL reader - TBD