Type Alias esp32p4::efuse::rd_repeat_data0::R

pub type R = R<RD_REPEAT_DATA0_SPEC>;

Register RD_REPEAT_DATA0 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn rd_dis(&self) -> RD_DIS_R

Bits 0:6 - Represents whether reading of individual eFuse block(block4~block10) is disabled or enabled. 1: disabled. 0: enabled.

pub fn usb_device_exchg_pins(&self) -> USB_DEVICE_EXCHG_PINS_R

Bit 7 - Enable usb device exchange pins of D+ and D-.

pub fn usb_otg11_exchg_pins(&self) -> USB_OTG11_EXCHG_PINS_R

Bit 8 - Enable usb otg11 exchange pins of D+ and D-.

pub fn dis_usb_jtag(&self) -> DIS_USB_JTAG_R

Bit 9 - Represents whether the function of usb switch to jtag is disabled or enabled. 1: disabled. 0: enabled.

pub fn powerglitch_en(&self) -> POWERGLITCH_EN_R

Bit 10 - Represents whether power glitch function is enabled. 1: enabled. 0: disabled.

pub fn dis_usb_serial_jtag(&self) -> DIS_USB_SERIAL_JTAG_R

Bit 11 - Represents whether USB-Serial-JTAG is disabled or enabled. 1: disabled. 0: enabled.

pub fn dis_force_download(&self) -> DIS_FORCE_DOWNLOAD_R

Bit 12 - Represents whether the function that forces chip into download mode is disabled or enabled. 1: disabled. 0: enabled.

pub fn spi_download_mspi_dis(&self) -> SPI_DOWNLOAD_MSPI_DIS_R

Bit 13 - Set this bit to disable accessing MSPI flash/MSPI ram by SYS AXI matrix during boot_mode_download.

pub fn dis_twai(&self) -> DIS_TWAI_R

Bit 14 - Represents whether TWAI function is disabled or enabled. 1: disabled. 0: enabled.

pub fn jtag_sel_enable(&self) -> JTAG_SEL_ENABLE_R

Bit 15 - Represents whether the selection between usb_to_jtag and pad_to_jtag through strapping gpio15 when both EFUSE_DIS_PAD_JTAG and EFUSE_DIS_USB_JTAG are equal to 0 is enabled or disabled. 1: enabled. 0: disabled.

pub fn soft_dis_jtag(&self) -> SOFT_DIS_JTAG_R

Bits 16:18 - Represents whether JTAG is disabled in soft way. Odd number: disabled. Even number: enabled.

pub fn dis_pad_jtag(&self) -> DIS_PAD_JTAG_R

Bit 19 - Represents whether JTAG is disabled in the hard way(permanently). 1: disabled. 0: enabled.

pub fn dis_download_manual_encrypt(&self) -> DIS_DOWNLOAD_MANUAL_ENCRYPT_R

Bit 20 - Represents whether flash encrypt function is disabled or enabled(except in SPI boot mode). 1: disabled. 0: enabled.

pub fn usb_device_drefh(&self) -> USB_DEVICE_DREFH_R

Bits 21:22 - USB intphy of usb device signle-end input high threshold, 1.76V to 2V. Step by 80mV

pub fn usb_otg11_drefh(&self) -> USB_OTG11_DREFH_R

Bits 23:24 - USB intphy of usb otg11 signle-end input high threshold, 1.76V to 2V. Step by 80mV

pub fn usb_phy_sel(&self) -> USB_PHY_SEL_R

Bit 25 - TBD

pub fn km_huk_gen_state_low(&self) -> KM_HUK_GEN_STATE_LOW_R

Bits 26:31 - Set this bit to control validation of HUK generate mode. Odd of 1 is invalid, even of 1 is valid.