Module esp32p4::efuse::rd_repeat_data1

BLOCK0 data register 2.

Structs

RD_REPEAT_DATA1_SPEC

BLOCK0 data register 2.

Type Aliases

FORCE_DISABLE_SW_INIT_KEY_R

Field FORCE_DISABLE_SW_INIT_KEY reader - Set this bit to disable software written init key, and force use efuse_init_key.

FORCE_USE_KEY_MANAGER_KEY_R

Field FORCE_USE_KEY_MANAGER_KEY reader - Set each bit to control whether corresponding key must come from key manager.. 1 is true, 0 is false. Bit0: ecdsa. Bit1: xts. Bit2: hmac. Bit3: ds.

KEY_PURPOSE_0_R

Field KEY_PURPOSE_0 reader - Represents the purpose of Key0.

KEY_PURPOSE_1_R

Field KEY_PURPOSE_1 reader - Represents the purpose of Key1.

KM_DEPLOY_ONLY_ONCE_R

Field KM_DEPLOY_ONLY_ONCE reader - Set each bit to control whether corresponding key can only be deployed once. 1 is true, 0 is false. Bit0: ecdsa. Bit1: xts. Bit2: hmac. Bit3: ds.

KM_HUK_GEN_STATE_HIGH_R

Field KM_HUK_GEN_STATE_HIGH reader - Set this bit to control validation of HUK generate mode. Odd of 1 is invalid, even of 1 is valid.

KM_RND_SWITCH_CYCLE_R

Field KM_RND_SWITCH_CYCLE reader - Set bits to control key manager random number switch cycle. 0: control by register. 1: 8 km clk cycles. 2: 16 km cycles. 3: 32 km cycles.

R

Register RD_REPEAT_DATA1 reader

SECURE_BOOT_KEY_REVOKE0_R

Field SECURE_BOOT_KEY_REVOKE0 reader - Represents whether revoking first secure boot key is enabled or disabled. 1: enabled. 0: disabled.

SECURE_BOOT_KEY_REVOKE1_R

Field SECURE_BOOT_KEY_REVOKE1 reader - Represents whether revoking second secure boot key is enabled or disabled. 1: enabled. 0: disabled.

SECURE_BOOT_KEY_REVOKE2_R

Field SECURE_BOOT_KEY_REVOKE2 reader - Represents whether revoking third secure boot key is enabled or disabled. 1: enabled. 0: disabled.

SPI_BOOT_CRYPT_CNT_R

Field SPI_BOOT_CRYPT_CNT reader - Represents whether SPI boot encrypt/decrypt is disabled or enabled. Odd number of 1: enabled. Even number of 1: disabled.

WDT_DELAY_SEL_R

Field WDT_DELAY_SEL reader - Represents whether RTC watchdog timeout threshold is selected at startup. 1: selected. 0: not selected.

XTS_KEY_LENGTH_256_R

Field XTS_KEY_LENGTH_256 reader - Set this bit to configure flash encryption use xts-128 key, else use xts-256 key.