Module esp32p4::efuse::rd_repeat_data2

BLOCK0 data register 3.

Structs

RD_REPEAT_DATA2_SPEC

BLOCK0 data register 3.

Type Aliases

CRYPT_DPA_ENABLE_R

Field CRYPT_DPA_ENABLE reader - Represents whether anti-dpa attack is enabled. 1:enabled. 0: disabled.

DIS_USB_OTG_DOWNLOAD_MODE_R

Field DIS_USB_OTG_DOWNLOAD_MODE reader - Set this bit to disable download via USB-OTG.

ECDSA_ENABLE_SOFT_K_R

Field ECDSA_ENABLE_SOFT_K reader - Represents whether hardware random number k is forced used in ESDCA. 1: force used. 0: not force used.

FLASH_ECC_EN_R

Field FLASH_ECC_EN reader - Set this bit to enable ecc for flash boot.

FLASH_PAGE_SIZE_R

Field FLASH_PAGE_SIZE reader - Set flash page size.

FLASH_TPUW_R

Field FLASH_TPUW reader - Represents the flash waiting time after power-up, in unit of ms. When the value less than 15, the waiting time is the programmed value. Otherwise, the waiting time is 2 times the programmed value.

FLASH_TYPE_R

Field FLASH_TYPE reader - The type of interfaced flash. 0: four data lines, 1: eight data lines.

KEY_PURPOSE_2_R

Field KEY_PURPOSE_2 reader - Represents the purpose of Key2.

KEY_PURPOSE_3_R

Field KEY_PURPOSE_3 reader - Represents the purpose of Key3.

KEY_PURPOSE_4_R

Field KEY_PURPOSE_4 reader - Represents the purpose of Key4.

KEY_PURPOSE_5_R

Field KEY_PURPOSE_5 reader - Represents the purpose of Key5.

R

Register RD_REPEAT_DATA2 reader

SECURE_BOOT_AGGRESSIVE_REVOKE_R

Field SECURE_BOOT_AGGRESSIVE_REVOKE reader - Represents whether revoking aggressive secure boot is enabled or disabled. 1: enabled. 0: disabled.

SECURE_BOOT_EN_R

Field SECURE_BOOT_EN reader - Represents whether secure boot is enabled or disabled. 1: enabled. 0: disabled.

SEC_DPA_LEVEL_R

Field SEC_DPA_LEVEL reader - Represents the spa secure level by configuring the clock random divide mode.