Type Alias esp32p4::efuse::rd_repeat_data2::R

pub type R = R<RD_REPEAT_DATA2_SPEC>;

Register RD_REPEAT_DATA2 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn key_purpose_2(&self) -> KEY_PURPOSE_2_R

Bits 0:3 - Represents the purpose of Key2.

pub fn key_purpose_3(&self) -> KEY_PURPOSE_3_R

Bits 4:7 - Represents the purpose of Key3.

pub fn key_purpose_4(&self) -> KEY_PURPOSE_4_R

Bits 8:11 - Represents the purpose of Key4.

pub fn key_purpose_5(&self) -> KEY_PURPOSE_5_R

Bits 12:15 - Represents the purpose of Key5.

pub fn sec_dpa_level(&self) -> SEC_DPA_LEVEL_R

Bits 16:17 - Represents the spa secure level by configuring the clock random divide mode.

pub fn ecdsa_enable_soft_k(&self) -> ECDSA_ENABLE_SOFT_K_R

Bit 18 - Represents whether hardware random number k is forced used in ESDCA. 1: force used. 0: not force used.

pub fn crypt_dpa_enable(&self) -> CRYPT_DPA_ENABLE_R

Bit 19 - Represents whether anti-dpa attack is enabled. 1:enabled. 0: disabled.

pub fn secure_boot_en(&self) -> SECURE_BOOT_EN_R

Bit 20 - Represents whether secure boot is enabled or disabled. 1: enabled. 0: disabled.

pub fn secure_boot_aggressive_revoke(&self) -> SECURE_BOOT_AGGRESSIVE_REVOKE_R

Bit 21 - Represents whether revoking aggressive secure boot is enabled or disabled. 1: enabled. 0: disabled.

pub fn flash_type(&self) -> FLASH_TYPE_R

Bit 23 - The type of interfaced flash. 0: four data lines, 1: eight data lines.

pub fn flash_page_size(&self) -> FLASH_PAGE_SIZE_R

Bits 24:25 - Set flash page size.

pub fn flash_ecc_en(&self) -> FLASH_ECC_EN_R

Bit 26 - Set this bit to enable ecc for flash boot.

pub fn dis_usb_otg_download_mode(&self) -> DIS_USB_OTG_DOWNLOAD_MODE_R

Bit 27 - Set this bit to disable download via USB-OTG.

pub fn flash_tpuw(&self) -> FLASH_TPUW_R

Bits 28:31 - Represents the flash waiting time after power-up, in unit of ms. When the value less than 15, the waiting time is the programmed value. Otherwise, the waiting time is 2 times the programmed value.