Type Alias esp32p4::efuse::rd_repeat_data3::R

pub type R = R<RD_REPEAT_DATA3_SPEC>;

Register RD_REPEAT_DATA3 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn dis_download_mode(&self) -> DIS_DOWNLOAD_MODE_R

Bit 0 - Represents whether Download mode is disabled or enabled. 1: disabled. 0: enabled.

pub fn dis_direct_boot(&self) -> DIS_DIRECT_BOOT_R

Bit 1 - Represents whether direct boot mode is disabled or enabled. 1: disabled. 0: enabled.

pub fn dis_usb_serial_jtag_rom_print(&self) -> DIS_USB_SERIAL_JTAG_ROM_PRINT_R

Bit 2 - Represents whether print from USB-Serial-JTAG is disabled or enabled. 1: disabled. 0: enabled.

pub fn lock_km_key(&self) -> LOCK_KM_KEY_R

Bit 3 - TBD

pub fn dis_usb_serial_jtag_download_mode( &self ) -> DIS_USB_SERIAL_JTAG_DOWNLOAD_MODE_R

Bit 4 - Represents whether the USB-Serial-JTAG download function is disabled or enabled. 1: disabled. 0: enabled.

pub fn enable_security_download(&self) -> ENABLE_SECURITY_DOWNLOAD_R

Bit 5 - Represents whether security download is enabled or disabled. 1: enabled. 0: disabled.

pub fn uart_print_control(&self) -> UART_PRINT_CONTROL_R

Bits 6:7 - Represents the type of UART printing. 00: force enable printing. 01: enable printing when GPIO8 is reset at low level. 10: enable printing when GPIO8 is reset at high level. 11: force disable printing.

pub fn force_send_resume(&self) -> FORCE_SEND_RESUME_R

Bit 8 - Represents whether ROM code is forced to send a resume command during SPI boot. 1: forced. 0:not forced.

pub fn secure_version(&self) -> SECURE_VERSION_R

Bits 9:24 - Represents the version used by ESP-IDF anti-rollback feature.

pub fn secure_boot_disable_fast_wake(&self) -> SECURE_BOOT_DISABLE_FAST_WAKE_R

Bit 25 - Represents whether FAST VERIFY ON WAKE is disabled or enabled when Secure Boot is enabled. 1: disabled. 0: enabled.

pub fn hys_en_pad(&self) -> HYS_EN_PAD_R

Bit 26 - Represents whether the hysteresis function of corresponding PAD is enabled. 1: enabled. 0:disabled.

pub fn dcdc_vset(&self) -> DCDC_VSET_R

Bits 27:31 - Set the dcdc voltage default.