Module esp32p4::h264::conf

General configuration register.

Structs

CONF_SPEC

General configuration register.

Type Aliases

BS_CLK_EN_R

Field BS_CLK_EN reader - Configures whether or not to open the clock gate for bs buffer.\0: Open the clock gate only when bs buffer work\1: Force open the clock gate for bs buffer

BS_CLK_EN_W

Field BS_CLK_EN writer - Configures whether or not to open the clock gate for bs buffer.\0: Open the clock gate only when bs buffer work\1: Force open the clock gate for bs buffer

CAVLC_RAM_CLK_EN_R

Field CAVLC_RAM_CLK_EN reader - Configures whether or not to open the clock gate for cavlc ram.\0: Open the clock gate only when application writes or reads cavlc ram\1: Force open the clock gate for cavlc ram

CAVLC_RAM_CLK_EN_W

Field CAVLC_RAM_CLK_EN writer - Configures whether or not to open the clock gate for cavlc ram.\0: Open the clock gate only when application writes or reads cavlc ram\1: Force open the clock gate for cavlc ram

CLAVLC_CLK_EN_R

Field CLAVLC_CLK_EN reader - Configures whether or not to open the clock gate for cavlc.\0: Open the clock gate only when cavlc work\1: Force open the clock gate for cavlc

CLAVLC_CLK_EN_W

Field CLAVLC_CLK_EN writer - Configures whether or not to open the clock gate for cavlc.\0: Open the clock gate only when cavlc work\1: Force open the clock gate for cavlc

CLK_EN_R

Field CLK_EN reader - Configures whether or not to open register clock gate.\0: Open the clock gate only when application writes registers\1: Force open the clock gate for register

CLK_EN_W

Field CLK_EN writer - Configures whether or not to open register clock gate.\0: Open the clock gate only when application writes registers\1: Force open the clock gate for register

CUR_MB_RAM_CLK_EN_R

Field CUR_MB_RAM_CLK_EN reader - Configures whether or not to open the clock gate for cur_mb ram.\0: Open the clock gate only when application writes or reads cur_mb ram\1: Force open the clock gate for cur_mb ram

CUR_MB_RAM_CLK_EN_W

Field CUR_MB_RAM_CLK_EN writer - Configures whether or not to open the clock gate for cur_mb ram.\0: Open the clock gate only when application writes or reads cur_mb ram\1: Force open the clock gate for cur_mb ram

DB_CLK_EN_R

Field DB_CLK_EN reader - Configures whether or not to open the clock gate for deblocking filter.\0: Open the clock gate only when deblocking filter work\1: Force open the clock gate for deblocking filter

DB_CLK_EN_W

Field DB_CLK_EN writer - Configures whether or not to open the clock gate for deblocking filter.\0: Open the clock gate only when deblocking filter work\1: Force open the clock gate for deblocking filter

DB_RAM_CLK_EN_R

Field DB_RAM_CLK_EN reader - Configures whether or not to open the clock gate for db ram.\0: Open the clock gate only when application writes or reads db ram\1: Force open the clock gate for db ram

DB_RAM_CLK_EN_W

Field DB_RAM_CLK_EN writer - Configures whether or not to open the clock gate for db ram.\0: Open the clock gate only when application writes or reads db ram\1: Force open the clock gate for db ram

DECI_CLK_EN_R

Field DECI_CLK_EN reader - Configures whether or not to open the clock gate for decimate.\0: Open the clock gate only when decimate work\1: Force open the clock gate for decimate

DECI_CLK_EN_W

Field DECI_CLK_EN writer - Configures whether or not to open the clock gate for decimate.\0: Open the clock gate only when decimate work\1: Force open the clock gate for decimate

FETCH_RAM_CLK_EN_R

Field FETCH_RAM_CLK_EN reader - Configures whether or not to open the clock gate for fetch ram.\0: Open the clock gate only when application writes or reads fetch ram\1: Force open the clock gate for fetch ram

FETCH_RAM_CLK_EN_W

Field FETCH_RAM_CLK_EN writer - Configures whether or not to open the clock gate for fetch ram.\0: Open the clock gate only when application writes or reads fetch ram\1: Force open the clock gate for fetch ram

FME_CLK_EN_R

Field FME_CLK_EN reader - Configures whether or not to open the clock gate for fme.\0: Open the clock gate only when fme work\1: Force open the clock gate for fme

FME_CLK_EN_W

Field FME_CLK_EN writer - Configures whether or not to open the clock gate for fme.\0: Open the clock gate only when fme work\1: Force open the clock gate for fme

FME_RAM_CLK_EN_R

Field FME_RAM_CLK_EN reader - Configures whether or not to open the clock gate for fme ram.\0: Open the clock gate only when application writes or readsfme ram\1: Force open the clock gate for fme ram

FME_RAM_CLK_EN_W

Field FME_RAM_CLK_EN writer - Configures whether or not to open the clock gate for fme ram.\0: Open the clock gate only when application writes or readsfme ram\1: Force open the clock gate for fme ram

I4X4_REF_RAM_CLK_EN_R

Field I4X4_REF_RAM_CLK_EN reader - Configures whether or not to open the clock gate for i4x4_mode ram.\0: Open the clock gate only when application writes or reads i4x4_mode ram\1: Force open the clock gate for i4x4_mode ram

I4X4_REF_RAM_CLK_EN_W

Field I4X4_REF_RAM_CLK_EN writer - Configures whether or not to open the clock gate for i4x4_mode ram.\0: Open the clock gate only when application writes or reads i4x4_mode ram\1: Force open the clock gate for i4x4_mode ram

IME_CLK_EN_R

Field IME_CLK_EN reader - Configures whether or not to open the clock gate for ime.\0: Open the clock gate only when ime work\1: Force open the clock gate for ime

IME_CLK_EN_W

Field IME_CLK_EN writer - Configures whether or not to open the clock gate for ime.\0: Open the clock gate only when ime work\1: Force open the clock gate for ime

IME_RAM_CLK_EN_R

Field IME_RAM_CLK_EN reader - Configures whether or not to open the clock gate for ime ram.\0: Open the clock gate only when application writes or reads ime ram\1: Force open the clock gate for ime ram

IME_RAM_CLK_EN_W

Field IME_RAM_CLK_EN writer - Configures whether or not to open the clock gate for ime ram.\0: Open the clock gate only when application writes or reads ime ram\1: Force open the clock gate for ime ram

INTERPOLATOR_CLK_EN_R

Field INTERPOLATOR_CLK_EN reader - Configures whether or not to open the clock gate for interpolator.\0: Open the clock gate only when interpolator work\1: Force open the clock gate for interpolator

INTERPOLATOR_CLK_EN_W

Field INTERPOLATOR_CLK_EN writer - Configures whether or not to open the clock gate for interpolator.\0: Open the clock gate only when interpolator work\1: Force open the clock gate for interpolator

INTRA_CLK_EN_R

Field INTRA_CLK_EN reader - Configures whether or not to open the clock gate for intra.\0: Open the clock gate only when intra work\1: Force open the clock gate for intra

INTRA_CLK_EN_W

Field INTRA_CLK_EN writer - Configures whether or not to open the clock gate for intra.\0: Open the clock gate only when intra work\1: Force open the clock gate for intra

MC_CLK_EN_R

Field MC_CLK_EN reader - Configures whether or not to open the clock gate for mc.\0: Open the clock gate only when mc work\1: Force open the clock gate for mc

MC_CLK_EN_W

Field MC_CLK_EN writer - Configures whether or not to open the clock gate for mc.\0: Open the clock gate only when mc work\1: Force open the clock gate for mc

MC_RAM_CLK_EN_R

Field MC_RAM_CLK_EN reader - Configures whether or not to open the clock gate for mc ram.\0: Open the clock gate only when application writes or reads mc ram\1: Force open the clock gate for mc ram

MC_RAM_CLK_EN_W

Field MC_RAM_CLK_EN writer - Configures whether or not to open the clock gate for mc ram.\0: Open the clock gate only when application writes or reads mc ram\1: Force open the clock gate for mc ram

MVD_RAM_CLK_EN_R

Field MVD_RAM_CLK_EN reader - Configures whether or not to open the clock gate for mvd ram.\0: Open the clock gate only when application writes or reads mvd ram\1: Force open the clock gate for mvd ram

MVD_RAM_CLK_EN_W

Field MVD_RAM_CLK_EN writer - Configures whether or not to open the clock gate for mvd ram.\0: Open the clock gate only when application writes or reads mvd ram\1: Force open the clock gate for mvd ram

MV_MERGE_CLK_EN_R

Field MV_MERGE_CLK_EN reader - Configures whether or not to open the clock gate for mv merge.\0: Open the clock gate only when mv merge work\1: Force open the clock gate for mv merge

MV_MERGE_CLK_EN_W

Field MV_MERGE_CLK_EN writer - Configures whether or not to open the clock gate for mv merge.\0: Open the clock gate only when mv merge work\1: Force open the clock gate for mv merge

PRE_RAM_CLK_EN_R

Field PRE_RAM_CLK_EN reader - Configures whether or not to open the clock gate for pre ram.\0: Open the clock gate only when application writes or reads pre ram\1: Force open the clock gate for pre ram

PRE_RAM_CLK_EN_W

Field PRE_RAM_CLK_EN writer - Configures whether or not to open the clock gate for pre ram.\0: Open the clock gate only when application writes or reads pre ram\1: Force open the clock gate for pre ram

QUANT_RAM_CLK_EN1_R

Field QUANT_RAM_CLK_EN1 reader - Configures whether or not to open the clock gate for quant ram1.\0: Open the clock gate only when application writes or reads quant ram1\1: Force open the clock gate for quant ram1

QUANT_RAM_CLK_EN1_W

Field QUANT_RAM_CLK_EN1 writer - Configures whether or not to open the clock gate for quant ram1.\0: Open the clock gate only when application writes or reads quant ram1\1: Force open the clock gate for quant ram1

QUANT_RAM_CLK_EN2_R

Field QUANT_RAM_CLK_EN2 reader - Configures whether or not to open the clock gate for quant ram2.\0: Open the clock gate only when application writes or reads quant ram2\1: Force open the clock gate for quant ram2

QUANT_RAM_CLK_EN2_W

Field QUANT_RAM_CLK_EN2 writer - Configures whether or not to open the clock gate for quant ram2.\0: Open the clock gate only when application writes or reads quant ram2\1: Force open the clock gate for quant ram2

R

Register CONF reader

REC_RAM_CLK_EN1_R

Field REC_RAM_CLK_EN1 reader - Configures whether or not to open the clock gate for rec ram1.\0: Open the clock gate only when application writes or reads rec ram1\1: Force open the clock gate for rec ram1

REC_RAM_CLK_EN1_W

Field REC_RAM_CLK_EN1 writer - Configures whether or not to open the clock gate for rec ram1.\0: Open the clock gate only when application writes or reads rec ram1\1: Force open the clock gate for rec ram1

REC_RAM_CLK_EN2_R

Field REC_RAM_CLK_EN2 reader - Configures whether or not to open the clock gate for rec ram2.\0: Open the clock gate only when application writes or reads rec ram2\1: Force open the clock gate for rec ram2

REC_RAM_CLK_EN2_W

Field REC_RAM_CLK_EN2 writer - Configures whether or not to open the clock gate for rec ram2.\0: Open the clock gate only when application writes or reads rec ram2\1: Force open the clock gate for rec ram2

REF_RAM_CLK_EN_R

Field REF_RAM_CLK_EN reader - Configures whether or not to open the clock gate for ref ram.\0: Open the clock gate only when application writes or reads ref ram\1: Force open the clock gate for ref ram

REF_RAM_CLK_EN_W

Field REF_RAM_CLK_EN writer - Configures whether or not to open the clock gate for ref ram.\0: Open the clock gate only when application writes or reads ref ram\1: Force open the clock gate for ref ram

W

Register CONF writer