Type Alias esp32p4::h264::conf::W

pub type W = W<CONF_SPEC>;

Register CONF writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn clk_en(&mut self) -> CLK_EN_W<'_, CONF_SPEC>

Bit 0 - Configures whether or not to open register clock gate.\0: Open the clock gate only when application writes registers\1: Force open the clock gate for register

pub fn rec_ram_clk_en2(&mut self) -> REC_RAM_CLK_EN2_W<'_, CONF_SPEC>

Bit 1 - Configures whether or not to open the clock gate for rec ram2.\0: Open the clock gate only when application writes or reads rec ram2\1: Force open the clock gate for rec ram2

pub fn rec_ram_clk_en1(&mut self) -> REC_RAM_CLK_EN1_W<'_, CONF_SPEC>

Bit 2 - Configures whether or not to open the clock gate for rec ram1.\0: Open the clock gate only when application writes or reads rec ram1\1: Force open the clock gate for rec ram1

pub fn quant_ram_clk_en2(&mut self) -> QUANT_RAM_CLK_EN2_W<'_, CONF_SPEC>

Bit 3 - Configures whether or not to open the clock gate for quant ram2.\0: Open the clock gate only when application writes or reads quant ram2\1: Force open the clock gate for quant ram2

pub fn quant_ram_clk_en1(&mut self) -> QUANT_RAM_CLK_EN1_W<'_, CONF_SPEC>

Bit 4 - Configures whether or not to open the clock gate for quant ram1.\0: Open the clock gate only when application writes or reads quant ram1\1: Force open the clock gate for quant ram1

pub fn pre_ram_clk_en(&mut self) -> PRE_RAM_CLK_EN_W<'_, CONF_SPEC>

Bit 5 - Configures whether or not to open the clock gate for pre ram.\0: Open the clock gate only when application writes or reads pre ram\1: Force open the clock gate for pre ram

pub fn mvd_ram_clk_en(&mut self) -> MVD_RAM_CLK_EN_W<'_, CONF_SPEC>

Bit 6 - Configures whether or not to open the clock gate for mvd ram.\0: Open the clock gate only when application writes or reads mvd ram\1: Force open the clock gate for mvd ram

pub fn mc_ram_clk_en(&mut self) -> MC_RAM_CLK_EN_W<'_, CONF_SPEC>

Bit 7 - Configures whether or not to open the clock gate for mc ram.\0: Open the clock gate only when application writes or reads mc ram\1: Force open the clock gate for mc ram

pub fn ref_ram_clk_en(&mut self) -> REF_RAM_CLK_EN_W<'_, CONF_SPEC>

Bit 8 - Configures whether or not to open the clock gate for ref ram.\0: Open the clock gate only when application writes or reads ref ram\1: Force open the clock gate for ref ram

pub fn i4x4_ref_ram_clk_en(&mut self) -> I4X4_REF_RAM_CLK_EN_W<'_, CONF_SPEC>

Bit 9 - Configures whether or not to open the clock gate for i4x4_mode ram.\0: Open the clock gate only when application writes or reads i4x4_mode ram\1: Force open the clock gate for i4x4_mode ram

pub fn ime_ram_clk_en(&mut self) -> IME_RAM_CLK_EN_W<'_, CONF_SPEC>

Bit 10 - Configures whether or not to open the clock gate for ime ram.\0: Open the clock gate only when application writes or reads ime ram\1: Force open the clock gate for ime ram

pub fn fme_ram_clk_en(&mut self) -> FME_RAM_CLK_EN_W<'_, CONF_SPEC>

Bit 11 - Configures whether or not to open the clock gate for fme ram.\0: Open the clock gate only when application writes or readsfme ram\1: Force open the clock gate for fme ram

pub fn fetch_ram_clk_en(&mut self) -> FETCH_RAM_CLK_EN_W<'_, CONF_SPEC>

Bit 12 - Configures whether or not to open the clock gate for fetch ram.\0: Open the clock gate only when application writes or reads fetch ram\1: Force open the clock gate for fetch ram

pub fn db_ram_clk_en(&mut self) -> DB_RAM_CLK_EN_W<'_, CONF_SPEC>

Bit 13 - Configures whether or not to open the clock gate for db ram.\0: Open the clock gate only when application writes or reads db ram\1: Force open the clock gate for db ram

pub fn cur_mb_ram_clk_en(&mut self) -> CUR_MB_RAM_CLK_EN_W<'_, CONF_SPEC>

Bit 14 - Configures whether or not to open the clock gate for cur_mb ram.\0: Open the clock gate only when application writes or reads cur_mb ram\1: Force open the clock gate for cur_mb ram

pub fn cavlc_ram_clk_en(&mut self) -> CAVLC_RAM_CLK_EN_W<'_, CONF_SPEC>

Bit 15 - Configures whether or not to open the clock gate for cavlc ram.\0: Open the clock gate only when application writes or reads cavlc ram\1: Force open the clock gate for cavlc ram

pub fn ime_clk_en(&mut self) -> IME_CLK_EN_W<'_, CONF_SPEC>

Bit 16 - Configures whether or not to open the clock gate for ime.\0: Open the clock gate only when ime work\1: Force open the clock gate for ime

pub fn fme_clk_en(&mut self) -> FME_CLK_EN_W<'_, CONF_SPEC>

Bit 17 - Configures whether or not to open the clock gate for fme.\0: Open the clock gate only when fme work\1: Force open the clock gate for fme

pub fn mc_clk_en(&mut self) -> MC_CLK_EN_W<'_, CONF_SPEC>

Bit 18 - Configures whether or not to open the clock gate for mc.\0: Open the clock gate only when mc work\1: Force open the clock gate for mc

pub fn interpolator_clk_en(&mut self) -> INTERPOLATOR_CLK_EN_W<'_, CONF_SPEC>

Bit 19 - Configures whether or not to open the clock gate for interpolator.\0: Open the clock gate only when interpolator work\1: Force open the clock gate for interpolator

pub fn db_clk_en(&mut self) -> DB_CLK_EN_W<'_, CONF_SPEC>

Bit 20 - Configures whether or not to open the clock gate for deblocking filter.\0: Open the clock gate only when deblocking filter work\1: Force open the clock gate for deblocking filter

pub fn clavlc_clk_en(&mut self) -> CLAVLC_CLK_EN_W<'_, CONF_SPEC>

Bit 21 - Configures whether or not to open the clock gate for cavlc.\0: Open the clock gate only when cavlc work\1: Force open the clock gate for cavlc

pub fn intra_clk_en(&mut self) -> INTRA_CLK_EN_W<'_, CONF_SPEC>

Bit 22 - Configures whether or not to open the clock gate for intra.\0: Open the clock gate only when intra work\1: Force open the clock gate for intra

pub fn deci_clk_en(&mut self) -> DECI_CLK_EN_W<'_, CONF_SPEC>

Bit 23 - Configures whether or not to open the clock gate for decimate.\0: Open the clock gate only when decimate work\1: Force open the clock gate for decimate

pub fn bs_clk_en(&mut self) -> BS_CLK_EN_W<'_, CONF_SPEC>

Bit 24 - Configures whether or not to open the clock gate for bs buffer.\0: Open the clock gate only when bs buffer work\1: Force open the clock gate for bs buffer

pub fn mv_merge_clk_en(&mut self) -> MV_MERGE_CLK_EN_W<'_, CONF_SPEC>

Bit 25 - Configures whether or not to open the clock gate for mv merge.\0: Open the clock gate only when mv merge work\1: Force open the clock gate for mv merge

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual