Type Alias esp32p4::h264_dma::in_conf0_ch3::W

pub type W = W<IN_CONF0_CH3_SPEC>;

Register IN_CONF0_CH3 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn indscr_burst_en_ch3( &mut self ) -> INDSCR_BURST_EN_CH3_W<'_, IN_CONF0_CH3_SPEC>

Bit 2 - Set this bit to 1 to enable INCR burst transfer for Rx transmitting link descriptor when accessing SRAM.

pub fn in_ecc_aes_en_ch3( &mut self ) -> IN_ECC_AES_EN_CH3_W<'_, IN_CONF0_CH3_SPEC>

Bit 3 - When access address space is ecc/aes area, this bit should be set to 1. In this case, the start address of square should be 16-bit aligned. The width of square multiply byte number of one pixel should be 16-bit aligned.

pub fn in_check_owner_ch3( &mut self ) -> IN_CHECK_OWNER_CH3_W<'_, IN_CONF0_CH3_SPEC>

Bit 4 - Set this bit to enable checking the owner attribute of the link descriptor.

pub fn in_mem_burst_length_ch3( &mut self ) -> IN_MEM_BURST_LENGTH_CH3_W<'_, IN_CONF0_CH3_SPEC>

Bits 6:8 - Block size of Rx channel 1. 0: single 1: 16 bytes 2: 32 bytes 3: 64 bytes 4: 128 bytes

pub fn in_page_bound_en_ch3( &mut self ) -> IN_PAGE_BOUND_EN_CH3_W<'_, IN_CONF0_CH3_SPEC>

Bit 12 - Set this bit to 1 to make sure AXI write data don’t cross the address boundary which define by mem_burst_length

pub fn in_rst_ch3(&mut self) -> IN_RST_CH3_W<'_, IN_CONF0_CH3_SPEC>

Bit 24 - Write 1 then write 0 to this bit to reset Rx channel

pub fn in_cmd_disable_ch3( &mut self ) -> IN_CMD_DISABLE_CH3_W<'_, IN_CONF0_CH3_SPEC>

Bit 25 - Write 1 before reset and write 0 after reset

pub fn in_arb_weight_opt_dis_ch3( &mut self ) -> IN_ARB_WEIGHT_OPT_DIS_CH3_W<'_, IN_CONF0_CH3_SPEC>

Bit 26 - Set this bit to 1 to disable arbiter optimum weight function.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual