Module esp32p4::h264_dma::out_conf0_ch4

TX CH4 config0 register

Structs

OUT_CONF0_CH4_SPEC

TX CH4 config0 register

Type Aliases

OUTDSCR_BURST_EN_CH4_R

Field OUTDSCR_BURST_EN_CH4 reader - Set this bit to 1 to enable INCR burst transfer for Tx channel 0 reading link descriptor when accessing internal SRAM.

OUTDSCR_BURST_EN_CH4_W

Field OUTDSCR_BURST_EN_CH4 writer - Set this bit to 1 to enable INCR burst transfer for Tx channel 0 reading link descriptor when accessing internal SRAM.

OUT_ARB_WEIGHT_OPT_DIS_CH4_R

Field OUT_ARB_WEIGHT_OPT_DIS_CH4 reader - Set this bit to 1 to disable arbiter optimum weight function.

OUT_ARB_WEIGHT_OPT_DIS_CH4_W

Field OUT_ARB_WEIGHT_OPT_DIS_CH4 writer - Set this bit to 1 to disable arbiter optimum weight function.

OUT_AUTO_WRBACK_CH4_R

Field OUT_AUTO_WRBACK_CH4 reader - Set this bit to enable automatic outlink-writeback when all the data pointed by outlink descriptor has been received.

OUT_AUTO_WRBACK_CH4_W

Field OUT_AUTO_WRBACK_CH4 writer - Set this bit to enable automatic outlink-writeback when all the data pointed by outlink descriptor has been received.

OUT_CHECK_OWNER_CH4_R

Field OUT_CHECK_OWNER_CH4 reader - Set this bit to enable checking the owner attribute of the link descriptor.

OUT_CHECK_OWNER_CH4_W

Field OUT_CHECK_OWNER_CH4 writer - Set this bit to enable checking the owner attribute of the link descriptor.

OUT_ECC_AES_EN_CH4_R

Field OUT_ECC_AES_EN_CH4 reader - When access address space is ecc/aes area, this bit should be set to 1. In this case, the start address of square should be 16-bit aligned. The width of square multiply byte number of one pixel should be 16-bit aligned.

OUT_ECC_AES_EN_CH4_W

Field OUT_ECC_AES_EN_CH4 writer - When access address space is ecc/aes area, this bit should be set to 1. In this case, the start address of square should be 16-bit aligned. The width of square multiply byte number of one pixel should be 16-bit aligned.

OUT_EOF_MODE_CH4_R

Field OUT_EOF_MODE_CH4 reader - EOF flag generation mode when receiving data. 1: EOF flag for Tx channel 0 is generated when data need to read has been popped from FIFO in DMA

OUT_EOF_MODE_CH4_W

Field OUT_EOF_MODE_CH4 writer - EOF flag generation mode when receiving data. 1: EOF flag for Tx channel 0 is generated when data need to read has been popped from FIFO in DMA

OUT_MEM_BURST_LENGTH_CH4_R

Field OUT_MEM_BURST_LENGTH_CH4 reader - Block size of Tx channel 4. 0: single 1: 16 bytes 2: 32 bytes 3: 64 bytes 4: 128 bytes

OUT_MEM_BURST_LENGTH_CH4_W

Field OUT_MEM_BURST_LENGTH_CH4 writer - Block size of Tx channel 4. 0: single 1: 16 bytes 2: 32 bytes 3: 64 bytes 4: 128 bytes

OUT_PAGE_BOUND_EN_CH4_R

Field OUT_PAGE_BOUND_EN_CH4 reader - Set this bit to 1 to make sure AXI read data don’t cross the address boundary which define by mem_burst_length

OUT_PAGE_BOUND_EN_CH4_W

Field OUT_PAGE_BOUND_EN_CH4 writer - Set this bit to 1 to make sure AXI read data don’t cross the address boundary which define by mem_burst_length

R

Register OUT_CONF0_CH4 reader

W

Register OUT_CONF0_CH4 writer