Type Alias esp32p4::h264_dma::out_int_raw_ch3::W

pub type W = W<OUT_INT_RAW_CH3_SPEC>;

Register OUT_INT_RAW_CH3 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn out_done_ch3_int_raw( &mut self ) -> OUT_DONE_CH3_INT_RAW_W<'_, OUT_INT_RAW_CH3_SPEC>

Bit 0 - The raw interrupt bit turns to high level when the last data pointed by one outlink descriptor has been transmitted to peripherals for Tx channel 0.

pub fn out_eof_ch3_int_raw( &mut self ) -> OUT_EOF_CH3_INT_RAW_W<'_, OUT_INT_RAW_CH3_SPEC>

Bit 1 - The raw interrupt bit turns to high level when the last data pointed by one outlink descriptor has been read from memory for Tx channel 0.

pub fn out_dscr_err_ch3_int_raw( &mut self ) -> OUT_DSCR_ERR_CH3_INT_RAW_W<'_, OUT_INT_RAW_CH3_SPEC>

Bit 2 - The raw interrupt bit turns to high level when detecting outlink descriptor error, including owner error, the second and third word error of outlink descriptor for Tx channel 0.

pub fn out_total_eof_ch3_int_raw( &mut self ) -> OUT_TOTAL_EOF_CH3_INT_RAW_W<'_, OUT_INT_RAW_CH3_SPEC>

Bit 3 - The raw interrupt bit turns to high level when data corresponding a outlink (includes one link descriptor or few link descriptors) is transmitted out for Tx channel 0.

pub fn outfifo_ovf_l1_ch3_int_raw( &mut self ) -> OUTFIFO_OVF_L1_CH3_INT_RAW_W<'_, OUT_INT_RAW_CH3_SPEC>

Bit 4 - The raw interrupt bit turns to high level when fifo is overflow.

pub fn outfifo_udf_l1_ch3_int_raw( &mut self ) -> OUTFIFO_UDF_L1_CH3_INT_RAW_W<'_, OUT_INT_RAW_CH3_SPEC>

Bit 5 - The raw interrupt bit turns to high level when fifo is underflow.

pub fn outfifo_ovf_l2_ch3_int_raw( &mut self ) -> OUTFIFO_OVF_L2_CH3_INT_RAW_W<'_, OUT_INT_RAW_CH3_SPEC>

Bit 6 - The raw interrupt bit turns to high level when fifo is overflow.

pub fn outfifo_udf_l2_ch3_int_raw( &mut self ) -> OUTFIFO_UDF_L2_CH3_INT_RAW_W<'_, OUT_INT_RAW_CH3_SPEC>

Bit 7 - The raw interrupt bit turns to high level when fifo is underflow.

pub fn out_dscr_task_ovf_ch3_int_raw( &mut self ) -> OUT_DSCR_TASK_OVF_CH3_INT_RAW_W<'_, OUT_INT_RAW_CH3_SPEC>

Bit 8 - The raw interrupt bit turns to high level when dscr ready task fifo is overflow.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual