Module esp32p4::i3c_mst::device_ctrl

DEVICE_CTRL register controls the transfer properties and disposition of controllers capabilities.

Structs

DEVICE_CTRL_SPEC

DEVICE_CTRL register controls the transfer properties and disposition of controllers capabilities.

Type Aliases

R

Register DEVICE_CTRL reader

REG_AUTO_DIS_IBI_EN_R

Field REG_AUTO_DIS_IBI_EN reader - NA

REG_AUTO_DIS_IBI_EN_W

Field REG_AUTO_DIS_IBI_EN writer - NA

REG_BA_INCLUDE_R

Field REG_BA_INCLUDE reader - This bit is used to include I3C broadcast address(0x7E) for private transfer.(If I3C broadcast address is not include for the private transfer, In-Band Interrupts driven from Slaves may not win address arbitration. Hence IBIs will get delayed)

REG_BA_INCLUDE_W

Field REG_BA_INCLUDE writer - This bit is used to include I3C broadcast address(0x7E) for private transfer.(If I3C broadcast address is not include for the private transfer, In-Band Interrupts driven from Slaves may not win address arbitration. Hence IBIs will get delayed)

REG_CLK_EN_R

Field REG_CLK_EN reader - NA

REG_CLK_EN_W

Field REG_CLK_EN writer - NA

REG_DATA_BYTE_CNT_UNLATCH_R

Field REG_DATA_BYTE_CNT_UNLATCH reader - 1: read current real-time updated value 0: read latch data byte cnt value

REG_DATA_BYTE_CNT_UNLATCH_W

Field REG_DATA_BYTE_CNT_UNLATCH writer - 1: read current real-time updated value 0: read latch data byte cnt value

REG_DMA_RX_EN_R

Field REG_DMA_RX_EN reader - NA

REG_DMA_RX_EN_W

Field REG_DMA_RX_EN writer - NA

REG_DMA_TX_EN_R

Field REG_DMA_TX_EN reader - NA

REG_DMA_TX_EN_W

Field REG_DMA_TX_EN writer - NA

REG_IBI_RSTART_TRANS_EN_R

Field REG_IBI_RSTART_TRANS_EN reader - NA

REG_IBI_RSTART_TRANS_EN_W

Field REG_IBI_RSTART_TRANS_EN writer - NA

REG_MEM_CLK_FORCE_ON_R

Field REG_MEM_CLK_FORCE_ON reader - 1: dev characteristic and address table memory clk date force on . 0 : clock gating by rd/wr.

REG_MEM_CLK_FORCE_ON_W

Field REG_MEM_CLK_FORCE_ON writer - 1: dev characteristic and address table memory clk date force on . 0 : clock gating by rd/wr.

REG_MULTI_SLV_SINGLE_CCC_EN_R

Field REG_MULTI_SLV_SINGLE_CCC_EN reader - 0: rx high bit first, 1: rx low bit first

REG_MULTI_SLV_SINGLE_CCC_EN_W

Field REG_MULTI_SLV_SINGLE_CCC_EN writer - 0: rx high bit first, 1: rx low bit first

REG_RX_BIT_ORDER_R

Field REG_RX_BIT_ORDER reader - 0: rx low byte fist, 1: rx high byte first

REG_RX_BIT_ORDER_W

Field REG_RX_BIT_ORDER writer - 0: rx low byte fist, 1: rx high byte first

REG_RX_BYTE_ORDER_R

Field REG_RX_BYTE_ORDER reader - NA

REG_RX_BYTE_ORDER_W

Field REG_RX_BYTE_ORDER writer - NA

REG_SCL_OE_FORCE_EN_R

Field REG_SCL_OE_FORCE_EN reader - This bit is used to force scl_oe

REG_SCL_OE_FORCE_EN_W

Field REG_SCL_OE_FORCE_EN writer - This bit is used to force scl_oe

REG_SCL_PULLUP_FORCE_EN_R

Field REG_SCL_PULLUP_FORCE_EN reader - This bit is used to force scl_pullup_en

REG_SCL_PULLUP_FORCE_EN_W

Field REG_SCL_PULLUP_FORCE_EN writer - This bit is used to force scl_pullup_en

REG_SDA_PP_RD_PULLUP_EN_R

Field REG_SDA_PP_RD_PULLUP_EN reader - NA

REG_SDA_PP_RD_PULLUP_EN_W

Field REG_SDA_PP_RD_PULLUP_EN writer - NA

REG_SDA_PP_WR_PULLUP_EN_R

Field REG_SDA_PP_WR_PULLUP_EN reader - NA

REG_SDA_PP_WR_PULLUP_EN_W

Field REG_SDA_PP_WR_PULLUP_EN writer - NA

REG_SDA_RD_TBIT_HLVL_PULLUP_EN_R

Field REG_SDA_RD_TBIT_HLVL_PULLUP_EN reader - NA

REG_SDA_RD_TBIT_HLVL_PULLUP_EN_W

Field REG_SDA_RD_TBIT_HLVL_PULLUP_EN writer - NA

REG_TRANS_START_R

Field REG_TRANS_START reader - Transfer Start

REG_TRANS_START_W

Field REG_TRANS_START writer - Transfer Start

W

Register DEVICE_CTRL writer