Type Alias esp32p4::i3c_mst::device_ctrl::R

pub type R = R<DEVICE_CTRL_SPEC>;

Register DEVICE_CTRL reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn reg_ba_include(&self) -> REG_BA_INCLUDE_R

Bit 1 - This bit is used to include I3C broadcast address(0x7E) for private transfer.(If I3C broadcast address is not include for the private transfer, In-Band Interrupts driven from Slaves may not win address arbitration. Hence IBIs will get delayed)

pub fn reg_trans_start(&self) -> REG_TRANS_START_R

Bit 2 - Transfer Start

pub fn reg_clk_en(&self) -> REG_CLK_EN_R

Bit 3 - NA

pub fn reg_ibi_rstart_trans_en(&self) -> REG_IBI_RSTART_TRANS_EN_R

Bit 4 - NA

pub fn reg_auto_dis_ibi_en(&self) -> REG_AUTO_DIS_IBI_EN_R

Bit 5 - NA

pub fn reg_dma_rx_en(&self) -> REG_DMA_RX_EN_R

Bit 6 - NA

pub fn reg_dma_tx_en(&self) -> REG_DMA_TX_EN_R

Bit 7 - NA

pub fn reg_multi_slv_single_ccc_en(&self) -> REG_MULTI_SLV_SINGLE_CCC_EN_R

Bit 8 - 0: rx high bit first, 1: rx low bit first

pub fn reg_rx_bit_order(&self) -> REG_RX_BIT_ORDER_R

Bit 9 - 0: rx low byte fist, 1: rx high byte first

pub fn reg_rx_byte_order(&self) -> REG_RX_BYTE_ORDER_R

Bit 10 - NA

pub fn reg_scl_pullup_force_en(&self) -> REG_SCL_PULLUP_FORCE_EN_R

Bit 11 - This bit is used to force scl_pullup_en

pub fn reg_scl_oe_force_en(&self) -> REG_SCL_OE_FORCE_EN_R

Bit 12 - This bit is used to force scl_oe

pub fn reg_sda_pp_rd_pullup_en(&self) -> REG_SDA_PP_RD_PULLUP_EN_R

Bit 13 - NA

pub fn reg_sda_rd_tbit_hlvl_pullup_en(&self) -> REG_SDA_RD_TBIT_HLVL_PULLUP_EN_R

Bit 14 - NA

pub fn reg_sda_pp_wr_pullup_en(&self) -> REG_SDA_PP_WR_PULLUP_EN_R

Bit 15 - NA

pub fn reg_data_byte_cnt_unlatch(&self) -> REG_DATA_BYTE_CNT_UNLATCH_R

Bit 16 - 1: read current real-time updated value 0: read latch data byte cnt value

pub fn reg_mem_clk_force_on(&self) -> REG_MEM_CLK_FORCE_ON_R

Bit 17 - 1: dev characteristic and address table memory clk date force on . 0 : clock gating by rd/wr.