Type Alias esp32p4::i3c_mst::device_ctrl::W

pub type W = W<DEVICE_CTRL_SPEC>;

Register DEVICE_CTRL writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn reg_ba_include(&mut self) -> REG_BA_INCLUDE_W<'_, DEVICE_CTRL_SPEC>

Bit 1 - This bit is used to include I3C broadcast address(0x7E) for private transfer.(If I3C broadcast address is not include for the private transfer, In-Band Interrupts driven from Slaves may not win address arbitration. Hence IBIs will get delayed)

pub fn reg_trans_start(&mut self) -> REG_TRANS_START_W<'_, DEVICE_CTRL_SPEC>

Bit 2 - Transfer Start

pub fn reg_clk_en(&mut self) -> REG_CLK_EN_W<'_, DEVICE_CTRL_SPEC>

Bit 3 - NA

pub fn reg_ibi_rstart_trans_en( &mut self ) -> REG_IBI_RSTART_TRANS_EN_W<'_, DEVICE_CTRL_SPEC>

Bit 4 - NA

pub fn reg_auto_dis_ibi_en( &mut self ) -> REG_AUTO_DIS_IBI_EN_W<'_, DEVICE_CTRL_SPEC>

Bit 5 - NA

pub fn reg_dma_rx_en(&mut self) -> REG_DMA_RX_EN_W<'_, DEVICE_CTRL_SPEC>

Bit 6 - NA

pub fn reg_dma_tx_en(&mut self) -> REG_DMA_TX_EN_W<'_, DEVICE_CTRL_SPEC>

Bit 7 - NA

pub fn reg_multi_slv_single_ccc_en( &mut self ) -> REG_MULTI_SLV_SINGLE_CCC_EN_W<'_, DEVICE_CTRL_SPEC>

Bit 8 - 0: rx high bit first, 1: rx low bit first

pub fn reg_rx_bit_order(&mut self) -> REG_RX_BIT_ORDER_W<'_, DEVICE_CTRL_SPEC>

Bit 9 - 0: rx low byte fist, 1: rx high byte first

pub fn reg_rx_byte_order(&mut self) -> REG_RX_BYTE_ORDER_W<'_, DEVICE_CTRL_SPEC>

Bit 10 - NA

pub fn reg_scl_pullup_force_en( &mut self ) -> REG_SCL_PULLUP_FORCE_EN_W<'_, DEVICE_CTRL_SPEC>

Bit 11 - This bit is used to force scl_pullup_en

pub fn reg_scl_oe_force_en( &mut self ) -> REG_SCL_OE_FORCE_EN_W<'_, DEVICE_CTRL_SPEC>

Bit 12 - This bit is used to force scl_oe

pub fn reg_sda_pp_rd_pullup_en( &mut self ) -> REG_SDA_PP_RD_PULLUP_EN_W<'_, DEVICE_CTRL_SPEC>

Bit 13 - NA

pub fn reg_sda_rd_tbit_hlvl_pullup_en( &mut self ) -> REG_SDA_RD_TBIT_HLVL_PULLUP_EN_W<'_, DEVICE_CTRL_SPEC>

Bit 14 - NA

pub fn reg_sda_pp_wr_pullup_en( &mut self ) -> REG_SDA_PP_WR_PULLUP_EN_W<'_, DEVICE_CTRL_SPEC>

Bit 15 - NA

pub fn reg_data_byte_cnt_unlatch( &mut self ) -> REG_DATA_BYTE_CNT_UNLATCH_W<'_, DEVICE_CTRL_SPEC>

Bit 16 - 1: read current real-time updated value 0: read latch data byte cnt value

pub fn reg_mem_clk_force_on( &mut self ) -> REG_MEM_CLK_FORCE_ON_W<'_, DEVICE_CTRL_SPEC>

Bit 17 - 1: dev characteristic and address table memory clk date force on . 0 : clock gating by rd/wr.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual