Module esp32p4::i3c_mst::int_st

NA

Structs

INT_ST_SPEC

NA

Type Aliases

CMD_BUF_EMPTY_THLD_INT_ST_R

Field CMD_BUF_EMPTY_THLD_INT_ST reader - This interrupt is generated when number of empty locations in command buffer is greater than or equal to threshold value specified by CMD_EMPTY_BUF_THLD field in BUFFER_THLD_CTRL register. This interrupt will be cleared automatically when number of empty locations in command buffer is less than threshold value.

CMD_CCC_MISMATCH_INT_ST_R

Field CMD_CCC_MISMATCH_INT_ST reader - NA

COMMAND_DONE_INT_ST_R

Field COMMAND_DONE_INT_ST reader - NA

DETECT_START_INT_ST_R

Field DETECT_START_INT_ST reader - NA

IBI_DATA_BUF_OVF_INT_ST_R

Field IBI_DATA_BUF_OVF_INT_ST reader - NA

IBI_DETECT_INT_ST_R

Field IBI_DETECT_INT_ST reader - NA

IBI_HANDLE_DONE_INT_ST_R

Field IBI_HANDLE_DONE_INT_ST reader - NA

IBI_STATUS_BUF_OVF_INT_ST_R

Field IBI_STATUS_BUF_OVF_INT_ST reader - NA

IBI_STATUS_THLD_INT_ST_R

Field IBI_STATUS_THLD_INT_ST reader - Only used in master mode. This interrupt is generated when number of entries in IBI buffer is greater than or equal to threshold value specified by IBI_BUF_THLD field in BUFFER_THLD_CTRL register. This interrupt will be cleared automatically when number of entries in IBI buffer is less than threshold value.

NXT_CMD_REQ_ERR_INT_ST_R

Field NXT_CMD_REQ_ERR_INT_ST reader - This interrupt is generated if toc is 0(master will restart next command), but command buf is empty.

R

Register INT_ST reader

RESP_BUF_OVF_INT_ST_R

Field RESP_BUF_OVF_INT_ST reader - NA

RESP_READY_INT_ST_R

Field RESP_READY_INT_ST reader - This interrupt is generated when number of entries in response buffer is greater than or equal to threshold value specified by RESP_BUF_THLD field in BUFFER_THLD_CTRL register. This interrupt will be cleared automatically when number of entries in response buffer is less than threshold value.

RX_DATA_BUF_THLD_INT_ST_R

Field RX_DATA_BUF_THLD_INT_ST reader - This interrupt is generated when number of entries in receive buffer is greater than or equal to threshold value specified by RX_BUF_THLD field in DATA_BUFFER_THLD_CTRL register. This interrupt will be cleared automatically when number of entries in receive buffer is less than threshold value.

TRANSFER_COMPLETE_INT_ST_R

Field TRANSFER_COMPLETE_INT_ST reader - NA

TRANSFER_ERR_INT_ST_R

Field TRANSFER_ERR_INT_ST reader - This interrupt is generated if any error occurs during transfer. The error type will be specified in the response packet associated with the command (in ERR_STATUS field of RESPONSE_BUFFER_PORT register). This bit can be cleared by writing 1’h1.

TX_DATA_BUF_THLD_INT_ST_R

Field TX_DATA_BUF_THLD_INT_ST reader - This interrupt is generated when number of empty locations in transmit buffer is greater than or equal to threshold value specified by TX_EMPTY_BUS_THLD field in DATA_BUFFER_THLD_CTRL register. This interrupt will be cleared automatically when number of empty locations in transmit buffer is less than threshold value.