Type Alias esp32p4::i3c_mst::present_state0::R

pub type R = R<PRESENT_STATE0_SPEC>;

Register PRESENT_STATE0 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn sda_lvl(&self) -> SDA_LVL_R

Bit 0 - This bit is used to check the SCL line level to recover from error and for debugging. This bit reflects the value of synchronized scl_in_a.

pub fn scl_lvl(&self) -> SCL_LVL_R

Bit 1 - This bit is used to check the SDA line level to recover from error and for debugging. This bit reflects the value of synchronized sda_in_a.

pub fn bus_busy(&self) -> BUS_BUSY_R

Bit 2 - NA

pub fn bus_free(&self) -> BUS_FREE_R

Bit 3 - NA

pub fn cmd_tid(&self) -> CMD_TID_R

Bits 9:12 - NA

pub fn scl_gen_fsm_state(&self) -> SCL_GEN_FSM_STATE_R

Bits 13:15 - NA

pub fn ibi_ev_handle_fsm_state(&self) -> IBI_EV_HANDLE_FSM_STATE_R

Bits 16:18 - NA

pub fn i2c_mode_fsm_state(&self) -> I2C_MODE_FSM_STATE_R

Bits 19:21 - NA

pub fn sdr_mode_fsm_state(&self) -> SDR_MODE_FSM_STATE_R

Bits 22:25 - NA

pub fn daa_mode_fsm_state(&self) -> DAA_MODE_FSM_STATE_R

Bits 26:28 - Reflects whether the Master Controller is in IDLE or not. This bit will be set when all the buffer(Command, Response, IBI, Transmit, Receive) are empty along with the Master State machine is in idle state. 0X0: not in idle 0x1: in idle

pub fn main_fsm_state(&self) -> MAIN_FSM_STATE_R

Bits 29:31 - NA