Module esp32p4::i3c_mst_mem::ibi_status_buf

In-Band Interrupt Buffer Status/Data Register. When receiving an IBI, IBI_PORT is used to both: Read the IBI Status Read the IBI Data(which is raw/opaque data)

Structs

IBI_STATUS_BUF_SPEC

In-Band Interrupt Buffer Status/Data Register. When receiving an IBI, IBI_PORT is used to both: Read the IBI Status Read the IBI Data(which is raw/opaque data)

Type Aliases

DATA_LENGTH_R

Field DATA_LENGTH reader - This field represents the length of data received along with IBI, in bytes.

IBI_ID_R

Field IBI_ID reader - IBI Identifier. The byte received after START which includes the address the R/W bit: Device address and R/W bit in case of Slave Interrupt or Master Request.

IBI_STS_R

Field IBI_STS reader - IBI received data/status. IBI Data register is mapped to the IBI Buffer. The IBI Data is always packed in4-byte aligned and put to the IBI Buffer. This register When read from, reads the data from the IBI buffer. IBI Status register when read from, returns the data from the IBI Buffer and indicates how the controller responded to incoming IBI(SIR, MR and HJ).

R

Register IBI_STATUS_BUF reader