Module esp32p4::i3c_slv::config

NA

Structs

CONFIG_SPEC

NA

Type Aliases

BAMATCH_R

Field BAMATCH reader - Bus Available condition match value for current ???Slow clock???. This provides the count of the slow clock to count out 1us (or more) to allow an IBI to drive SDA Low when the Master is not doing so. The max width , and so max value, is controlled by the block. Only if enabled for events such IBI or MR or HJ, and if enabled to provide this as a register. With is limited to CLK_SLOW_BITS

BAMATCH_W

Field BAMATCH writer - Bus Available condition match value for current ???Slow clock???. This provides the count of the slow clock to count out 1us (or more) to allow an IBI to drive SDA Low when the Master is not doing so. The max width , and so max value, is controlled by the block. Only if enabled for events such IBI or MR or HJ, and if enabled to provide this as a register. With is limited to CLK_SLOW_BITS

DDROK_R

Field DDROK reader - NA

DDROK_W

Field DDROK writer - NA

IDRAND_R

Field IDRAND reader - NA

IDRAND_W

Field IDRAND writer - NA

MATCHSS_R

Field MATCHSS reader - 1: the START and STOP sticky STATUS bits will only be set if MATCHED is set..This allows START and STOP to be used to detect end of a message to /from this slave.

MATCHSS_W

Field MATCHSS writer - 1: the START and STOP sticky STATUS bits will only be set if MATCHED is set..This allows START and STOP to be used to detect end of a message to /from this slave.

NACK_R

Field NACK reader - 1:the slave will NACK all requests to it except CCC broadcast. This should be used with caution as the Master may determine the slave is missing if overused.

NACK_W

Field NACK writer - 1:the slave will NACK all requests to it except CCC broadcast. This should be used with caution as the Master may determine the slave is missing if overused.

OFFLINE_R

Field OFFLINE reader - NA

OFFLINE_W

Field OFFLINE writer - NA

R

Register CONFIG reader

S0IGNORE_R

Field S0IGNORE reader - If 1, the Slave will not detect S0 or S1 errors and so not lock up waiting on an Exit Pattern. This should only be used when the bus will not use HDR.

S0IGNORE_W

Field S0IGNORE writer - If 1, the Slave will not detect S0 or S1 errors and so not lock up waiting on an Exit Pattern. This should only be used when the bus will not use HDR.

SADDR_R

Field SADDR reader - If allowed by the block:sets i2c 7 bits static address,else should be 0. If enabled to use one and to be provided by SW. Block may provide in HW as well.

SADDR_W

Field SADDR writer - If allowed by the block:sets i2c 7 bits static address,else should be 0. If enabled to use one and to be provided by SW. Block may provide in HW as well.

SLVENA_R

Field SLVENA reader - 1: allow the slave to operate on i2c or i3c bus. 0: the slave will ignore the bus. This should be not set until registers such as PARTNO, IDEXT and the like are set 1st -if used- since they impact data to the master

SLVENA_W

Field SLVENA writer - 1: allow the slave to operate on i2c or i3c bus. 0: the slave will ignore the bus. This should be not set until registers such as PARTNO, IDEXT and the like are set 1st -if used- since they impact data to the master

W

Register CONFIG writer