Module esp32p4::i3c_slv::ctrl

NA

Structs

CTRL_SPEC

NA

Type Aliases

ACTSTATE_R

Field ACTSTATE reader - NA

ACTSTATE_W

Field ACTSTATE writer - NA

EXTDATA_R

Field EXTDATA reader - reserved

EXTDATA_W

Field EXTDATA writer - reserved

IBIDATA_R

Field IBIDATA reader - Data byte to go with an IBI, if enabled for it. If enabled (was in BCR), then it is required.

IBIDATA_W

Field IBIDATA writer - Data byte to go with an IBI, if enabled for it. If enabled (was in BCR), then it is required.

MAPIDX_R

Field MAPIDX reader - Index of Dynamic Address that IBI is for. This is 0 for the main or base Dynamic Address, or can be any valid index.

MAPIDX_W

Field MAPIDX writer - Index of Dynamic Address that IBI is for. This is 0 for the main or base Dynamic Address, or can be any valid index.

PENDINT_R

Field PENDINT reader - Should be set to the pending interrupt that GETSTATUS CCC will return. This should be maintained by the application if used and configured, as the Master will read this. If not configured, the GETSTATUS field will return 1 if an IBI is pending, and 0 otherwise.

PENDINT_W

Field PENDINT writer - Should be set to the pending interrupt that GETSTATUS CCC will return. This should be maintained by the application if used and configured, as the Master will read this. If not configured, the GETSTATUS field will return 1 if an IBI is pending, and 0 otherwise.

R

Register CTRL reader

SLV_EVENT_R

Field SLV_EVENT reader - If set to non-0, will request an event. Once requested, STATUS.EVENT and EVDET will show the status as it progresses. Once completed, the field will automatically return to 0. Once non-0, only 0 can be written(to cancel) until done. 0: Normal mode. If set to 0 after was a non-0 value, will cancel if not already in flight. 1: start an IBI. This will try to push through an IBI on the bus. If data associate with the IBI, it will be drawn from the IBIDATA field. Note that if Time control is enabled, this will include anytime control related bytes further, the IBIDATA byte will have bit7 set to 1.

SLV_EVENT_W

Field SLV_EVENT writer - If set to non-0, will request an event. Once requested, STATUS.EVENT and EVDET will show the status as it progresses. Once completed, the field will automatically return to 0. Once non-0, only 0 can be written(to cancel) until done. 0: Normal mode. If set to 0 after was a non-0 value, will cancel if not already in flight. 1: start an IBI. This will try to push through an IBI on the bus. If data associate with the IBI, it will be drawn from the IBIDATA field. Note that if Time control is enabled, this will include anytime control related bytes further, the IBIDATA byte will have bit7 set to 1.

VENDINFO_R

Field VENDINFO reader - NA

VENDINFO_W

Field VENDINFO writer - NA

W

Register CTRL writer