Type Alias esp32p4::i3c_slv::ctrl::R

pub type R = R<CTRL_SPEC>;

Register CTRL reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn slv_event(&self) -> SLV_EVENT_R

Bits 0:1 - If set to non-0, will request an event. Once requested, STATUS.EVENT and EVDET will show the status as it progresses. Once completed, the field will automatically return to 0. Once non-0, only 0 can be written(to cancel) until done. 0: Normal mode. If set to 0 after was a non-0 value, will cancel if not already in flight. 1: start an IBI. This will try to push through an IBI on the bus. If data associate with the IBI, it will be drawn from the IBIDATA field. Note that if Time control is enabled, this will include anytime control related bytes further, the IBIDATA byte will have bit7 set to 1.

pub fn extdata(&self) -> EXTDATA_R

Bit 3 - reserved

pub fn mapidx(&self) -> MAPIDX_R

Bits 4:7 - Index of Dynamic Address that IBI is for. This is 0 for the main or base Dynamic Address, or can be any valid index.

pub fn ibidata(&self) -> IBIDATA_R

Bits 8:15 - Data byte to go with an IBI, if enabled for it. If enabled (was in BCR), then it is required.

pub fn pendint(&self) -> PENDINT_R

Bits 16:19 - Should be set to the pending interrupt that GETSTATUS CCC will return. This should be maintained by the application if used and configured, as the Master will read this. If not configured, the GETSTATUS field will return 1 if an IBI is pending, and 0 otherwise.

pub fn actstate(&self) -> ACTSTATE_R

Bits 20:21 - NA

pub fn vendinfo(&self) -> VENDINFO_R

Bits 24:31 - NA