Module esp32p4::i3c_slv::intclr

NA

Structs

INTCLR_SPEC

NA

Type Aliases

RXPEND_CLR_W

Field RXPEND_CLR writer - Interrupt when receiving a message from Master, which is not being handled by the block (excludes CCCs being handled automatically). If FIFO, then RX fullness trigger. If DMA, then message end.

STOP_CLR_W

Field STOP_CLR writer - Interrupt on STOP state on the bus. See Start as the preferred interrupt when needed. This interrupt may not trigger for quick STOP/START combination, as it relates to the state of being stopped.

TXSEND_CLR_W

Field TXSEND_CLR writer - NA

W

Register INTCLR writer