Module esp32p4::i3c_slv::status

NA

Structs

STATUS_SPEC

NA

Type Aliases

CCC_R

Field CCC reader - A common -command-code(CCC), not handled by block, has been received. This acts differently between: *Broadcasted ones, which will then also correspond with RXPEND and the 1st byte will be the CCC(command) . *Direct ones, which may never be directed to this device. If it is, then the TXSEND or RXPEND will be triggered with this end the RXPEND will contain the command.

CCC_W

Field CCC writer - A common -command-code(CCC), not handled by block, has been received. This acts differently between: *Broadcasted ones, which will then also correspond with RXPEND and the 1st byte will be the CCC(command) . *Direct ones, which may never be directed to this device. If it is, then the TXSEND or RXPEND will be triggered with this end the RXPEND will contain the command.

DACHG_R

Field DACHG reader - The Slv Dynamic Address has been assigned, reassigned, or reset(lost) and is now in that state of being valid or none. Actual DA can be seen in the DYNADDR register. Note that this will also be used when MAP Auto feature is configured. This will be changing one or more MAP items. See DYNADDR and/or MAPCTRLn. DYNAADDR for the main DA(0) will indicate if last change was due to Auto MAP.

DACHG_W

Field DACHG writer - The Slv Dynamic Address has been assigned, reassigned, or reset(lost) and is now in that state of being valid or none. Actual DA can be seen in the DYNADDR register. Note that this will also be used when MAP Auto feature is configured. This will be changing one or more MAP items. See DYNADDR and/or MAPCTRLn. DYNAADDR for the main DA(0) will indicate if last change was due to Auto MAP.

ERRWARN_R

Field ERRWARN reader - NA

HDRMATCH_R

Field HDRMATCH reader - NA

HDRMATCH_W

Field HDRMATCH writer - NA

MATCHED_R

Field MATCHED reader - NA

MATCHED_W

Field MATCHED writer - NA

R

Register STATUS reader

RXPEND_R

Field RXPEND reader - Receiving a message from master,which is not being handled by block(not a CCC internally processed). For all but External FIFO, this uses DATACTRL RXTRIG, which defaults to not-empty. If DMA is enabled for RX, DMA will be signaled as well. Will self-clear if data is read(FIFO and non-FIFO)

START_R

Field START reader - NA

START_W

Field START writer - NA

STCCCH_R

Field STCCCH reader - Is 1 if a CCC message is being handled automatically.

STDAA_R

Field STDAA reader - NA

STHDR_R

Field STHDR reader - NA

STMSG_R

Field STMSG reader - Is 1 if this bus Slave is listening to the bus traffic or repsonding, If STNOSTOP=1, then this will be 0 when a non-matching address seen until next respeated START it STOP.

STNOTSTOP_R

Field STNOTSTOP reader - Is 1 if bus is busy(activity) and 0 when in a STOP condition. Other bits may also set when busy. Note that this can also be true from an S0 or S1 error, which waits for an Exit Pattern.

STOP_R

Field STOP reader - NA

STOP_W

Field STOP writer - NA

STREQRD_R

Field STREQRD reader - 1 if the req in process is an sdr read from this slave or an IBI is being pushed out,

STREQWR_R

Field STREQWR reader - NA

TXNOTFULL_R

Field TXNOTFULL reader - Is 1 when the To-bus buffer/FIFO can accept more data to go out. Defau:1. For all but External FIFO, this uses DATACTRL TXTRIG,which defaults to not-full. If DMA is enabled for TX, it will also be signaled to provide more.

W

Register STATUS writer