Type Alias esp32p4::i3c_slv::status::R

pub type R = R<STATUS_SPEC>;

Register STATUS reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn stnotstop(&self) -> STNOTSTOP_R

Bit 0 - Is 1 if bus is busy(activity) and 0 when in a STOP condition. Other bits may also set when busy. Note that this can also be true from an S0 or S1 error, which waits for an Exit Pattern.

pub fn stmsg(&self) -> STMSG_R

Bit 1 - Is 1 if this bus Slave is listening to the bus traffic or repsonding, If STNOSTOP=1, then this will be 0 when a non-matching address seen until next respeated START it STOP.

pub fn stccch(&self) -> STCCCH_R

Bit 2 - Is 1 if a CCC message is being handled automatically.

pub fn streqrd(&self) -> STREQRD_R

Bit 3 - 1 if the req in process is an sdr read from this slave or an IBI is being pushed out,

pub fn streqwr(&self) -> STREQWR_R

Bit 4 - NA

pub fn stdaa(&self) -> STDAA_R

Bit 5 - NA

pub fn sthdr(&self) -> STHDR_R

Bit 6 - NA

pub fn start(&self) -> START_R

Bit 8 - NA

pub fn matched(&self) -> MATCHED_R

Bit 9 - NA

pub fn stop(&self) -> STOP_R

Bit 10 - NA

pub fn rxpend(&self) -> RXPEND_R

Bit 11 - Receiving a message from master,which is not being handled by block(not a CCC internally processed). For all but External FIFO, this uses DATACTRL RXTRIG, which defaults to not-empty. If DMA is enabled for RX, DMA will be signaled as well. Will self-clear if data is read(FIFO and non-FIFO)

pub fn txnotfull(&self) -> TXNOTFULL_R

Bit 12 - Is 1 when the To-bus buffer/FIFO can accept more data to go out. Defau:1. For all but External FIFO, this uses DATACTRL TXTRIG,which defaults to not-full. If DMA is enabled for TX, it will also be signaled to provide more.

pub fn dachg(&self) -> DACHG_R

Bit 13 - The Slv Dynamic Address has been assigned, reassigned, or reset(lost) and is now in that state of being valid or none. Actual DA can be seen in the DYNADDR register. Note that this will also be used when MAP Auto feature is configured. This will be changing one or more MAP items. See DYNADDR and/or MAPCTRLn. DYNAADDR for the main DA(0) will indicate if last change was due to Auto MAP.

pub fn ccc(&self) -> CCC_R

Bit 14 - A common -command-code(CCC), not handled by block, has been received. This acts differently between: *Broadcasted ones, which will then also correspond with RXPEND and the 1st byte will be the CCC(command) . *Direct ones, which may never be directed to this device. If it is, then the TXSEND or RXPEND will be triggered with this end the RXPEND will contain the command.

pub fn errwarn(&self) -> ERRWARN_R

Bit 15 - NA

pub fn hdrmatch(&self) -> HDRMATCH_R

Bit 16 - NA