Type Alias esp32p4::i3c_slv::status::W

pub type W = W<STATUS_SPEC>;

Register STATUS writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn start(&mut self) -> START_W<'_, STATUS_SPEC>

Bit 8 - NA

pub fn matched(&mut self) -> MATCHED_W<'_, STATUS_SPEC>

Bit 9 - NA

pub fn stop(&mut self) -> STOP_W<'_, STATUS_SPEC>

Bit 10 - NA

pub fn dachg(&mut self) -> DACHG_W<'_, STATUS_SPEC>

Bit 13 - The Slv Dynamic Address has been assigned, reassigned, or reset(lost) and is now in that state of being valid or none. Actual DA can be seen in the DYNADDR register. Note that this will also be used when MAP Auto feature is configured. This will be changing one or more MAP items. See DYNADDR and/or MAPCTRLn. DYNAADDR for the main DA(0) will indicate if last change was due to Auto MAP.

pub fn ccc(&mut self) -> CCC_W<'_, STATUS_SPEC>

Bit 14 - A common -command-code(CCC), not handled by block, has been received. This acts differently between: *Broadcasted ones, which will then also correspond with RXPEND and the 1st byte will be the CCC(command) . *Direct ones, which may never be directed to this device. If it is, then the TXSEND or RXPEND will be triggered with this end the RXPEND will contain the command.

pub fn hdrmatch(&mut self) -> HDRMATCH_W<'_, STATUS_SPEC>

Bit 16 - NA

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual