Module esp32p4::io_mux::gpio

IO_MUX Control Register

Structs

GPIO_SPEC

IO_MUX Control Register

Type Aliases

FILTER_EN_R

Field FILTER_EN reader - Configures whether or not to enable filter for pin input signals. 0: Disable 1: Enable

FILTER_EN_W

Field FILTER_EN writer - Configures whether or not to enable filter for pin input signals. 0: Disable 1: Enable

FUN_DRV_R

Field FUN_DRV reader - Configures the drive strength of GPIOn. 0: ~5 mA 1: ~10 mA 2: ~20 mA 3: ~40 mA

FUN_DRV_W

Field FUN_DRV writer - Configures the drive strength of GPIOn. 0: ~5 mA 1: ~10 mA 2: ~20 mA 3: ~40 mA

FUN_IE_R

Field FUN_IE reader - Configures whether or not to enable input of GPIOn. 0: Disable 1: Enable

FUN_IE_W

Field FUN_IE writer - Configures whether or not to enable input of GPIOn. 0: Disable 1: Enable

FUN_WPD_R

Field FUN_WPD reader - Configures whether or not to enable pull-down resistor of GPIOn. 0: Disable 1: Enable

FUN_WPD_W

Field FUN_WPD writer - Configures whether or not to enable pull-down resistor of GPIOn. 0: Disable 1: Enable

FUN_WPU_R

Field FUN_WPU reader - Configures whether or not enable pull-up resistor of GPIOn. 0: Disable 1: Enable

FUN_WPU_W

Field FUN_WPU writer - Configures whether or not enable pull-up resistor of GPIOn. 0: Disable 1: Enable

MCU_DRV_R

Field MCU_DRV reader - Configures the drive strength of GPIOn during sleep mode. 0: ~5 mA 1: ~10 mA 2: ~20 mA 3: ~40 mA

MCU_DRV_W

Field MCU_DRV writer - Configures the drive strength of GPIOn during sleep mode. 0: ~5 mA 1: ~10 mA 2: ~20 mA 3: ~40 mA

MCU_IE_R

Field MCU_IE reader - Configures whether or not to enable the input of GPIOn during sleep mode. 0: Disable 1: Enable

MCU_IE_W

Field MCU_IE writer - Configures whether or not to enable the input of GPIOn during sleep mode. 0: Disable 1: Enable

MCU_OE_R

Field MCU_OE reader - Configures whether or not to enable the output of GPIOn in sleep mode. 0: Disable 1: Enable

MCU_OE_W

Field MCU_OE writer - Configures whether or not to enable the output of GPIOn in sleep mode. 0: Disable 1: Enable

MCU_SEL_R

Field MCU_SEL reader - Configures to select IO MUX function for this pin. 0: Select Function 0 1: Select Function 1 ……

MCU_SEL_W

Field MCU_SEL writer - Configures to select IO MUX function for this pin. 0: Select Function 0 1: Select Function 1 ……

MCU_WPD_R

Field MCU_WPD reader - Configure whether or not to enable pull-down resistor of GPIOn during sleep mode. 0: Disable 1: Enable

MCU_WPD_W

Field MCU_WPD writer - Configure whether or not to enable pull-down resistor of GPIOn during sleep mode. 0: Disable 1: Enable

MCU_WPU_R

Field MCU_WPU reader - Configures whether or not to enable pull-up resistor of GPIOn during sleep mode. 0: Disable 1: Enable

MCU_WPU_W

Field MCU_WPU writer - Configures whether or not to enable pull-up resistor of GPIOn during sleep mode. 0: Disable 1: Enable

R

Register GPIO%s reader

SLP_SEL_R

Field SLP_SEL reader - Configures whether or not to enter sleep mode for GPIOn. 0: Not enter 1: Enter

SLP_SEL_W

Field SLP_SEL writer - Configures whether or not to enter sleep mode for GPIOn. 0: Not enter 1: Enter

W

Register GPIO%s writer