Module esp32p4::isp

ISP Peripheral

Modules

ae_block_mean_0

ae statistic result register 0

ae_block_mean_1

ae statistic result register 1

ae_block_mean_2

ae statistic result register 2

ae_block_mean_3

ae statistic result register 3

ae_block_mean_4

ae statistic result register 4

ae_block_mean_5

ae statistic result register 5

ae_block_mean_6

ae statistic result register 6

ae_bx

ae window register in x-direction

ae_by

ae window register in y-direction

ae_ctrl

ae control register

ae_monitor

ae monitor control register

ae_win_reciprocal

reciprocal of ae sub-window pixel number

ae_winpixnum

ae sub-window pix num register

af_ctrl0

af control register 0

af_ctrl1

af control register 1

af_env_user_th_lum

af monitor user lum threshold register

af_env_user_th_sum

af monitor user sum threshold register

af_gen_th_ctrl

af gen threshold control register

af_hscale_a

h-scale of af window a register

af_hscale_b

h-scale of af window b register

af_hscale_c

v-scale of af window c register

af_lum_a

result of lum of af window a

af_lum_b

result of lum of af window b

af_lum_c

result of lum of af window c

af_sum_a

result of sum of af window a

af_sum_b

result of sum of af window b

af_sum_c

result of sum of af window c

af_threshold

af threshold register

af_vscale_a

v-scale of af window a register

af_vscale_b

v-scale of af window b register

af_vscale_c

v-scale of af window c register

awb0_acc_b

result of accumulate of b channel of all white points

awb0_acc_g

result of accumulate of g channel of all white points

awb0_acc_r

result of accumulate of r channel of all white points

awb0_white_cnt

result of awb white point number

awb_hscale

h-scale of awb window

awb_mode

awb mode control register

awb_th_bg

awb b/g threshold register

awb_th_lum

awb lum threshold register

awb_th_rg

awb r/g threshold register

awb_vscale

v-scale of awb window

bf_gau0

bf gau template register 0

bf_gau1

bf gau template register 1

bf_matrix_ctrl

bf pix2matrix ctrl

bf_sigma

bf denoising level control register

blc_ctrl0

blc stretch control register

blc_ctrl1

blc window control register

blc_ctrl2

blc black threshold control register

blc_mean

results of the average of black window

blc_value

blc black level register

cam_cntl

isp cam source control register

cam_conf

isp cam source config register

ccm_coef0

ccm coef register 0

ccm_coef1

ccm coef register 1

ccm_coef3

ccm coef register 3

ccm_coef4

ccm coef register 4

ccm_coef5

ccm coef register 5

clk_en

isp clk control register

cntl

isp module enable control register

color_ctrl

color control register

demosaic_grad_ratio

demosaic gradient select ratio

demosaic_matrix_ctrl

demosaic pix2matrix ctrl

dma_cntl

isp dma source trans control register

dma_raw_data

isp dma source total raw number set register

dpc_conf

DPC parameter config register

dpc_ctrl

DPC mode control register

dpc_deadpix_cnt

DPC dead-pix number register

dpc_matrix_ctrl

dpc pix2matrix ctrl

frame_cfg

frame control parameter register

gamma_bx1

point of X-axis of b channel gamma curve register 1

gamma_bx2

point of X-axis of b channel gamma curve register 2

gamma_by1

point of Y-axis of b channel gamma curve register 1

gamma_by2

point of Y-axis of b channel gamma curve register 2

gamma_by3

point of Y-axis of b channel gamma curve register 3

gamma_by4

point of Y-axis of b channel gamma curve register 4

gamma_ctrl

gamma control register

gamma_gx1

point of X-axis of g channel gamma curve register 1

gamma_gx2

point of X-axis of g channel gamma curve register 2

gamma_gy1

point of Y-axis of g channel gamma curve register 1

gamma_gy2

point of Y-axis of g channel gamma curve register 2

gamma_gy3

point of Y-axis of g channel gamma curve register 3

gamma_gy4

point of Y-axis of g channel gamma curve register 4

gamma_rx1

point of X-axis of r channel gamma curve register 1

gamma_rx2

point of X-axis of r channel gamma curve register 2

gamma_ry1

point of Y-axis of r channel gamma curve register 1

gamma_ry2

point of Y-axis of r channel gamma curve register 2

gamma_ry3

point of Y-axis of r channel gamma curve register 3

gamma_ry4

point of Y-axis of r channel gamma curve register 4

hist_bin0

result of histogram bin 0

hist_bin1

result of histogram bin 1

hist_bin2

result of histogram bin 2

hist_bin3

result of histogram bin 3

hist_bin4

result of histogram bin 4

hist_bin5

result of histogram bin 5

hist_bin6

result of histogram bin 6

hist_bin7

result of histogram bin 7

hist_bin8

result of histogram bin 8

hist_bin9

result of histogram bin 9

hist_bin10

result of histogram bin 10

hist_bin11

result of histogram bin 11

hist_bin12

result of histogram bin 12

hist_bin13

result of histogram bin 13

hist_bin14

result of histogram bin 14

hist_bin15

result of histogram bin 15

hist_coeff

histogram rgb to gray coefficients register

hist_mode

histogram mode control register

hist_offs

histogram window offsets register

hist_seg0

histogram bin control register 0

hist_seg1

histogram bin control register 1

hist_seg2

histogram bin control register 2

hist_seg3

histogram bin control register 3

hist_size

histogram sub-window size register

hist_weight0

histogram sub-window weight register 0

hist_weight1

histogram sub-window weight register 1

hist_weight2

histogram sub-window weight register 2

hist_weight3

histogram sub-window weight register 3

hist_weight4

histogram sub-window weight register 4

hist_weight5

histogram sub-window weight register 5

hist_weight6

histogram sub-window weight register 6

hsync_cnt

header hsync interval control register

int_clr

interrupt clear register

int_ena

interrupt enable register

int_raw

raw interrupt register

int_st

masked interrupt register

lsc_tablesize

LSC point in x-direction

lut_cmd

LUT command register

lut_rdata

LUT read data register

lut_wdata

LUT write data register

median_matrix_ctrl

median pix2matrix ctrl

mem_aux_ctrl_0

mem aux control register 0

mem_aux_ctrl_1

mem aux control register 1

mem_aux_ctrl_2

mem aux control register 2

mem_aux_ctrl_3

mem aux control register 3

mem_aux_ctrl_4

mem aux control register 4

rdn_eco_cs

rdn eco cs register

rdn_eco_high

rdn eco all high register

rdn_eco_low

rdn eco all low register

sharp_ctrl0

sharp control register 0

sharp_ctrl1

sharp control register 1

sharp_filter0

sharp usm config register 0

sharp_filter1

sharp usm config register 1

sharp_filter2

sharp usm config register 2

sharp_matrix_ctrl

sharp pix2matrix ctrl

ver_date

version control register

yuv_format

yuv format control register

Structs

RegisterBlock

Register block

Type Aliases

AE_BLOCK_MEAN_0

AE_BLOCK_MEAN_0 (r) register accessor: ae statistic result register 0

AE_BLOCK_MEAN_1

AE_BLOCK_MEAN_1 (r) register accessor: ae statistic result register 1

AE_BLOCK_MEAN_2

AE_BLOCK_MEAN_2 (r) register accessor: ae statistic result register 2

AE_BLOCK_MEAN_3

AE_BLOCK_MEAN_3 (r) register accessor: ae statistic result register 3

AE_BLOCK_MEAN_4

AE_BLOCK_MEAN_4 (r) register accessor: ae statistic result register 4

AE_BLOCK_MEAN_5

AE_BLOCK_MEAN_5 (r) register accessor: ae statistic result register 5

AE_BLOCK_MEAN_6

AE_BLOCK_MEAN_6 (r) register accessor: ae statistic result register 6

AE_BX

AE_BX (rw) register accessor: ae window register in x-direction

AE_BY

AE_BY (rw) register accessor: ae window register in y-direction

AE_CTRL

AE_CTRL (rw) register accessor: ae control register

AE_MONITOR

AE_MONITOR (rw) register accessor: ae monitor control register

AE_WINPIXNUM

AE_WINPIXNUM (rw) register accessor: ae sub-window pix num register

AE_WIN_RECIPROCAL

AE_WIN_RECIPROCAL (rw) register accessor: reciprocal of ae sub-window pixel number

AF_CTRL0

AF_CTRL0 (rw) register accessor: af control register 0

AF_CTRL1

AF_CTRL1 (rw) register accessor: af control register 1

AF_ENV_USER_TH_LUM

AF_ENV_USER_TH_LUM (rw) register accessor: af monitor user lum threshold register

AF_ENV_USER_TH_SUM

AF_ENV_USER_TH_SUM (rw) register accessor: af monitor user sum threshold register

AF_GEN_TH_CTRL

AF_GEN_TH_CTRL (rw) register accessor: af gen threshold control register

AF_HSCALE_A

AF_HSCALE_A (rw) register accessor: h-scale of af window a register

AF_HSCALE_B

AF_HSCALE_B (rw) register accessor: h-scale of af window b register

AF_HSCALE_C

AF_HSCALE_C (rw) register accessor: v-scale of af window c register

AF_LUM_A

AF_LUM_A (r) register accessor: result of lum of af window a

AF_LUM_B

AF_LUM_B (r) register accessor: result of lum of af window b

AF_LUM_C

AF_LUM_C (r) register accessor: result of lum of af window c

AF_SUM_A

AF_SUM_A (r) register accessor: result of sum of af window a

AF_SUM_B

AF_SUM_B (r) register accessor: result of sum of af window b

AF_SUM_C

AF_SUM_C (r) register accessor: result of sum of af window c

AF_THRESHOLD

AF_THRESHOLD (rw) register accessor: af threshold register

AF_VSCALE_A

AF_VSCALE_A (rw) register accessor: v-scale of af window a register

AF_VSCALE_B

AF_VSCALE_B (rw) register accessor: v-scale of af window b register

AF_VSCALE_C

AF_VSCALE_C (rw) register accessor: v-scale of af window c register

AWB0_ACC_B

AWB0_ACC_B (r) register accessor: result of accumulate of b channel of all white points

AWB0_ACC_G

AWB0_ACC_G (r) register accessor: result of accumulate of g channel of all white points

AWB0_ACC_R

AWB0_ACC_R (r) register accessor: result of accumulate of r channel of all white points

AWB0_WHITE_CNT

AWB0_WHITE_CNT (r) register accessor: result of awb white point number

AWB_HSCALE

AWB_HSCALE (rw) register accessor: h-scale of awb window

AWB_MODE

AWB_MODE (rw) register accessor: awb mode control register

AWB_TH_BG

AWB_TH_BG (rw) register accessor: awb b/g threshold register

AWB_TH_LUM

AWB_TH_LUM (rw) register accessor: awb lum threshold register

AWB_TH_RG

AWB_TH_RG (rw) register accessor: awb r/g threshold register

AWB_VSCALE

AWB_VSCALE (rw) register accessor: v-scale of awb window

BF_GAU0

BF_GAU0 (rw) register accessor: bf gau template register 0

BF_GAU1

BF_GAU1 (rw) register accessor: bf gau template register 1

BF_MATRIX_CTRL

BF_MATRIX_CTRL (rw) register accessor: bf pix2matrix ctrl

BF_SIGMA

BF_SIGMA (rw) register accessor: bf denoising level control register

BLC_CTRL0

BLC_CTRL0 (rw) register accessor: blc stretch control register

BLC_CTRL1

BLC_CTRL1 (rw) register accessor: blc window control register

BLC_CTRL2

BLC_CTRL2 (rw) register accessor: blc black threshold control register

BLC_MEAN

BLC_MEAN (r) register accessor: results of the average of black window

BLC_VALUE

BLC_VALUE (rw) register accessor: blc black level register

CAM_CNTL

CAM_CNTL (rw) register accessor: isp cam source control register

CAM_CONF

CAM_CONF (rw) register accessor: isp cam source config register

CCM_COEF0

CCM_COEF0 (rw) register accessor: ccm coef register 0

CCM_COEF1

CCM_COEF1 (rw) register accessor: ccm coef register 1

CCM_COEF3

CCM_COEF3 (rw) register accessor: ccm coef register 3

CCM_COEF4

CCM_COEF4 (rw) register accessor: ccm coef register 4

CCM_COEF5

CCM_COEF5 (rw) register accessor: ccm coef register 5

CLK_EN

CLK_EN (rw) register accessor: isp clk control register

CNTL

CNTL (rw) register accessor: isp module enable control register

COLOR_CTRL

COLOR_CTRL (rw) register accessor: color control register

DEMOSAIC_GRAD_RATIO

DEMOSAIC_GRAD_RATIO (rw) register accessor: demosaic gradient select ratio

DEMOSAIC_MATRIX_CTRL

DEMOSAIC_MATRIX_CTRL (rw) register accessor: demosaic pix2matrix ctrl

DMA_CNTL

DMA_CNTL (rw) register accessor: isp dma source trans control register

DMA_RAW_DATA

DMA_RAW_DATA (rw) register accessor: isp dma source total raw number set register

DPC_CONF

DPC_CONF (rw) register accessor: DPC parameter config register

DPC_CTRL

DPC_CTRL (rw) register accessor: DPC mode control register

DPC_DEADPIX_CNT

DPC_DEADPIX_CNT (r) register accessor: DPC dead-pix number register

DPC_MATRIX_CTRL

DPC_MATRIX_CTRL (rw) register accessor: dpc pix2matrix ctrl

FRAME_CFG

FRAME_CFG (rw) register accessor: frame control parameter register

GAMMA_BX1

GAMMA_BX1 (rw) register accessor: point of X-axis of b channel gamma curve register 1

GAMMA_BX2

GAMMA_BX2 (rw) register accessor: point of X-axis of b channel gamma curve register 2

GAMMA_BY1

GAMMA_BY1 (rw) register accessor: point of Y-axis of b channel gamma curve register 1

GAMMA_BY2

GAMMA_BY2 (rw) register accessor: point of Y-axis of b channel gamma curve register 2

GAMMA_BY3

GAMMA_BY3 (rw) register accessor: point of Y-axis of b channel gamma curve register 3

GAMMA_BY4

GAMMA_BY4 (rw) register accessor: point of Y-axis of b channel gamma curve register 4

GAMMA_CTRL

GAMMA_CTRL (rw) register accessor: gamma control register

GAMMA_GX1

GAMMA_GX1 (rw) register accessor: point of X-axis of g channel gamma curve register 1

GAMMA_GX2

GAMMA_GX2 (rw) register accessor: point of X-axis of g channel gamma curve register 2

GAMMA_GY1

GAMMA_GY1 (rw) register accessor: point of Y-axis of g channel gamma curve register 1

GAMMA_GY2

GAMMA_GY2 (rw) register accessor: point of Y-axis of g channel gamma curve register 2

GAMMA_GY3

GAMMA_GY3 (rw) register accessor: point of Y-axis of g channel gamma curve register 3

GAMMA_GY4

GAMMA_GY4 (rw) register accessor: point of Y-axis of g channel gamma curve register 4

GAMMA_RX1

GAMMA_RX1 (rw) register accessor: point of X-axis of r channel gamma curve register 1

GAMMA_RX2

GAMMA_RX2 (rw) register accessor: point of X-axis of r channel gamma curve register 2

GAMMA_RY1

GAMMA_RY1 (rw) register accessor: point of Y-axis of r channel gamma curve register 1

GAMMA_RY2

GAMMA_RY2 (rw) register accessor: point of Y-axis of r channel gamma curve register 2

GAMMA_RY3

GAMMA_RY3 (rw) register accessor: point of Y-axis of r channel gamma curve register 3

GAMMA_RY4

GAMMA_RY4 (rw) register accessor: point of Y-axis of r channel gamma curve register 4

HIST_BIN0

HIST_BIN0 (r) register accessor: result of histogram bin 0

HIST_BIN1

HIST_BIN1 (r) register accessor: result of histogram bin 1

HIST_BIN2

HIST_BIN2 (r) register accessor: result of histogram bin 2

HIST_BIN3

HIST_BIN3 (r) register accessor: result of histogram bin 3

HIST_BIN4

HIST_BIN4 (r) register accessor: result of histogram bin 4

HIST_BIN5

HIST_BIN5 (r) register accessor: result of histogram bin 5

HIST_BIN6

HIST_BIN6 (r) register accessor: result of histogram bin 6

HIST_BIN7

HIST_BIN7 (r) register accessor: result of histogram bin 7

HIST_BIN8

HIST_BIN8 (r) register accessor: result of histogram bin 8

HIST_BIN9

HIST_BIN9 (r) register accessor: result of histogram bin 9

HIST_BIN10

HIST_BIN10 (r) register accessor: result of histogram bin 10

HIST_BIN11

HIST_BIN11 (r) register accessor: result of histogram bin 11

HIST_BIN12

HIST_BIN12 (r) register accessor: result of histogram bin 12

HIST_BIN13

HIST_BIN13 (r) register accessor: result of histogram bin 13

HIST_BIN14

HIST_BIN14 (r) register accessor: result of histogram bin 14

HIST_BIN15

HIST_BIN15 (r) register accessor: result of histogram bin 15

HIST_COEFF

HIST_COEFF (rw) register accessor: histogram rgb to gray coefficients register

HIST_MODE

HIST_MODE (rw) register accessor: histogram mode control register

HIST_OFFS

HIST_OFFS (rw) register accessor: histogram window offsets register

HIST_SEG0

HIST_SEG0 (rw) register accessor: histogram bin control register 0

HIST_SEG1

HIST_SEG1 (rw) register accessor: histogram bin control register 1

HIST_SEG2

HIST_SEG2 (rw) register accessor: histogram bin control register 2

HIST_SEG3

HIST_SEG3 (rw) register accessor: histogram bin control register 3

HIST_SIZE

HIST_SIZE (rw) register accessor: histogram sub-window size register

HIST_WEIGHT0

HIST_WEIGHT0 (rw) register accessor: histogram sub-window weight register 0

HIST_WEIGHT1

HIST_WEIGHT1 (rw) register accessor: histogram sub-window weight register 1

HIST_WEIGHT2

HIST_WEIGHT2 (rw) register accessor: histogram sub-window weight register 2

HIST_WEIGHT3

HIST_WEIGHT3 (rw) register accessor: histogram sub-window weight register 3

HIST_WEIGHT4

HIST_WEIGHT4 (rw) register accessor: histogram sub-window weight register 4

HIST_WEIGHT5

HIST_WEIGHT5 (rw) register accessor: histogram sub-window weight register 5

HIST_WEIGHT6

HIST_WEIGHT6 (rw) register accessor: histogram sub-window weight register 6

HSYNC_CNT

HSYNC_CNT (rw) register accessor: header hsync interval control register

INT_CLR

INT_CLR (w) register accessor: interrupt clear register

INT_ENA

INT_ENA (rw) register accessor: interrupt enable register

INT_RAW

INT_RAW (r) register accessor: raw interrupt register

INT_ST

INT_ST (r) register accessor: masked interrupt register

LSC_TABLESIZE

LSC_TABLESIZE (rw) register accessor: LSC point in x-direction

LUT_CMD

LUT_CMD (w) register accessor: LUT command register

LUT_RDATA

LUT_RDATA (r) register accessor: LUT read data register

LUT_WDATA

LUT_WDATA (rw) register accessor: LUT write data register

MEDIAN_MATRIX_CTRL

MEDIAN_MATRIX_CTRL (rw) register accessor: median pix2matrix ctrl

MEM_AUX_CTRL_0

MEM_AUX_CTRL_0 (rw) register accessor: mem aux control register 0

MEM_AUX_CTRL_1

MEM_AUX_CTRL_1 (rw) register accessor: mem aux control register 1

MEM_AUX_CTRL_2

MEM_AUX_CTRL_2 (rw) register accessor: mem aux control register 2

MEM_AUX_CTRL_3

MEM_AUX_CTRL_3 (rw) register accessor: mem aux control register 3

MEM_AUX_CTRL_4

MEM_AUX_CTRL_4 (rw) register accessor: mem aux control register 4

RDN_ECO_CS

RDN_ECO_CS (rw) register accessor: rdn eco cs register

RDN_ECO_HIGH

RDN_ECO_HIGH (rw) register accessor: rdn eco all high register

RDN_ECO_LOW

RDN_ECO_LOW (rw) register accessor: rdn eco all low register

SHARP_CTRL0

SHARP_CTRL0 (rw) register accessor: sharp control register 0

SHARP_CTRL1

SHARP_CTRL1 (r) register accessor: sharp control register 1

SHARP_FILTER0

SHARP_FILTER0 (rw) register accessor: sharp usm config register 0

SHARP_FILTER1

SHARP_FILTER1 (rw) register accessor: sharp usm config register 1

SHARP_FILTER2

SHARP_FILTER2 (rw) register accessor: sharp usm config register 2

SHARP_MATRIX_CTRL

SHARP_MATRIX_CTRL (rw) register accessor: sharp pix2matrix ctrl

VER_DATE

VER_DATE (rw) register accessor: version control register

YUV_FORMAT

YUV_FORMAT (rw) register accessor: yuv format control register