Module esp32p4::isp::blc_value

blc black level register

Structs

BLC_VALUE_SPEC

blc black level register

Type Aliases

BLC_R0_VALUE_R

Field BLC_R0_VALUE reader - this field configures the black level of top left channel of bayer img

BLC_R0_VALUE_W

Field BLC_R0_VALUE writer - this field configures the black level of top left channel of bayer img

BLC_R1_VALUE_R

Field BLC_R1_VALUE reader - this field configures the black level of top right channel of bayer img

BLC_R1_VALUE_W

Field BLC_R1_VALUE writer - this field configures the black level of top right channel of bayer img

BLC_R2_VALUE_R

Field BLC_R2_VALUE reader - this field configures the black level of bottom left channel of bayer img

BLC_R2_VALUE_W

Field BLC_R2_VALUE writer - this field configures the black level of bottom left channel of bayer img

BLC_R3_VALUE_R

Field BLC_R3_VALUE reader - this field configures the black level of bottom right channel of bayer img

BLC_R3_VALUE_W

Field BLC_R3_VALUE writer - this field configures the black level of bottom right channel of bayer img

R

Register BLC_VALUE reader

W

Register BLC_VALUE writer