Module esp32p4::isp::dpc_ctrl

DPC mode control register

Structs

DPC_CTRL_SPEC

DPC mode control register

Type Aliases

DPC_BLACK_EN_R

Field DPC_BLACK_EN reader - this bit configures input image type select when in check mode, 0: white img, 1: black img

DPC_BLACK_EN_W

Field DPC_BLACK_EN writer - this bit configures input image type select when in check mode, 0: white img, 1: black img

DPC_CHECK_EN_R

Field DPC_CHECK_EN reader - this bit configures the check mode enable. 0: disable, 1: enable

DPC_CHECK_EN_W

Field DPC_CHECK_EN writer - this bit configures the check mode enable. 0: disable, 1: enable

DPC_CHECK_OD_EN_R

Field DPC_CHECK_OD_EN reader - this bit configures output pixel data when in check mode or not. 0: no data output, 1: data output

DPC_CHECK_OD_EN_W

Field DPC_CHECK_OD_EN writer - this bit configures output pixel data when in check mode or not. 0: no data output, 1: data output

DPC_METHOD_SEL_R

Field DPC_METHOD_SEL reader - this bit configures dyn dpc method select. 0: simple method, 1: hard method

DPC_METHOD_SEL_W

Field DPC_METHOD_SEL writer - this bit configures dyn dpc method select. 0: simple method, 1: hard method

DYN_EN_R

Field DYN_EN reader - this bit configures the dyn dpc enable. 0: disable, 1: enable

DYN_EN_W

Field DYN_EN writer - this bit configures the dyn dpc enable. 0: disable, 1: enable

R

Register DPC_CTRL reader

STA_EN_R

Field STA_EN reader - this bit configures the sta dpc enable. 0: disable, 1: enable

STA_EN_W

Field STA_EN writer - this bit configures the sta dpc enable. 0: disable, 1: enable

W

Register DPC_CTRL writer