Type Alias esp32p4::isp::frame_cfg::W

pub type W = W<FRAME_CFG_SPEC>;

Register FRAME_CFG writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn vadr_num(&mut self) -> VADR_NUM_W<'_, FRAME_CFG_SPEC>

Bits 0:11 - this field configures input image size in y-direction, image row number - 1

pub fn hadr_num(&mut self) -> HADR_NUM_W<'_, FRAME_CFG_SPEC>

Bits 12:23 - this field configures input image size in x-direction, image line number - 1

pub fn bayer_mode(&mut self) -> BAYER_MODE_W<'_, FRAME_CFG_SPEC>

Bits 27:28 - this field configures the bayer mode of input pixel. 00 : BG/GR 01 : GB/RG 10 : GR/BG 11 : RG/GB

pub fn hsync_start_exist(&mut self) -> HSYNC_START_EXIST_W<'_, FRAME_CFG_SPEC>

Bit 29 - this bit configures the line end packet exist or not. 0: not exist, 1: exist

pub fn hsync_end_exist(&mut self) -> HSYNC_END_EXIST_W<'_, FRAME_CFG_SPEC>

Bit 30 - this bit configures the line start packet exist or not. 0: not exist, 1: exist

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual