Type Alias esp32p4::isp::gamma_gx1::W

pub type W = W<GAMMA_GX1_SPEC>;

Register GAMMA_GX1 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn gamma_g_x07(&mut self) -> GAMMA_G_X07_W<'_, GAMMA_GX1_SPEC>

Bits 0:2 - this field configures the point 7 of X-axis of g channel gamma curve, it represents the power of the distance from the previous point

pub fn gamma_g_x06(&mut self) -> GAMMA_G_X06_W<'_, GAMMA_GX1_SPEC>

Bits 3:5 - this field configures the point 6 of X-axis of g channel gamma curve, it represents the power of the distance from the previous point

pub fn gamma_g_x05(&mut self) -> GAMMA_G_X05_W<'_, GAMMA_GX1_SPEC>

Bits 6:8 - this field configures the point 5 of X-axis of g channel gamma curve, it represents the power of the distance from the previous point

pub fn gamma_g_x04(&mut self) -> GAMMA_G_X04_W<'_, GAMMA_GX1_SPEC>

Bits 9:11 - this field configures the point 4 of X-axis of g channel gamma curve, it represents the power of the distance from the previous point

pub fn gamma_g_x03(&mut self) -> GAMMA_G_X03_W<'_, GAMMA_GX1_SPEC>

Bits 12:14 - this field configures the point 3 of X-axis of g channel gamma curve, it represents the power of the distance from the previous point

pub fn gamma_g_x02(&mut self) -> GAMMA_G_X02_W<'_, GAMMA_GX1_SPEC>

Bits 15:17 - this field configures the point 2 of X-axis of g channel gamma curve, it represents the power of the distance from the previous point

pub fn gamma_g_x01(&mut self) -> GAMMA_G_X01_W<'_, GAMMA_GX1_SPEC>

Bits 18:20 - this field configures the point 1 of X-axis of g channel gamma curve, it represents the power of the distance from the previous point

pub fn gamma_g_x00(&mut self) -> GAMMA_G_X00_W<'_, GAMMA_GX1_SPEC>

Bits 21:23 - this field configures the point 0 of X-axis of g channel gamma curve, it represents the power of the distance from the previous point

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual