Type Alias esp32p4::isp::int_raw::R

pub type R = R<INT_RAW_SPEC>;

Register INT_RAW reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn isp_data_type_err_int_raw(&self) -> ISP_DATA_TYPE_ERR_INT_RAW_R

Bit 0 - the raw interrupt status of input data type error. isp only support RGB bayer data type, other type will report type_err_int

pub fn isp_async_fifo_ovf_int_raw(&self) -> ISP_ASYNC_FIFO_OVF_INT_RAW_R

Bit 1 - the raw interrupt status of isp input fifo overflow

pub fn isp_buf_full_int_raw(&self) -> ISP_BUF_FULL_INT_RAW_R

Bit 2 - the raw interrupt status of isp input buffer full

pub fn isp_hvnum_setting_err_int_raw(&self) -> ISP_HVNUM_SETTING_ERR_INT_RAW_R

Bit 3 - the raw interrupt status of hnum and vnum setting format error

pub fn isp_data_type_setting_err_int_raw( &self ) -> ISP_DATA_TYPE_SETTING_ERR_INT_RAW_R

Bit 4 - the raw interrupt status of setting invalid reg_data_type

pub fn isp_mipi_hnum_unmatch_int_raw(&self) -> ISP_MIPI_HNUM_UNMATCH_INT_RAW_R

Bit 5 - the raw interrupt status of hnum setting unmatch with mipi input

pub fn dpc_check_done_int_raw(&self) -> DPC_CHECK_DONE_INT_RAW_R

Bit 6 - the raw interrupt status of dpc check done

pub fn gamma_xcoord_err_int_raw(&self) -> GAMMA_XCOORD_ERR_INT_RAW_R

Bit 7 - the raw interrupt status of gamma setting error. it report the sum of the lengths represented by reg_gamma_x00~x0F isn’t equal to 256

pub fn ae_monitor_int_raw(&self) -> AE_MONITOR_INT_RAW_R

Bit 8 - the raw interrupt status of ae monitor

pub fn ae_frame_done_int_raw(&self) -> AE_FRAME_DONE_INT_RAW_R

Bit 9 - the raw interrupt status of ae.

pub fn af_fdone_int_raw(&self) -> AF_FDONE_INT_RAW_R

Bit 10 - the raw interrupt status of af statistic. when auto_update enable, each frame done will send one int pulse when manual_update, each time when write 1 to reg_manual_update will send a int pulse when next frame done

pub fn af_env_int_raw(&self) -> AF_ENV_INT_RAW_R

Bit 11 - the raw interrupt status of af monitor. send a int pulse when env_det function enabled and environment changes detected

pub fn awb_fdone_int_raw(&self) -> AWB_FDONE_INT_RAW_R

Bit 12 - the raw interrupt status of awb. send a int pulse when statistic of one awb frame done

pub fn hist_fdone_int_raw(&self) -> HIST_FDONE_INT_RAW_R

Bit 13 - the raw interrupt status of histogram. send a int pulse when statistic of one frame histogram done

pub fn frame_int_raw(&self) -> FRAME_INT_RAW_R

Bit 14 - the raw interrupt status of isp frame end

pub fn blc_frame_int_raw(&self) -> BLC_FRAME_INT_RAW_R

Bit 15 - the raw interrupt status of blc frame done

pub fn lsc_frame_int_raw(&self) -> LSC_FRAME_INT_RAW_R

Bit 16 - the raw interrupt status of lsc frame done

pub fn dpc_frame_int_raw(&self) -> DPC_FRAME_INT_RAW_R

Bit 17 - the raw interrupt status of dpc frame done

pub fn bf_frame_int_raw(&self) -> BF_FRAME_INT_RAW_R

Bit 18 - the raw interrupt status of bf frame done

pub fn demosaic_frame_int_raw(&self) -> DEMOSAIC_FRAME_INT_RAW_R

Bit 19 - the raw interrupt status of demosaic frame done

pub fn median_frame_int_raw(&self) -> MEDIAN_FRAME_INT_RAW_R

Bit 20 - the raw interrupt status of median frame done

pub fn ccm_frame_int_raw(&self) -> CCM_FRAME_INT_RAW_R

Bit 21 - the raw interrupt status of ccm frame done

pub fn gamma_frame_int_raw(&self) -> GAMMA_FRAME_INT_RAW_R

Bit 22 - the raw interrupt status of gamma frame done

pub fn rgb2yuv_frame_int_raw(&self) -> RGB2YUV_FRAME_INT_RAW_R

Bit 23 - the raw interrupt status of rgb2yuv frame done

pub fn sharp_frame_int_raw(&self) -> SHARP_FRAME_INT_RAW_R

Bit 24 - the raw interrupt status of sharp frame done

pub fn color_frame_int_raw(&self) -> COLOR_FRAME_INT_RAW_R

Bit 25 - the raw interrupt status of color frame done

pub fn yuv2rgb_frame_int_raw(&self) -> YUV2RGB_FRAME_INT_RAW_R

Bit 26 - the raw interrupt status of yuv2rgb frame done

pub fn tail_idi_frame_int_raw(&self) -> TAIL_IDI_FRAME_INT_RAW_R

Bit 27 - the raw interrupt status of isp_tail idi frame_end

pub fn header_idi_frame_int_raw(&self) -> HEADER_IDI_FRAME_INT_RAW_R

Bit 28 - the raw interrupt status of real input frame end of isp_input