Module esp32p4::isp::int_st

masked interrupt register

Structs

INT_ST_SPEC

masked interrupt register

Type Aliases

AE_FRAME_DONE_INT_ST_R

Field AE_FRAME_DONE_INT_ST reader - the masked interrupt status of ae

AE_MONITOR_INT_ST_R

Field AE_MONITOR_INT_ST reader - the masked interrupt status of ae monitor

AF_ENV_INT_ST_R

Field AF_ENV_INT_ST reader - the masked interrupt status of af monitor

AF_FDONE_INT_ST_R

Field AF_FDONE_INT_ST reader - the masked interrupt status of af statistic

AWB_FDONE_INT_ST_R

Field AWB_FDONE_INT_ST reader - the masked interrupt status of awb

BF_FRAME_INT_ST_R

Field BF_FRAME_INT_ST reader - the masked interrupt status of bf frame done

BLC_FRAME_INT_ST_R

Field BLC_FRAME_INT_ST reader - the masked interrupt status of blc frame done

CCM_FRAME_INT_ST_R

Field CCM_FRAME_INT_ST reader - the masked interrupt status of ccm frame done

COLOR_FRAME_INT_ST_R

Field COLOR_FRAME_INT_ST reader - the masked interrupt status of color frame done

DEMOSAIC_FRAME_INT_ST_R

Field DEMOSAIC_FRAME_INT_ST reader - the masked interrupt status of demosaic frame done

DPC_CHECK_DONE_INT_ST_R

Field DPC_CHECK_DONE_INT_ST reader - the masked interrupt status of dpc check done

DPC_FRAME_INT_ST_R

Field DPC_FRAME_INT_ST reader - the masked interrupt status of dpc frame done

FRAME_INT_ST_R

Field FRAME_INT_ST reader - the masked interrupt status of isp frame end

GAMMA_FRAME_INT_ST_R

Field GAMMA_FRAME_INT_ST reader - the masked interrupt status of gamma frame done

GAMMA_XCOORD_ERR_INT_ST_R

Field GAMMA_XCOORD_ERR_INT_ST reader - the masked interrupt status of gamma setting error

HEADER_IDI_FRAME_INT_ST_R

Field HEADER_IDI_FRAME_INT_ST reader - the masked interrupt status of real input frame end of isp_input

HIST_FDONE_INT_ST_R

Field HIST_FDONE_INT_ST reader - the masked interrupt status of histogram

ISP_ASYNC_FIFO_OVF_INT_ST_R

Field ISP_ASYNC_FIFO_OVF_INT_ST reader - the masked interrupt status of isp input fifo overflow

ISP_BUF_FULL_INT_ST_R

Field ISP_BUF_FULL_INT_ST reader - the masked interrupt status of isp input buffer full

ISP_DATA_TYPE_ERR_INT_ST_R

Field ISP_DATA_TYPE_ERR_INT_ST reader - the masked interrupt status of input data type error

ISP_DATA_TYPE_SETTING_ERR_INT_ST_R

Field ISP_DATA_TYPE_SETTING_ERR_INT_ST reader - the masked interrupt status of setting invalid reg_data_type

ISP_HVNUM_SETTING_ERR_INT_ST_R

Field ISP_HVNUM_SETTING_ERR_INT_ST reader - the masked interrupt status of hnum and vnum setting format error

ISP_MIPI_HNUM_UNMATCH_INT_ST_R

Field ISP_MIPI_HNUM_UNMATCH_INT_ST reader - the masked interrupt status of hnum setting unmatch with mipi input

LSC_FRAME_INT_ST_R

Field LSC_FRAME_INT_ST reader - the masked interrupt status of lsc frame done

MEDIAN_FRAME_INT_ST_R

Field MEDIAN_FRAME_INT_ST reader - the masked interrupt status of median frame done

R

Register INT_ST reader

RGB2YUV_FRAME_INT_ST_R

Field RGB2YUV_FRAME_INT_ST reader - the masked interrupt status of rgb2yuv frame done

SHARP_FRAME_INT_ST_R

Field SHARP_FRAME_INT_ST reader - the masked interrupt status of sharp frame done

TAIL_IDI_FRAME_INT_ST_R

Field TAIL_IDI_FRAME_INT_ST reader - the masked interrupt status of isp_tail idi frame_end

YUV2RGB_FRAME_INT_ST_R

Field YUV2RGB_FRAME_INT_ST reader - the masked interrupt status of yuv2rgb frame done