Type Alias esp32p4::isp::int_st::R

pub type R = R<INT_ST_SPEC>;

Register INT_ST reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn isp_data_type_err_int_st(&self) -> ISP_DATA_TYPE_ERR_INT_ST_R

Bit 0 - the masked interrupt status of input data type error

pub fn isp_async_fifo_ovf_int_st(&self) -> ISP_ASYNC_FIFO_OVF_INT_ST_R

Bit 1 - the masked interrupt status of isp input fifo overflow

pub fn isp_buf_full_int_st(&self) -> ISP_BUF_FULL_INT_ST_R

Bit 2 - the masked interrupt status of isp input buffer full

pub fn isp_hvnum_setting_err_int_st(&self) -> ISP_HVNUM_SETTING_ERR_INT_ST_R

Bit 3 - the masked interrupt status of hnum and vnum setting format error

pub fn isp_data_type_setting_err_int_st( &self ) -> ISP_DATA_TYPE_SETTING_ERR_INT_ST_R

Bit 4 - the masked interrupt status of setting invalid reg_data_type

pub fn isp_mipi_hnum_unmatch_int_st(&self) -> ISP_MIPI_HNUM_UNMATCH_INT_ST_R

Bit 5 - the masked interrupt status of hnum setting unmatch with mipi input

pub fn dpc_check_done_int_st(&self) -> DPC_CHECK_DONE_INT_ST_R

Bit 6 - the masked interrupt status of dpc check done

pub fn gamma_xcoord_err_int_st(&self) -> GAMMA_XCOORD_ERR_INT_ST_R

Bit 7 - the masked interrupt status of gamma setting error

pub fn ae_monitor_int_st(&self) -> AE_MONITOR_INT_ST_R

Bit 8 - the masked interrupt status of ae monitor

pub fn ae_frame_done_int_st(&self) -> AE_FRAME_DONE_INT_ST_R

Bit 9 - the masked interrupt status of ae

pub fn af_fdone_int_st(&self) -> AF_FDONE_INT_ST_R

Bit 10 - the masked interrupt status of af statistic

pub fn af_env_int_st(&self) -> AF_ENV_INT_ST_R

Bit 11 - the masked interrupt status of af monitor

pub fn awb_fdone_int_st(&self) -> AWB_FDONE_INT_ST_R

Bit 12 - the masked interrupt status of awb

pub fn hist_fdone_int_st(&self) -> HIST_FDONE_INT_ST_R

Bit 13 - the masked interrupt status of histogram

pub fn frame_int_st(&self) -> FRAME_INT_ST_R

Bit 14 - the masked interrupt status of isp frame end

pub fn blc_frame_int_st(&self) -> BLC_FRAME_INT_ST_R

Bit 15 - the masked interrupt status of blc frame done

pub fn lsc_frame_int_st(&self) -> LSC_FRAME_INT_ST_R

Bit 16 - the masked interrupt status of lsc frame done

pub fn dpc_frame_int_st(&self) -> DPC_FRAME_INT_ST_R

Bit 17 - the masked interrupt status of dpc frame done

pub fn bf_frame_int_st(&self) -> BF_FRAME_INT_ST_R

Bit 18 - the masked interrupt status of bf frame done

pub fn demosaic_frame_int_st(&self) -> DEMOSAIC_FRAME_INT_ST_R

Bit 19 - the masked interrupt status of demosaic frame done

pub fn median_frame_int_st(&self) -> MEDIAN_FRAME_INT_ST_R

Bit 20 - the masked interrupt status of median frame done

pub fn ccm_frame_int_st(&self) -> CCM_FRAME_INT_ST_R

Bit 21 - the masked interrupt status of ccm frame done

pub fn gamma_frame_int_st(&self) -> GAMMA_FRAME_INT_ST_R

Bit 22 - the masked interrupt status of gamma frame done

pub fn rgb2yuv_frame_int_st(&self) -> RGB2YUV_FRAME_INT_ST_R

Bit 23 - the masked interrupt status of rgb2yuv frame done

pub fn sharp_frame_int_st(&self) -> SHARP_FRAME_INT_ST_R

Bit 24 - the masked interrupt status of sharp frame done

pub fn color_frame_int_st(&self) -> COLOR_FRAME_INT_ST_R

Bit 25 - the masked interrupt status of color frame done

pub fn yuv2rgb_frame_int_st(&self) -> YUV2RGB_FRAME_INT_ST_R

Bit 26 - the masked interrupt status of yuv2rgb frame done

pub fn tail_idi_frame_int_st(&self) -> TAIL_IDI_FRAME_INT_ST_R

Bit 27 - the masked interrupt status of isp_tail idi frame_end

pub fn header_idi_frame_int_st(&self) -> HEADER_IDI_FRAME_INT_ST_R

Bit 28 - the masked interrupt status of real input frame end of isp_input