Module esp32p4::jpeg::decode_conf

Control and configuration registers

Structs

DECODE_CONF_SPEC

Control and configuration registers

Type Aliases

COMPONENT_NUM_R

Field COMPONENT_NUM reader - configure number of components in frame when decode

COMPONENT_NUM_W

Field COMPONENT_NUM writer - configure number of components in frame when decode

DEZIGZAG_READY_CTL_R

Field DEZIGZAG_READY_CTL reader - reserved for decoder

DEZIGZAG_READY_CTL_W

Field DEZIGZAG_READY_CTL writer - reserved for decoder

MULTI_SCAN_ERR_CHECK_R

Field MULTI_SCAN_ERR_CHECK reader - reserved for decoder

MULTI_SCAN_ERR_CHECK_W

Field MULTI_SCAN_ERR_CHECK writer - reserved for decoder

R

Register DECODE_CONF reader

RESTART_INTERVAL_R

Field RESTART_INTERVAL reader - configure restart interval in DRI marker when decode

RESTART_INTERVAL_W

Field RESTART_INTERVAL writer - configure restart interval in DRI marker when decode

RST_CHECK_BYTE_NUM_R

Field RST_CHECK_BYTE_NUM reader - Configure the byte number to check next rst marker after one rst interval is decoded down. The real check number is reg_rst_check_byte_num+1

RST_CHECK_BYTE_NUM_W

Field RST_CHECK_BYTE_NUM writer - Configure the byte number to check next rst marker after one rst interval is decoded down. The real check number is reg_rst_check_byte_num+1

SOS_CHECK_BYTE_NUM_R

Field SOS_CHECK_BYTE_NUM reader - Configure the byte number to check next sos marker in the multi-scan picture after one scan is decoded down. The real check number is reg_sos_check_byte_num+1

SOS_CHECK_BYTE_NUM_W

Field SOS_CHECK_BYTE_NUM writer - Configure the byte number to check next sos marker in the multi-scan picture after one scan is decoded down. The real check number is reg_sos_check_byte_num+1

SW_DHT_EN_R

Field SW_DHT_EN reader - software decode dht table enable

W

Register DECODE_CONF writer