Type Alias esp32p4::jpeg::int_ena::W

pub type W = W<INT_ENA_SPEC>;

Register INT_ENA writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn done_int_ena(&mut self) -> DONE_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 0 - This enable interrupt bit turns to high level when JPEG finishes encoding a picture..

pub fn rle_parallel_err_int_ena( &mut self ) -> RLE_PARALLEL_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 1 - The enable interrupt bit to sign that rle parallel error when decoding.

pub fn cid_err_int_ena(&mut self) -> CID_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 2 - The enable interrupt bit to sign that scan id check with component fails when decoding.

pub fn c_dht_dc_id_err_int_ena( &mut self ) -> C_DHT_DC_ID_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 3 - The enable interrupt bit to sign that scan component’s dc dht id check with dc dht table’s id fails when decoding.

pub fn c_dht_ac_id_err_int_ena( &mut self ) -> C_DHT_AC_ID_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 4 - The enable interrupt bit to sign that scan component’s ac dht id check with ac dht table’s id fails when decoding.

pub fn c_dqt_id_err_int_ena( &mut self ) -> C_DQT_ID_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 5 - The enable interrupt bit to sign that scan component’s dqt id check with dqt table’s id fails when decoding.

pub fn rst_uxp_err_int_ena(&mut self) -> RST_UXP_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 6 - The enable interrupt bit to sign that RST header marker is detected but restart interval is 0 when decoding.

pub fn rst_check_none_err_int_ena( &mut self ) -> RST_CHECK_NONE_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 7 - The enable interrupt bit to sign that RST header marker is not detected but restart interval is not 0 when decoding.

pub fn rst_check_pos_err_int_ena( &mut self ) -> RST_CHECK_POS_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 8 - The enable interrupt bit to sign that RST header marker position mismatches with restart interval when decoding.

pub fn out_eof_int_ena(&mut self) -> OUT_EOF_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 9 - The enable interrupt bit turns to high level when the last pixel of one square has been transmitted for Tx channel.

pub fn sr_color_mode_err_int_ena( &mut self ) -> SR_COLOR_MODE_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 10 - The enable interrupt bit to sign that the selected source color mode is not supported.

pub fn dct_done_int_ena(&mut self) -> DCT_DONE_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 11 - The enable interrupt bit to sign that one dct calculation is finished.

pub fn bs_last_block_eof_int_ena( &mut self ) -> BS_LAST_BLOCK_EOF_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 12 - The enable interrupt bit to sign that the coding process for last block is finished.

pub fn scan_check_none_err_int_ena( &mut self ) -> SCAN_CHECK_NONE_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 13 - The enable interrupt bit to sign that SOS header marker is not detected but there are still components left to be decoded.

pub fn scan_check_pos_err_int_ena( &mut self ) -> SCAN_CHECK_POS_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 14 - The enable interrupt bit to sign that SOS header marker position wrong when decoding.

pub fn uxp_det_int_ena(&mut self) -> UXP_DET_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 15 - The enable interrupt bit to sign that unsupported header marker is detected when decoding.

pub fn en_frame_eof_err_int_ena( &mut self ) -> EN_FRAME_EOF_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 16 - The enable interrupt bit to sign that received pixel blocks are smaller than expected when encoding.

pub fn en_frame_eof_lack_int_ena( &mut self ) -> EN_FRAME_EOF_LACK_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 17 - The enable interrupt bit to sign that the frame eof sign bit from dma input is missing when encoding. But the number of pixel blocks is enough.

pub fn de_frame_eof_err_int_ena( &mut self ) -> DE_FRAME_EOF_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 18 - The enable interrupt bit to sign that decoded blocks are smaller than expected when decoding.

pub fn de_frame_eof_lack_int_ena( &mut self ) -> DE_FRAME_EOF_LACK_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 19 - The enable interrupt bit to sign that the either frame eof from dma input or eoi marker is missing when encoding. But the number of decoded blocks is enough.

pub fn sos_unmatch_err_int_ena( &mut self ) -> SOS_UNMATCH_ERR_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 20 - The enable interrupt bit to sign that the component number of a scan is 0 or does not match the sos marker’s length when decoding.

pub fn marker_err_fst_scan_int_ena( &mut self ) -> MARKER_ERR_FST_SCAN_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 21 - The enable interrupt bit to sign that the first scan has header marker error when decoding.

pub fn marker_err_other_scan_int_ena( &mut self ) -> MARKER_ERR_OTHER_SCAN_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 22 - The enable interrupt bit to sign that the following scans but not the first scan have header marker error when decoding.

pub fn undet_int_ena(&mut self) -> UNDET_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 23 - The enable interrupt bit to sign that JPEG format is not detected at the eof data of a packet when decoding.

pub fn decode_timeout_int_ena( &mut self ) -> DECODE_TIMEOUT_INT_ENA_W<'_, INT_ENA_SPEC>

Bit 24 - The enable interrupt bit to sign that decode pause time is longer than the setting decode timeout time when decoding.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual