Module esp32p4::jpeg::int_st

Interrupt status registers

Structs

INT_ST_SPEC

Interrupt status registers

Type Aliases

BS_LAST_BLOCK_EOF_INT_ST_R

Field BS_LAST_BLOCK_EOF_INT_ST reader - The status interrupt bit to sign that the coding process for last block is finished.

CID_ERR_INT_ST_R

Field CID_ERR_INT_ST reader - The status interrupt bit to sign that scan id check with component fails when decoding.

C_DHT_AC_ID_ERR_INT_ST_R

Field C_DHT_AC_ID_ERR_INT_ST reader - The status interrupt bit to sign that scan component’s ac dht id check with ac dht table’s id fails when decoding.

C_DHT_DC_ID_ERR_INT_ST_R

Field C_DHT_DC_ID_ERR_INT_ST reader - The status interrupt bit to sign that scan component’s dc dht id check with dc dht table’s id fails when decoding.

C_DQT_ID_ERR_INT_ST_R

Field C_DQT_ID_ERR_INT_ST reader - The status interrupt bit to sign that scan component’s dqt id check with dqt table’s id fails when decoding.

DCT_DONE_INT_ST_R

Field DCT_DONE_INT_ST reader - The status interrupt bit to sign that one dct calculation is finished.

DECODE_TIMEOUT_INT_ST_R

Field DECODE_TIMEOUT_INT_ST reader - The status interrupt bit to sign that decode pause time is longer than the setting decode timeout time when decoding.

DE_FRAME_EOF_ERR_INT_ST_R

Field DE_FRAME_EOF_ERR_INT_ST reader - The status interrupt bit to sign that decoded blocks are smaller than expected when decoding.

DE_FRAME_EOF_LACK_INT_ST_R

Field DE_FRAME_EOF_LACK_INT_ST reader - The status interrupt bit to sign that the either frame eof from dma input or eoi marker is missing when encoding. But the number of decoded blocks is enough.

DONE_INT_ST_R

Field DONE_INT_ST reader - This status interrupt bit turns to high level when JPEG finishes encoding a picture..

EN_FRAME_EOF_ERR_INT_ST_R

Field EN_FRAME_EOF_ERR_INT_ST reader - The status interrupt bit to sign that received pixel blocks are smaller than expected when encoding.

EN_FRAME_EOF_LACK_INT_ST_R

Field EN_FRAME_EOF_LACK_INT_ST reader - The status interrupt bit to sign that the frame eof sign bit from dma input is missing when encoding. But the number of pixel blocks is enough.

MARKER_ERR_FST_SCAN_INT_ST_R

Field MARKER_ERR_FST_SCAN_INT_ST reader - The status interrupt bit to sign that the first scan has header marker error when decoding.

MARKER_ERR_OTHER_SCAN_INT_ST_R

Field MARKER_ERR_OTHER_SCAN_INT_ST reader - The status interrupt bit to sign that the following scans but not the first scan have header marker error when decoding.

OUT_EOF_INT_ST_R

Field OUT_EOF_INT_ST reader - The status interrupt bit turns to high level when the last pixel of one square has been transmitted for Tx channel.

R

Register INT_ST reader

RLE_PARALLEL_ERR_INT_ST_R

Field RLE_PARALLEL_ERR_INT_ST reader - The status interrupt bit to sign that rle parallel error when decoding.

RST_CHECK_NONE_ERR_INT_ST_R

Field RST_CHECK_NONE_ERR_INT_ST reader - The status interrupt bit to sign that RST header marker is not detected but restart interval is not 0 when decoding.

RST_CHECK_POS_ERR_INT_ST_R

Field RST_CHECK_POS_ERR_INT_ST reader - The status interrupt bit to sign that RST header marker position mismatches with restart interval when decoding.

RST_UXP_ERR_INT_ST_R

Field RST_UXP_ERR_INT_ST reader - The status interrupt bit to sign that RST header marker is detected but restart interval is 0 when decoding.

SCAN_CHECK_NONE_ERR_INT_ST_R

Field SCAN_CHECK_NONE_ERR_INT_ST reader - The status interrupt bit to sign that SOS header marker is not detected but there are still components left to be decoded.

SCAN_CHECK_POS_ERR_INT_ST_R

Field SCAN_CHECK_POS_ERR_INT_ST reader - The status interrupt bit to sign that SOS header marker position wrong when decoding.

SOS_UNMATCH_ERR_INT_ST_R

Field SOS_UNMATCH_ERR_INT_ST reader - The status interrupt bit to sign that the component number of a scan is 0 or does not match the sos marker’s length when decoding.

SR_COLOR_MODE_ERR_INT_ST_R

Field SR_COLOR_MODE_ERR_INT_ST reader - The status interrupt bit to sign that the selected source color mode is not supported.

UNDET_INT_ST_R

Field UNDET_INT_ST reader - The status interrupt bit to sign that JPEG format is not detected at the eof data of a packet when decoding.

UXP_DET_INT_ST_R

Field UXP_DET_INT_ST reader - The status interrupt bit to sign that unsupported header marker is detected when decoding.