Type Alias esp32p4::jpeg::int_st::R

pub type R = R<INT_ST_SPEC>;

Register INT_ST reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn done_int_st(&self) -> DONE_INT_ST_R

Bit 0 - This status interrupt bit turns to high level when JPEG finishes encoding a picture..

pub fn rle_parallel_err_int_st(&self) -> RLE_PARALLEL_ERR_INT_ST_R

Bit 1 - The status interrupt bit to sign that rle parallel error when decoding.

pub fn cid_err_int_st(&self) -> CID_ERR_INT_ST_R

Bit 2 - The status interrupt bit to sign that scan id check with component fails when decoding.

pub fn c_dht_dc_id_err_int_st(&self) -> C_DHT_DC_ID_ERR_INT_ST_R

Bit 3 - The status interrupt bit to sign that scan component’s dc dht id check with dc dht table’s id fails when decoding.

pub fn c_dht_ac_id_err_int_st(&self) -> C_DHT_AC_ID_ERR_INT_ST_R

Bit 4 - The status interrupt bit to sign that scan component’s ac dht id check with ac dht table’s id fails when decoding.

pub fn c_dqt_id_err_int_st(&self) -> C_DQT_ID_ERR_INT_ST_R

Bit 5 - The status interrupt bit to sign that scan component’s dqt id check with dqt table’s id fails when decoding.

pub fn rst_uxp_err_int_st(&self) -> RST_UXP_ERR_INT_ST_R

Bit 6 - The status interrupt bit to sign that RST header marker is detected but restart interval is 0 when decoding.

pub fn rst_check_none_err_int_st(&self) -> RST_CHECK_NONE_ERR_INT_ST_R

Bit 7 - The status interrupt bit to sign that RST header marker is not detected but restart interval is not 0 when decoding.

pub fn rst_check_pos_err_int_st(&self) -> RST_CHECK_POS_ERR_INT_ST_R

Bit 8 - The status interrupt bit to sign that RST header marker position mismatches with restart interval when decoding.

pub fn out_eof_int_st(&self) -> OUT_EOF_INT_ST_R

Bit 9 - The status interrupt bit turns to high level when the last pixel of one square has been transmitted for Tx channel.

pub fn sr_color_mode_err_int_st(&self) -> SR_COLOR_MODE_ERR_INT_ST_R

Bit 10 - The status interrupt bit to sign that the selected source color mode is not supported.

pub fn dct_done_int_st(&self) -> DCT_DONE_INT_ST_R

Bit 11 - The status interrupt bit to sign that one dct calculation is finished.

pub fn bs_last_block_eof_int_st(&self) -> BS_LAST_BLOCK_EOF_INT_ST_R

Bit 12 - The status interrupt bit to sign that the coding process for last block is finished.

pub fn scan_check_none_err_int_st(&self) -> SCAN_CHECK_NONE_ERR_INT_ST_R

Bit 13 - The status interrupt bit to sign that SOS header marker is not detected but there are still components left to be decoded.

pub fn scan_check_pos_err_int_st(&self) -> SCAN_CHECK_POS_ERR_INT_ST_R

Bit 14 - The status interrupt bit to sign that SOS header marker position wrong when decoding.

pub fn uxp_det_int_st(&self) -> UXP_DET_INT_ST_R

Bit 15 - The status interrupt bit to sign that unsupported header marker is detected when decoding.

pub fn en_frame_eof_err_int_st(&self) -> EN_FRAME_EOF_ERR_INT_ST_R

Bit 16 - The status interrupt bit to sign that received pixel blocks are smaller than expected when encoding.

pub fn en_frame_eof_lack_int_st(&self) -> EN_FRAME_EOF_LACK_INT_ST_R

Bit 17 - The status interrupt bit to sign that the frame eof sign bit from dma input is missing when encoding. But the number of pixel blocks is enough.

pub fn de_frame_eof_err_int_st(&self) -> DE_FRAME_EOF_ERR_INT_ST_R

Bit 18 - The status interrupt bit to sign that decoded blocks are smaller than expected when decoding.

pub fn de_frame_eof_lack_int_st(&self) -> DE_FRAME_EOF_LACK_INT_ST_R

Bit 19 - The status interrupt bit to sign that the either frame eof from dma input or eoi marker is missing when encoding. But the number of decoded blocks is enough.

pub fn sos_unmatch_err_int_st(&self) -> SOS_UNMATCH_ERR_INT_ST_R

Bit 20 - The status interrupt bit to sign that the component number of a scan is 0 or does not match the sos marker’s length when decoding.

pub fn marker_err_fst_scan_int_st(&self) -> MARKER_ERR_FST_SCAN_INT_ST_R

Bit 21 - The status interrupt bit to sign that the first scan has header marker error when decoding.

pub fn marker_err_other_scan_int_st(&self) -> MARKER_ERR_OTHER_SCAN_INT_ST_R

Bit 22 - The status interrupt bit to sign that the following scans but not the first scan have header marker error when decoding.

pub fn undet_int_st(&self) -> UNDET_INT_ST_R

Bit 23 - The status interrupt bit to sign that JPEG format is not detected at the eof data of a packet when decoding.

pub fn decode_timeout_int_st(&self) -> DECODE_TIMEOUT_INT_ST_R

Bit 24 - The status interrupt bit to sign that decode pause time is longer than the setting decode timeout time when decoding.