Module esp32p4::lcd_cam::lcd_ctrl2

LCD config register.

Structs

LCD_CTRL2_SPEC

LCD config register.

Type Aliases

LCD_DE_IDLE_POL_R

Field LCD_DE_IDLE_POL reader - It is the idle value of LCD_DE.

LCD_DE_IDLE_POL_W

Field LCD_DE_IDLE_POL writer - It is the idle value of LCD_DE.

LCD_HSYNC_IDLE_POL_R

Field LCD_HSYNC_IDLE_POL reader - It is the idle value of LCD_HSYNC.

LCD_HSYNC_IDLE_POL_W

Field LCD_HSYNC_IDLE_POL writer - It is the idle value of LCD_HSYNC.

LCD_HSYNC_POSITION_R

Field LCD_HSYNC_POSITION reader - It is the position of LCD_HSYNC active pulse in a line.

LCD_HSYNC_POSITION_W

Field LCD_HSYNC_POSITION writer - It is the position of LCD_HSYNC active pulse in a line.

LCD_HSYNC_WIDTH_R

Field LCD_HSYNC_WIDTH reader - It is the position of LCD_HSYNC active pulse in a line.

LCD_HSYNC_WIDTH_W

Field LCD_HSYNC_WIDTH writer - It is the position of LCD_HSYNC active pulse in a line.

LCD_HS_BLANK_EN_R

Field LCD_HS_BLANK_EN reader - 1: The pulse of LCD_HSYNC is out in vertical blanking lines RGB mode. 0: LCD_HSYNC pulse is valid only in active region lines in RGB mode.

LCD_HS_BLANK_EN_W

Field LCD_HS_BLANK_EN writer - 1: The pulse of LCD_HSYNC is out in vertical blanking lines RGB mode. 0: LCD_HSYNC pulse is valid only in active region lines in RGB mode.

LCD_VSYNC_IDLE_POL_R

Field LCD_VSYNC_IDLE_POL reader - It is the idle value of LCD_VSYNC.

LCD_VSYNC_IDLE_POL_W

Field LCD_VSYNC_IDLE_POL writer - It is the idle value of LCD_VSYNC.

LCD_VSYNC_WIDTH_R

Field LCD_VSYNC_WIDTH reader - It is the position of LCD_VSYNC active pulse in a line.

LCD_VSYNC_WIDTH_W

Field LCD_VSYNC_WIDTH writer - It is the position of LCD_VSYNC active pulse in a line.

R

Register LCD_CTRL2 reader

W

Register LCD_CTRL2 writer