Module esp32p4::lcd_cam::lcd_misc

LCD config register.

Structs

LCD_MISC_SPEC

LCD config register.

Type Aliases

LCD_AFIFO_RESET_W

Field LCD_AFIFO_RESET writer - LCD AFIFO reset signal.

LCD_BK_EN_R

Field LCD_BK_EN reader - 1: Enable blank region when LCD sends data out. 0: No blank region.

LCD_BK_EN_W

Field LCD_BK_EN writer - 1: Enable blank region when LCD sends data out. 0: No blank region.

LCD_CD_CMD_SET_R

Field LCD_CD_CMD_SET reader - 1: LCD_CD = !reg_cd_idle_edge when lcd_st[2:0] is in LCD_CMD state. 0: LCD_CD = reg_cd_idle_edge.

LCD_CD_CMD_SET_W

Field LCD_CD_CMD_SET writer - 1: LCD_CD = !reg_cd_idle_edge when lcd_st[2:0] is in LCD_CMD state. 0: LCD_CD = reg_cd_idle_edge.

LCD_CD_DATA_SET_R

Field LCD_CD_DATA_SET reader - 1: LCD_CD = !reg_cd_idle_edge when lcd_st[2:0] is in LCD_DOUT state. 0: LCD_CD = reg_cd_idle_edge.

LCD_CD_DATA_SET_W

Field LCD_CD_DATA_SET writer - 1: LCD_CD = !reg_cd_idle_edge when lcd_st[2:0] is in LCD_DOUT state. 0: LCD_CD = reg_cd_idle_edge.

LCD_CD_DUMMY_SET_R

Field LCD_CD_DUMMY_SET reader - 1: LCD_CD = !reg_cd_idle_edge when lcd_st[2:0] is in LCD_DUMMY state. 0: LCD_CD = reg_cd_idle_edge.

LCD_CD_DUMMY_SET_W

Field LCD_CD_DUMMY_SET writer - 1: LCD_CD = !reg_cd_idle_edge when lcd_st[2:0] is in LCD_DUMMY state. 0: LCD_CD = reg_cd_idle_edge.

LCD_CD_IDLE_EDGE_R

Field LCD_CD_IDLE_EDGE reader - The default value of LCD_CD.

LCD_CD_IDLE_EDGE_W

Field LCD_CD_IDLE_EDGE writer - The default value of LCD_CD.

LCD_NEXT_FRAME_EN_R

Field LCD_NEXT_FRAME_EN reader - 1: Send the next frame data when the current frame is sent out. 0: LCD stops when the current frame is sent out.

LCD_NEXT_FRAME_EN_W

Field LCD_NEXT_FRAME_EN writer - 1: Send the next frame data when the current frame is sent out. 0: LCD stops when the current frame is sent out.

LCD_VBK_CYCLELEN_R

Field LCD_VBK_CYCLELEN reader - The vertical back blank region cycle length minus 1 in LCD RGB mode, or the hold time cycle length in LCD non-RGB mode.

LCD_VBK_CYCLELEN_W

Field LCD_VBK_CYCLELEN writer - The vertical back blank region cycle length minus 1 in LCD RGB mode, or the hold time cycle length in LCD non-RGB mode.

LCD_VFK_CYCLELEN_R

Field LCD_VFK_CYCLELEN reader - The setup cycle length minus 1 in LCD non-RGB mode.

LCD_VFK_CYCLELEN_W

Field LCD_VFK_CYCLELEN writer - The setup cycle length minus 1 in LCD non-RGB mode.

LCD_WIRE_MODE_R

Field LCD_WIRE_MODE reader - The wire width of LCD output. 0: 8bit. 1: 16bit 2: 24bit

LCD_WIRE_MODE_W

Field LCD_WIRE_MODE writer - The wire width of LCD output. 0: 8bit. 1: 16bit 2: 24bit

R

Register LCD_MISC reader

W

Register LCD_MISC writer