Type Alias esp32p4::lcd_cam::lcd_misc::R

pub type R = R<LCD_MISC_SPEC>;

Register LCD_MISC reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn lcd_wire_mode(&self) -> LCD_WIRE_MODE_R

Bits 4:5 - The wire width of LCD output. 0: 8bit. 1: 16bit 2: 24bit

pub fn lcd_vfk_cyclelen(&self) -> LCD_VFK_CYCLELEN_R

Bits 6:11 - The setup cycle length minus 1 in LCD non-RGB mode.

pub fn lcd_vbk_cyclelen(&self) -> LCD_VBK_CYCLELEN_R

Bits 12:24 - The vertical back blank region cycle length minus 1 in LCD RGB mode, or the hold time cycle length in LCD non-RGB mode.

pub fn lcd_next_frame_en(&self) -> LCD_NEXT_FRAME_EN_R

Bit 25 - 1: Send the next frame data when the current frame is sent out. 0: LCD stops when the current frame is sent out.

pub fn lcd_bk_en(&self) -> LCD_BK_EN_R

Bit 26 - 1: Enable blank region when LCD sends data out. 0: No blank region.

pub fn lcd_cd_data_set(&self) -> LCD_CD_DATA_SET_R

Bit 28 - 1: LCD_CD = !reg_cd_idle_edge when lcd_st[2:0] is in LCD_DOUT state. 0: LCD_CD = reg_cd_idle_edge.

pub fn lcd_cd_dummy_set(&self) -> LCD_CD_DUMMY_SET_R

Bit 29 - 1: LCD_CD = !reg_cd_idle_edge when lcd_st[2:0] is in LCD_DUMMY state. 0: LCD_CD = reg_cd_idle_edge.

pub fn lcd_cd_cmd_set(&self) -> LCD_CD_CMD_SET_R

Bit 30 - 1: LCD_CD = !reg_cd_idle_edge when lcd_st[2:0] is in LCD_CMD state. 0: LCD_CD = reg_cd_idle_edge.

pub fn lcd_cd_idle_edge(&self) -> LCD_CD_IDLE_EDGE_R

Bit 31 - The default value of LCD_CD.