Module esp32p4::lcd_cam::lcd_user

LCD config register.

Structs

LCD_USER_SPEC

LCD config register.

Type Aliases

LCD_ALWAYS_OUT_EN_R

Field LCD_ALWAYS_OUT_EN reader - LCD always output when LCD is in LCD_DOUT state, unless reg_lcd_start is cleared or reg_lcd_reset is set.

LCD_ALWAYS_OUT_EN_W

Field LCD_ALWAYS_OUT_EN writer - LCD always output when LCD is in LCD_DOUT state, unless reg_lcd_start is cleared or reg_lcd_reset is set.

LCD_BIT_ORDER_R

Field LCD_BIT_ORDER reader - 1: Change data bit order, change LCD_DATA_out[7:0] to LCD_DATA_out[0:7] in one byte mode, and bits[15:0] to bits[0:15] in two byte mode. 0: Not change.

LCD_BIT_ORDER_W

Field LCD_BIT_ORDER writer - 1: Change data bit order, change LCD_DATA_out[7:0] to LCD_DATA_out[0:7] in one byte mode, and bits[15:0] to bits[0:15] in two byte mode. 0: Not change.

LCD_BYTE_MODE_R

Field LCD_BYTE_MODE reader - 2: 24bit mode. 1: 16bit mode. 0: 8bit mode

LCD_BYTE_MODE_W

Field LCD_BYTE_MODE writer - 2: 24bit mode. 1: 16bit mode. 0: 8bit mode

LCD_BYTE_ORDER_R

Field LCD_BYTE_ORDER reader - 1: invert data byte order, only valid in 2 byte mode. 0: Not change.

LCD_BYTE_ORDER_W

Field LCD_BYTE_ORDER writer - 1: invert data byte order, only valid in 2 byte mode. 0: Not change.

LCD_CMD_2_CYCLE_EN_R

Field LCD_CMD_2_CYCLE_EN reader - The cycle length of command phase. 1: 2 cycles. 0: 1 cycle.

LCD_CMD_2_CYCLE_EN_W

Field LCD_CMD_2_CYCLE_EN writer - The cycle length of command phase. 1: 2 cycles. 0: 1 cycle.

LCD_CMD_R

Field LCD_CMD reader - 1: Be able to send command in LCD sequence when LCD starts. 0: Disable.

LCD_CMD_W

Field LCD_CMD writer - 1: Be able to send command in LCD sequence when LCD starts. 0: Disable.

LCD_DOUT_BIT_ORDER_R

Field LCD_DOUT_BIT_ORDER reader - 1: change bit order in every byte. 0: Not change.

LCD_DOUT_BIT_ORDER_W

Field LCD_DOUT_BIT_ORDER writer - 1: change bit order in every byte. 0: Not change.

LCD_DOUT_BYTE_SWIZZLE_ENABLE_R

Field LCD_DOUT_BYTE_SWIZZLE_ENABLE reader - 1: enable byte swizzle 0: disable

LCD_DOUT_BYTE_SWIZZLE_ENABLE_W

Field LCD_DOUT_BYTE_SWIZZLE_ENABLE writer - 1: enable byte swizzle 0: disable

LCD_DOUT_BYTE_SWIZZLE_MODE_R

Field LCD_DOUT_BYTE_SWIZZLE_MODE reader - 0: ABAB->BABA. 1: ABC->ACB. 2: ABC->BAC. 3: ABC->BCA. 4:ABC->CAB. 5:ABC->CBA

LCD_DOUT_BYTE_SWIZZLE_MODE_W

Field LCD_DOUT_BYTE_SWIZZLE_MODE writer - 0: ABAB->BABA. 1: ABC->ACB. 2: ABC->BAC. 3: ABC->BCA. 4:ABC->CAB. 5:ABC->CBA

LCD_DOUT_CYCLELEN_R

Field LCD_DOUT_CYCLELEN reader - The output data cycles minus 1 of LCD module.

LCD_DOUT_CYCLELEN_W

Field LCD_DOUT_CYCLELEN writer - The output data cycles minus 1 of LCD module.

LCD_DOUT_R

Field LCD_DOUT reader - 1: Be able to send data out in LCD sequence when LCD starts. 0: Disable.

LCD_DOUT_W

Field LCD_DOUT writer - 1: Be able to send data out in LCD sequence when LCD starts. 0: Disable.

LCD_DUMMY_CYCLELEN_R

Field LCD_DUMMY_CYCLELEN reader - The dummy cycle length minus 1.

LCD_DUMMY_CYCLELEN_W

Field LCD_DUMMY_CYCLELEN writer - The dummy cycle length minus 1.

LCD_DUMMY_R

Field LCD_DUMMY reader - 1: Enable DUMMY phase in LCD sequence when LCD starts. 0: Disable.

LCD_DUMMY_W

Field LCD_DUMMY writer - 1: Enable DUMMY phase in LCD sequence when LCD starts. 0: Disable.

LCD_RESET_W

Field LCD_RESET writer - The value of command.

LCD_START_R

Field LCD_START reader - LCD start sending data enable signal, valid in high level.

LCD_START_W

Field LCD_START writer - LCD start sending data enable signal, valid in high level.

LCD_UPDATE_R

Field LCD_UPDATE reader - 1: Update LCD registers, will be cleared by hardware. 0 : Not care.

LCD_UPDATE_W

Field LCD_UPDATE writer - 1: Update LCD registers, will be cleared by hardware. 0 : Not care.

R

Register LCD_USER reader

W

Register LCD_USER writer