Module esp32p4::ledc::ch_conf0

Configuration register 0 for channel %s

Structs

CH_CONF0_SPEC

Configuration register 0 for channel %s

Type Aliases

IDLE_LV_CH_R

Field IDLE_LV_CH reader - Configures the output value when channel %s is inactive. Valid only when LEDC_SIG_OUT_EN_CH%s is 0.\0: Output level is low\1: Output level is high

IDLE_LV_CH_W

Field IDLE_LV_CH writer - Configures the output value when channel %s is inactive. Valid only when LEDC_SIG_OUT_EN_CH%s is 0.\0: Output level is low\1: Output level is high

OVF_CNT_EN_CH_R

Field OVF_CNT_EN_CH reader - Configures whether or not to enable the ovf_cnt of channel %s.\0: Disable\1: Enable

OVF_CNT_EN_CH_W

Field OVF_CNT_EN_CH writer - Configures whether or not to enable the ovf_cnt of channel %s.\0: Disable\1: Enable

OVF_CNT_RESET_CH_W

Field OVF_CNT_RESET_CH writer - Configures whether or not to reset the ovf_cnt of channel %s.\0: Invalid. No effect\1: Reset the ovf_cnt

OVF_NUM_CH_R

Field OVF_NUM_CH reader - Configures the maximum times of overflow minus 1.The LEDC_OVF_CNT_CH%s_INT interrupt will be triggered when channel %s overflows for (LEDC_OVF_NUM_CH%s + 1) times.

OVF_NUM_CH_W

Field OVF_NUM_CH writer - Configures the maximum times of overflow minus 1.The LEDC_OVF_CNT_CH%s_INT interrupt will be triggered when channel %s overflows for (LEDC_OVF_NUM_CH%s + 1) times.

PARA_UP_CH_W

Field PARA_UP_CH writer - Configures whether or not to update LEDC_HPOINT_CH%s, LEDC_DUTY_START_CH%s, LEDC_SIG_OUT_EN_CH%s, LEDC_TIMER_SEL_CH%s, LEDC_DUTY_NUM_CH%s, LEDC_DUTY_CYCLE_CH%s, LEDC_DUTY_SCALE_CH%s, LEDC_DUTY_INC_CH%s, and LEDC_OVF_CNT_EN_CH%s fields for channel %s, and will be automatically cleared by hardware.\0: Invalid. No effect\1: Update

R

Register CH%s_CONF0 reader

SIG_OUT_EN_CH_R

Field SIG_OUT_EN_CH reader - Configures whether or not to enable signal output on channel %s.\0: Signal output disable\1: Signal output enable

SIG_OUT_EN_CH_W

Field SIG_OUT_EN_CH writer - Configures whether or not to enable signal output on channel %s.\0: Signal output disable\1: Signal output enable

TIMER_SEL_CH_R

Field TIMER_SEL_CH reader - Configures which timer is channel %s selected.\0: Select timer0\1: Select timer1\2: Select timer2\3: Select timer3

TIMER_SEL_CH_W

Field TIMER_SEL_CH writer - Configures which timer is channel %s selected.\0: Select timer0\1: Select timer1\2: Select timer2\3: Select timer3

W

Register CH%s_CONF0 writer