Module esp32p4::ledc::timer_conf

Timer %s configuration register

Structs

TIMER_CONF_SPEC

Timer %s configuration register

Type Aliases

CLK_DIV_TIMER_R

Field CLK_DIV_TIMER reader - Configures the divisor for the divider in timer %s.The least significant eight bits represent the fractional part.

CLK_DIV_TIMER_W

Field CLK_DIV_TIMER writer - Configures the divisor for the divider in timer %s.The least significant eight bits represent the fractional part.

R

Register TIMER%s_CONF reader

TICK_SEL_TIMER_R

Field TICK_SEL_TIMER reader - Configures which clock is timer %s selected. Unused.

TICK_SEL_TIMER_W

Field TICK_SEL_TIMER writer - Configures which clock is timer %s selected. Unused.

TIMER_DUTY_RES_R

Field TIMER_DUTY_RES reader - Configures the range of the counter in timer %s.

TIMER_DUTY_RES_W

Field TIMER_DUTY_RES writer - Configures the range of the counter in timer %s.

TIMER_PARA_UP_W

Field TIMER_PARA_UP writer - Configures whether or not to update LEDC_CLK_DIV_TIMER%s and LEDC_TIMER%s_DUTY_RES.\0: Invalid. No effect\1: Update

TIMER_PAUSE_R

Field TIMER_PAUSE reader - Configures whether or not to pause the counter in timer %s.\0: Normal\1: Pause

TIMER_PAUSE_W

Field TIMER_PAUSE writer - Configures whether or not to pause the counter in timer %s.\0: Normal\1: Pause

TIMER_RST_R

Field TIMER_RST reader - Configures whether or not to reset timer %s. The counter will show 0 after reset.\0: Not reset\1: Reset

TIMER_RST_W

Field TIMER_RST writer - Configures whether or not to reset timer %s. The counter will show 0 after reset.\0: Not reset\1: Reset

W

Register TIMER%s_CONF writer