Module esp32p4::lp_adc

Low-power Analog to Digital Converter

Modules

amp_ctrl1

N/A

amp_ctrl2

N/A

amp_ctrl3

N/A

atten1

ADC1 attenuation registers.

atten2

ADC1 attenuation registers.

cocpu_int_raw

Interrupt raw registers.

force_wpd_sar

In sleep, force to use rtc to control ADC

int_clr

Interrupt clear registers.

int_ena

Interrupt enable registers.

int_ena_w1tc

Interrupt enable deassert registers.

int_ena_w1ts

Interrupt enable assert registers.

int_st

Interrupt status registers.

meas1_ctrl1

N/A

meas1_ctrl2

ADC1 configuration registers.

meas1_mux

SAR ADC1 MUX register.

meas2_ctrl1

ADC2 configuration registers.

meas2_ctrl2

ADC2 configuration registers.

meas2_mux

SAR ADC2 MUX register.

meas_status

N/A

reader1_ctrl

Control the read operation of ADC1.

reader1_status

N/A

reader2_ctrl

Control the read operation of ADC2.

reader2_status

N/A

reg_clken

N/A

rnd_eco_cs

N/A

rnd_eco_high

N/A

rnd_eco_low

N/A

sar1_hw_wakeup

Hardware automatic sampling registers for wakeup function.

sar2_hw_wakeup

Hardware automatic sampling registers for wakeup function.

wakeup1

ADC1 wakeup configuration registers.

wakeup2

ADC2 wakeup configuration registers.

wakeup_sel

Wakeup source select register.

Structs

RegisterBlock

Register block

Type Aliases

AMP_CTRL1

AMP_CTRL1 (rw) register accessor: N/A

AMP_CTRL2

AMP_CTRL2 (rw) register accessor: N/A

AMP_CTRL3

AMP_CTRL3 (rw) register accessor: N/A

ATTEN1

ATTEN1 (rw) register accessor: ADC1 attenuation registers.

ATTEN2

ATTEN2 (rw) register accessor: ADC1 attenuation registers.

COCPU_INT_RAW

COCPU_INT_RAW (rw) register accessor: Interrupt raw registers.

FORCE_WPD_SAR

FORCE_WPD_SAR (rw) register accessor: In sleep, force to use rtc to control ADC

INT_CLR

INT_CLR (w) register accessor: Interrupt clear registers.

INT_ENA

INT_ENA (rw) register accessor: Interrupt enable registers.

INT_ENA_W1TC

INT_ENA_W1TC (w) register accessor: Interrupt enable deassert registers.

INT_ENA_W1TS

INT_ENA_W1TS (w) register accessor: Interrupt enable assert registers.

INT_ST

INT_ST (r) register accessor: Interrupt status registers.

MEAS1_CTRL1

MEAS1_CTRL1 (rw) register accessor: N/A

MEAS1_CTRL2

MEAS1_CTRL2 (rw) register accessor: ADC1 configuration registers.

MEAS1_MUX

MEAS1_MUX (rw) register accessor: SAR ADC1 MUX register.

MEAS2_CTRL1

MEAS2_CTRL1 (rw) register accessor: ADC2 configuration registers.

MEAS2_CTRL2

MEAS2_CTRL2 (rw) register accessor: ADC2 configuration registers.

MEAS2_MUX

MEAS2_MUX (rw) register accessor: SAR ADC2 MUX register.

MEAS_STATUS

MEAS_STATUS (r) register accessor: N/A

READER1_CTRL

READER1_CTRL (rw) register accessor: Control the read operation of ADC1.

READER1_STATUS

READER1_STATUS (r) register accessor: N/A

READER2_CTRL

READER2_CTRL (rw) register accessor: Control the read operation of ADC2.

READER2_STATUS

READER2_STATUS (r) register accessor: N/A

REG_CLKEN

REG_CLKEN (rw) register accessor: N/A

RND_ECO_CS

RND_ECO_CS (rw) register accessor: N/A

RND_ECO_HIGH

RND_ECO_HIGH (rw) register accessor: N/A

RND_ECO_LOW

RND_ECO_LOW (rw) register accessor: N/A

SAR1_HW_WAKEUP

SAR1_HW_WAKEUP (rw) register accessor: Hardware automatic sampling registers for wakeup function.

SAR2_HW_WAKEUP

SAR2_HW_WAKEUP (rw) register accessor: Hardware automatic sampling registers for wakeup function.

WAKEUP1

WAKEUP1 (rw) register accessor: ADC1 wakeup configuration registers.

WAKEUP2

WAKEUP2 (rw) register accessor: ADC2 wakeup configuration registers.

WAKEUP_SEL

WAKEUP_SEL (rw) register accessor: Wakeup source select register.