Module esp32p4::lp_i2c0::scl_start_hold

Configures the delay between the SDA and SCL negative edge for a start condition

Structs

SCL_START_HOLD_SPEC

Configures the delay between the SDA and SCL negative edge for a start condition

Type Aliases

R

Register SCL_START_HOLD reader

TIME_R

Field TIME reader - Configures the time between the falling edge of SDA and the falling edge of SCL for a START condition. Measurement unit: i2c_sclk.

TIME_W

Field TIME writer - Configures the time between the falling edge of SDA and the falling edge of SCL for a START condition. Measurement unit: i2c_sclk.

W

Register SCL_START_HOLD writer