Type Alias esp32p4::lp_i2s0::rx_conf1::R

pub type R = R<RX_CONF1_SPEC>;

Register RX_CONF1 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn rx_tdm_ws_width(&self) -> RX_TDM_WS_WIDTH_R

Bits 0:6 - The width of rx_ws_out in TDM mode is (I2S_RX_TDM_WS_WIDTH[6:0] +1) * T_bck

pub fn rx_bck_div_num(&self) -> RX_BCK_DIV_NUM_R

Bits 7:12 - Bit clock configuration bits in receiver mode.

pub fn rx_bits_mod(&self) -> RX_BITS_MOD_R

Bits 13:17 - Set the bits to configure the valid data bit length of I2S receiver channel. 7: all the valid channel data is in 8-bit-mode. 15: all the valid channel data is in 16-bit-mode. 23: all the valid channel data is in 24-bit-mode. 31:all the valid channel data is in 32-bit-mode.

pub fn rx_half_sample_bits(&self) -> RX_HALF_SAMPLE_BITS_R

Bits 18:23 - I2S Rx half sample bits -1.

pub fn rx_tdm_chan_bits(&self) -> RX_TDM_CHAN_BITS_R

Bits 24:28 - The Rx bit number for each channel minus 1in TDM mode.

pub fn rx_msb_shift(&self) -> RX_MSB_SHIFT_R

Bit 29 - Set this bit to enable receiver in Phillips standard mode