Type Alias esp32p4::lp_uart::conf0_sync::R

pub type R = R<CONF0_SYNC_SPEC>;

Register CONF0_SYNC reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn parity(&self) -> PARITY_R

Bit 0 - This register is used to configure the parity check mode.

pub fn parity_en(&self) -> PARITY_EN_R

Bit 1 - Set this bit to enable uart parity check.

pub fn bit_num(&self) -> BIT_NUM_R

Bits 2:3 - This register is used to set the length of data.

pub fn stop_bit_num(&self) -> STOP_BIT_NUM_R

Bits 4:5 - This register is used to set the length of stop bit.

pub fn txd_brk(&self) -> TXD_BRK_R

Bit 6 - Set this bit to enbale transmitter to send NULL when the process of sending data is done.

pub fn loopback(&self) -> LOOPBACK_R

Bit 12 - Set this bit to enable uart loopback test mode.

pub fn tx_flow_en(&self) -> TX_FLOW_EN_R

Bit 13 - Set this bit to enable flow control function for transmitter.

pub fn rxd_inv(&self) -> RXD_INV_R

Bit 15 - Set this bit to inverse the level value of uart rxd signal.

pub fn txd_inv(&self) -> TXD_INV_R

Bit 16 - Set this bit to inverse the level value of uart txd signal.

pub fn dis_rx_dat_ovf(&self) -> DIS_RX_DAT_OVF_R

Bit 17 - Disable UART Rx data overflow detect.

pub fn err_wr_mask(&self) -> ERR_WR_MASK_R

Bit 18 - 1’h1: Receiver stops storing data into FIFO when data is wrong. 1’h0: Receiver stores the data even if the received data is wrong.

pub fn mem_clk_en(&self) -> MEM_CLK_EN_R

Bit 20 - UART memory clock gate enable signal.

pub fn sw_rts(&self) -> SW_RTS_R

Bit 21 - This register is used to configure the software rts signal which is used in software flow control.

pub fn rxfifo_rst(&self) -> RXFIFO_RST_R

Bit 22 - Set this bit to reset the uart receive-FIFO.

pub fn txfifo_rst(&self) -> TXFIFO_RST_R

Bit 23 - Set this bit to reset the uart transmit-FIFO.