Type Alias esp32p4::lp_uart::conf1::R

pub type R = R<CONF1_SPEC>;

Register CONF1 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn rxfifo_full_thrhd(&self) -> RXFIFO_FULL_THRHD_R

Bits 3:7 - It will produce rxfifo_full_int interrupt when receiver receives more data than this register value.

pub fn txfifo_empty_thrhd(&self) -> TXFIFO_EMPTY_THRHD_R

Bits 11:15 - It will produce txfifo_empty_int interrupt when the data amount in Tx-FIFO is less than this register value.

pub fn cts_inv(&self) -> CTS_INV_R

Bit 16 - Set this bit to inverse the level value of uart cts signal.

pub fn dsr_inv(&self) -> DSR_INV_R

Bit 17 - Set this bit to inverse the level value of uart dsr signal.

pub fn rts_inv(&self) -> RTS_INV_R

Bit 18 - Set this bit to inverse the level value of uart rts signal.

pub fn dtr_inv(&self) -> DTR_INV_R

Bit 19 - Set this bit to inverse the level value of uart dtr signal.

pub fn sw_dtr(&self) -> SW_DTR_R

Bit 20 - This register is used to configure the software dtr signal which is used in software flow control.

pub fn clk_en(&self) -> CLK_EN_R

Bit 21 - 1’h1: Force clock on for register. 1’h0: Support clock only when application writes registers.