Module esp32p4::mcpwm0

Motor Control Pulse-Width Modulation 0

Modules

cap_ch

CAP%s capture value register

cap_ch_cfg

Capture channel %s configuration register

cap_status

Last capture trigger edge information register

cap_timer_cfg

Capture timer configuration register

cap_timer_phase

Capture timer sync phase register

carrier_cfg

Carrier%s configuration register

clk

Global configuration register

clk_cfg

PWM clock prescaler register.

dt_cfg

Dead time configuration register

dt_fed_cfg

Falling edge delay (FED) shadow register

dt_red_cfg

Rising edge delay (RED) shadow register

evt_en

Event enable register

evt_en2

Event enable register2

fault_detect

Fault detection configuration and status register

fh_cfg0

PWM%s A and PWM%s B trip events actions configuration register

fh_cfg1

Software triggers for fault handler actions configuration register

fh_status

Fault events status register

gen_a

PWM%s output signal A actions configuration register

gen_b

PWM%s output signal B actions configuration register

gen_cfg0

Generator%s fault event T0 and T1 configuration register

gen_force

Generator%s output signal force mode register.

gen_stmp_cfg

Generator%s time stamp registers A and B transfer status and update method register

gen_tstmp_a

Generator%s time stamp A’s shadow register

gen_tstmp_b

Generator%s time stamp B’s shadow register

int_clr

Interrupt clear register

int_ena

Interrupt enable register

int_raw

Interrupt raw status register

int_st

Interrupt masked status register

op_tstmp_e1

Generator%s timer stamp E1 value register

op_tstmp_e2

Generator%s timer stamp E2 value register

operator_timersel

PWM operator’s timer select register

task_en

Task enable register

timer_cfg0

PWM timer%s period and update method configuration register.

timer_cfg1

PWM timer%s working mode and start/stop control register.

timer_status

PWM timer%s status register.

timer_sync

PWM timer%s sync function configuration register.

timer_synci_cfg

Synchronization input selection register for PWM timers.

update_cfg

Generator Update configuration register

version

Version register.

Structs

RegisterBlock

Register block

Type Aliases

CAP_CH

CAP_CH (r) register accessor: CAP%s capture value register

CAP_CH_CFG

CAP_CH_CFG (rw) register accessor: Capture channel %s configuration register

CAP_STATUS

CAP_STATUS (r) register accessor: Last capture trigger edge information register

CAP_TIMER_CFG

CAP_TIMER_CFG (rw) register accessor: Capture timer configuration register

CAP_TIMER_PHASE

CAP_TIMER_PHASE (rw) register accessor: Capture timer sync phase register

CARRIER_CFG

CARRIER_CFG (rw) register accessor: Carrier%s configuration register

CLK

CLK (rw) register accessor: Global configuration register

CLK_CFG

CLK_CFG (rw) register accessor: PWM clock prescaler register.

DT_CFG

DT_CFG (rw) register accessor: Dead time configuration register

DT_FED_CFG

DT_FED_CFG (rw) register accessor: Falling edge delay (FED) shadow register

DT_RED_CFG

DT_RED_CFG (rw) register accessor: Rising edge delay (RED) shadow register

EVT_EN

EVT_EN (rw) register accessor: Event enable register

EVT_EN2

EVT_EN2 (rw) register accessor: Event enable register2

FAULT_DETECT

FAULT_DETECT (rw) register accessor: Fault detection configuration and status register

FH_CFG0

FH_CFG0 (rw) register accessor: PWM%s A and PWM%s B trip events actions configuration register

FH_CFG1

FH_CFG1 (rw) register accessor: Software triggers for fault handler actions configuration register

FH_STATUS

FH_STATUS (r) register accessor: Fault events status register

GEN_A

GEN_A (rw) register accessor: PWM%s output signal A actions configuration register

GEN_B

GEN_B (rw) register accessor: PWM%s output signal B actions configuration register

GEN_CFG0

GEN_CFG0 (rw) register accessor: Generator%s fault event T0 and T1 configuration register

GEN_FORCE

GEN_FORCE (rw) register accessor: Generator%s output signal force mode register.

GEN_STMP_CFG

GEN_STMP_CFG (rw) register accessor: Generator%s time stamp registers A and B transfer status and update method register

GEN_TSTMP_A

GEN_TSTMP_A (rw) register accessor: Generator%s time stamp A’s shadow register

GEN_TSTMP_B

GEN_TSTMP_B (rw) register accessor: Generator%s time stamp B’s shadow register

INT_CLR

INT_CLR (w) register accessor: Interrupt clear register

INT_ENA

INT_ENA (rw) register accessor: Interrupt enable register

INT_RAW

INT_RAW (rw) register accessor: Interrupt raw status register

INT_ST

INT_ST (r) register accessor: Interrupt masked status register

OPERATOR_TIMERSEL

OPERATOR_TIMERSEL (rw) register accessor: PWM operator’s timer select register

OP_TSTMP_E1

OP_TSTMP_E1 (rw) register accessor: Generator%s timer stamp E1 value register

OP_TSTMP_E2

OP_TSTMP_E2 (rw) register accessor: Generator%s timer stamp E2 value register

TASK_EN

TASK_EN (rw) register accessor: Task enable register

TIMER_CFG0

TIMER_CFG0 (rw) register accessor: PWM timer%s period and update method configuration register.

TIMER_CFG1

TIMER_CFG1 (rw) register accessor: PWM timer%s working mode and start/stop control register.

TIMER_STATUS

TIMER_STATUS (r) register accessor: PWM timer%s status register.

TIMER_SYNC

TIMER_SYNC (rw) register accessor: PWM timer%s sync function configuration register.

TIMER_SYNCI_CFG

TIMER_SYNCI_CFG (rw) register accessor: Synchronization input selection register for PWM timers.

UPDATE_CFG

UPDATE_CFG (rw) register accessor: Generator Update configuration register

VERSION

VERSION (rw) register accessor: Version register.