Module esp32p4::mcpwm0::cap_ch_cfg

Capture channel %s configuration register

Structs

CAP_CH_CFG_SPEC

Capture channel %s configuration register

Type Aliases

CAP_EN_R

Field CAP_EN reader - Configures whether or not to enable capture on channel %s.\0: Disable\1: Enable

CAP_EN_W

Field CAP_EN writer - Configures whether or not to enable capture on channel %s.\0: Disable\1: Enable

CAP_IN_INVERT_R

Field CAP_IN_INVERT reader - Configures whether or not to invert CAP%s from GPIO matrix before prescale.\0: Normal\1: Invert

CAP_IN_INVERT_W

Field CAP_IN_INVERT writer - Configures whether or not to invert CAP%s from GPIO matrix before prescale.\0: Normal\1: Invert

CAP_MODE_R

Field CAP_MODE reader - Configures which edge of capture on channel %s after prescaling is used.\0: None\Bit0 is set to 1: Rnable capture on the negative edge\Bit1 is set to 1: Enable capture on the positive edge

CAP_MODE_W

Field CAP_MODE writer - Configures which edge of capture on channel %s after prescaling is used.\0: None\Bit0 is set to 1: Rnable capture on the negative edge\Bit1 is set to 1: Enable capture on the positive edge

CAP_PRESCALE_R

Field CAP_PRESCALE reader - Configures prescale value on possitive edge of CAP%s. Prescale value = PWM_CAP%s_PRESCALE + 1

CAP_PRESCALE_W

Field CAP_PRESCALE writer - Configures prescale value on possitive edge of CAP%s. Prescale value = PWM_CAP%s_PRESCALE + 1

CAP_SW_W

Field CAP_SW writer - Configures the generation of software capture.\0: Invalid, No effect\1: Trigger a software forced capture on channel %s

R

Register CAP_CH%s_CFG reader

W

Register CAP_CH%s_CFG writer