Module esp32p4::mcpwm0::carrier_cfg

Carrier%s configuration register

Structs

CARRIER_CFG_SPEC

Carrier%s configuration register

Type Aliases

CHOPPER_DUTY_R

Field CHOPPER_DUTY reader - Configures carrier duty. Duty = PWM_CARRIER%s_DUTY / 8

CHOPPER_DUTY_W

Field CHOPPER_DUTY writer - Configures carrier duty. Duty = PWM_CARRIER%s_DUTY / 8

CHOPPER_EN_R

Field CHOPPER_EN reader - Configures whether or not to enable carrier%s.\0: Bypassed\1: Enabled

CHOPPER_EN_W

Field CHOPPER_EN writer - Configures whether or not to enable carrier%s.\0: Bypassed\1: Enabled

CHOPPER_IN_INVERT_R

Field CHOPPER_IN_INVERT reader - Configures whether or not to invert the input of PWM%s A and PWM%s B for this submodule.\0: Normal\1: Invert

CHOPPER_IN_INVERT_W

Field CHOPPER_IN_INVERT writer - Configures whether or not to invert the input of PWM%s A and PWM%s B for this submodule.\0: Normal\1: Invert

CHOPPER_OSHTWTH_R

Field CHOPPER_OSHTWTH reader - Configures width of the first pulse. Measurement unit: Periods of the carrier.

CHOPPER_OSHTWTH_W

Field CHOPPER_OSHTWTH writer - Configures width of the first pulse. Measurement unit: Periods of the carrier.

CHOPPER_OUT_INVERT_R

Field CHOPPER_OUT_INVERT reader - Configures whether or not to invert the output of PWM%s A and PWM%s B for this submodule.\0: Normal\1: Invert

CHOPPER_OUT_INVERT_W

Field CHOPPER_OUT_INVERT writer - Configures whether or not to invert the output of PWM%s A and PWM%s B for this submodule.\0: Normal\1: Invert

CHOPPER_PRESCALE_R

Field CHOPPER_PRESCALE reader - Configures the prescale value of PWM carrier%s clock (PC_clk), so that period of PC_clk = period of PWM_clk * (PWM_CARRIER%s_PRESCALE + 1)

CHOPPER_PRESCALE_W

Field CHOPPER_PRESCALE writer - Configures the prescale value of PWM carrier%s clock (PC_clk), so that period of PC_clk = period of PWM_clk * (PWM_CARRIER%s_PRESCALE + 1)

R

Register CARRIER%s_CFG reader

W

Register CARRIER%s_CFG writer