Type Alias esp32p4::mcpwm0::carrier_cfg::R

pub type R = R<CARRIER_CFG_SPEC>;

Register CARRIER%s_CFG reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn chopper_en(&self) -> CHOPPER_EN_R

Bit 0 - Configures whether or not to enable carrier%s.\0: Bypassed\1: Enabled

pub fn chopper_prescale(&self) -> CHOPPER_PRESCALE_R

Bits 1:4 - Configures the prescale value of PWM carrier%s clock (PC_clk), so that period of PC_clk = period of PWM_clk * (PWM_CARRIER%s_PRESCALE + 1)

pub fn chopper_duty(&self) -> CHOPPER_DUTY_R

Bits 5:7 - Configures carrier duty. Duty = PWM_CARRIER%s_DUTY / 8

pub fn chopper_oshtwth(&self) -> CHOPPER_OSHTWTH_R

Bits 8:11 - Configures width of the first pulse. Measurement unit: Periods of the carrier.

pub fn chopper_out_invert(&self) -> CHOPPER_OUT_INVERT_R

Bit 12 - Configures whether or not to invert the output of PWM%s A and PWM%s B for this submodule.\0: Normal\1: Invert

pub fn chopper_in_invert(&self) -> CHOPPER_IN_INVERT_R

Bit 13 - Configures whether or not to invert the input of PWM%s A and PWM%s B for this submodule.\0: Normal\1: Invert