Module esp32p4::mcpwm0::gen_force

Generator%s output signal force mode register.

Structs

GEN_FORCE_SPEC

Generator%s output signal force mode register.

Type Aliases

GEN_A_CNTUFORCE_MODE_R

Field GEN_A_CNTUFORCE_MODE reader - Configures continuous software force mode for PWM%s A.\0: Disabled\1: Low\2: High\3: Disabled

GEN_A_CNTUFORCE_MODE_W

Field GEN_A_CNTUFORCE_MODE writer - Configures continuous software force mode for PWM%s A.\0: Disabled\1: Low\2: High\3: Disabled

GEN_A_NCIFORCE_MODE_R

Field GEN_A_NCIFORCE_MODE reader - Configures non-continuous immediate software force mode for PWM%s A.\0: Disabled\1: Low\2: High\3: Disabled

GEN_A_NCIFORCE_MODE_W

Field GEN_A_NCIFORCE_MODE writer - Configures non-continuous immediate software force mode for PWM%s A.\0: Disabled\1: Low\2: High\3: Disabled

GEN_A_NCIFORCE_R

Field GEN_A_NCIFORCE reader - Configures the generation of non-continuous immediate software-force event for PWM%s A, a toggle will trigger a force event.

GEN_A_NCIFORCE_W

Field GEN_A_NCIFORCE writer - Configures the generation of non-continuous immediate software-force event for PWM%s A, a toggle will trigger a force event.

GEN_B_CNTUFORCE_MODE_R

Field GEN_B_CNTUFORCE_MODE reader - Configures continuous software force mode for PWM%s B.\0: Disabled\1: Low\2: High\3: Disabled

GEN_B_CNTUFORCE_MODE_W

Field GEN_B_CNTUFORCE_MODE writer - Configures continuous software force mode for PWM%s B.\0: Disabled\1: Low\2: High\3: Disabled

GEN_B_NCIFORCE_MODE_R

Field GEN_B_NCIFORCE_MODE reader - Configures non-continuous immediate software force mode for PWM%s B.\0: Disabled\1: Low\2: High\3: Disabled

GEN_B_NCIFORCE_MODE_W

Field GEN_B_NCIFORCE_MODE writer - Configures non-continuous immediate software force mode for PWM%s B.\0: Disabled\1: Low\2: High\3: Disabled

GEN_B_NCIFORCE_R

Field GEN_B_NCIFORCE reader - Configures the generation of non-continuous immediate software-force event for PWM%s B, a toggle will trigger a force event.

GEN_B_NCIFORCE_W

Field GEN_B_NCIFORCE writer - Configures the generation of non-continuous immediate software-force event for PWM%s B, a toggle will trigger a force event.

GEN_CNTUFORCE_UPMETHOD_R

Field GEN_CNTUFORCE_UPMETHOD reader - Configures update method for continuous software force of PWM generator%s.\0: Immediately\Bit0 is set to 1: TEZ\Bit1 is set to 1: TEP\Bit2 is set to 1: TEA\Bit3 is set to 1: TEB\Bit4 is set to 1: Sync\Bit5 is set to 1: Disable update. TEA/B here and below means an event generated when the timer’s value equals to that of register A/B.

GEN_CNTUFORCE_UPMETHOD_W

Field GEN_CNTUFORCE_UPMETHOD writer - Configures update method for continuous software force of PWM generator%s.\0: Immediately\Bit0 is set to 1: TEZ\Bit1 is set to 1: TEP\Bit2 is set to 1: TEA\Bit3 is set to 1: TEB\Bit4 is set to 1: Sync\Bit5 is set to 1: Disable update. TEA/B here and below means an event generated when the timer’s value equals to that of register A/B.

R

Register GEN%s_FORCE reader

W

Register GEN%s_FORCE writer