Type Alias esp32p4::mcpwm0::int_clr::W

pub type W = W<INT_CLR_SPEC>;

Register INT_CLR writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn timer0_stop_int_clr(&mut self) -> TIMER0_STOP_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 0 - Clear bit: Write 1 to clear the interrupt triggered when the timer 0 stops.

pub fn timer1_stop_int_clr(&mut self) -> TIMER1_STOP_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 1 - Clear bit: Write 1 to clear the interrupt triggered when the timer 1 stops.

pub fn timer2_stop_int_clr(&mut self) -> TIMER2_STOP_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 2 - Clear bit: Write 1 to clear the interrupt triggered when the timer 2 stops.

pub fn timer0_tez_int_clr(&mut self) -> TIMER0_TEZ_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 3 - Clear bit: Write 1 to clear the interrupt triggered by a PWM timer 0 TEZ event.

pub fn timer1_tez_int_clr(&mut self) -> TIMER1_TEZ_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 4 - Clear bit: Write 1 to clear the interrupt triggered by a PWM timer 1 TEZ event.

pub fn timer2_tez_int_clr(&mut self) -> TIMER2_TEZ_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 5 - Clear bit: Write 1 to clear the interrupt triggered by a PWM timer 2 TEZ event.

pub fn timer0_tep_int_clr(&mut self) -> TIMER0_TEP_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 6 - Clear bit: Write 1 to clear the interrupt triggered by a PWM timer 0 TEP event.

pub fn timer1_tep_int_clr(&mut self) -> TIMER1_TEP_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 7 - Clear bit: Write 1 to clear the interrupt triggered by a PWM timer 1 TEP event.

pub fn timer2_tep_int_clr(&mut self) -> TIMER2_TEP_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 8 - Clear bit: Write 1 to clear the interrupt triggered by a PWM timer 2 TEP event.

pub fn fault0_int_clr(&mut self) -> FAULT0_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 9 - Clear bit: Write 1 to clear the interrupt triggered when event_f0 starts.

pub fn fault1_int_clr(&mut self) -> FAULT1_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 10 - Clear bit: Write 1 to clear the interrupt triggered when event_f1 starts.

pub fn fault2_int_clr(&mut self) -> FAULT2_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 11 - Clear bit: Write 1 to clear the interrupt triggered when event_f2 starts.

pub fn fault0_clr_int_clr(&mut self) -> FAULT0_CLR_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 12 - Clear bit: Write 1 to clear the interrupt triggered when event_f0 clears.

pub fn fault1_clr_int_clr(&mut self) -> FAULT1_CLR_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 13 - Clear bit: Write 1 to clear the interrupt triggered when event_f1 clears.

pub fn fault2_clr_int_clr(&mut self) -> FAULT2_CLR_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 14 - Clear bit: Write 1 to clear the interrupt triggered when event_f2 clears.

pub fn cmpr0_tea_int_clr(&mut self) -> CMPR0_TEA_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 15 - Clear bit: Write 1 to clear the interrupt triggered by a PWM operator 0 TEA event

pub fn cmpr1_tea_int_clr(&mut self) -> CMPR1_TEA_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 16 - Clear bit: Write 1 to clear the interrupt triggered by a PWM operator 1 TEA event

pub fn cmpr2_tea_int_clr(&mut self) -> CMPR2_TEA_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 17 - Clear bit: Write 1 to clear the interrupt triggered by a PWM operator 2 TEA event

pub fn cmpr0_teb_int_clr(&mut self) -> CMPR0_TEB_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 18 - Clear bit: Write 1 to clear the interrupt triggered by a PWM operator 0 TEB event

pub fn cmpr1_teb_int_clr(&mut self) -> CMPR1_TEB_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 19 - Clear bit: Write 1 to clear the interrupt triggered by a PWM operator 1 TEB event

pub fn cmpr2_teb_int_clr(&mut self) -> CMPR2_TEB_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 20 - Clear bit: Write 1 to clear the interrupt triggered by a PWM operator 2 TEB event

pub fn tz0_cbc_int_clr(&mut self) -> TZ0_CBC_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 21 - Clear bit: Write 1 to clear the interrupt triggered by a cycle-by-cycle mode action on PWM0.

pub fn tz1_cbc_int_clr(&mut self) -> TZ1_CBC_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 22 - Clear bit: Write 1 to clear the interrupt triggered by a cycle-by-cycle mode action on PWM1.

pub fn tz2_cbc_int_clr(&mut self) -> TZ2_CBC_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 23 - Clear bit: Write 1 to clear the interrupt triggered by a cycle-by-cycle mode action on PWM2.

pub fn tz0_ost_int_clr(&mut self) -> TZ0_OST_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 24 - Clear bit: Write 1 to clear the interrupt triggered by a one-shot mode action on PWM0.

pub fn tz1_ost_int_clr(&mut self) -> TZ1_OST_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 25 - Clear bit: Write 1 to clear the interrupt triggered by a one-shot mode action on PWM1.

pub fn tz2_ost_int_clr(&mut self) -> TZ2_OST_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 26 - Clear bit: Write 1 to clear the interrupt triggered by a one-shot mode action on PWM2.

pub fn cap0_int_clr(&mut self) -> CAP0_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 27 - Clear bit: Write 1 to clear the interrupt triggered by capture on CAP0.

pub fn cap1_int_clr(&mut self) -> CAP1_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 28 - Clear bit: Write 1 to clear the interrupt triggered by capture on CAP1.

pub fn cap2_int_clr(&mut self) -> CAP2_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 29 - Clear bit: Write 1 to clear the interrupt triggered by capture on CAP2.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual