Module esp32p4::mcpwm0::int_ena

Interrupt enable register

Structs

INT_ENA_SPEC

Interrupt enable register

Type Aliases

CAP0_INT_ENA_R

Field CAP0_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by capture on CAP0.

CAP0_INT_ENA_W

Field CAP0_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by capture on CAP0.

CAP1_INT_ENA_R

Field CAP1_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by capture on CAP1.

CAP1_INT_ENA_W

Field CAP1_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by capture on CAP1.

CAP2_INT_ENA_R

Field CAP2_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by capture on CAP2.

CAP2_INT_ENA_W

Field CAP2_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by capture on CAP2.

CMPR0_TEA_INT_ENA_R

Field CMPR0_TEA_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a PWM operator 0 TEA event.

CMPR0_TEA_INT_ENA_W

Field CMPR0_TEA_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a PWM operator 0 TEA event.

CMPR0_TEB_INT_ENA_R

Field CMPR0_TEB_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a PWM operator 0 TEB event.

CMPR0_TEB_INT_ENA_W

Field CMPR0_TEB_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a PWM operator 0 TEB event.

CMPR1_TEA_INT_ENA_R

Field CMPR1_TEA_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a PWM operator 1 TEA event.

CMPR1_TEA_INT_ENA_W

Field CMPR1_TEA_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a PWM operator 1 TEA event.

CMPR1_TEB_INT_ENA_R

Field CMPR1_TEB_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a PWM operator 1 TEB event.

CMPR1_TEB_INT_ENA_W

Field CMPR1_TEB_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a PWM operator 1 TEB event.

CMPR2_TEA_INT_ENA_R

Field CMPR2_TEA_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a PWM operator 2 TEA event.

CMPR2_TEA_INT_ENA_W

Field CMPR2_TEA_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a PWM operator 2 TEA event.

CMPR2_TEB_INT_ENA_R

Field CMPR2_TEB_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a PWM operator 2 TEB event.

CMPR2_TEB_INT_ENA_W

Field CMPR2_TEB_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a PWM operator 2 TEB event.

FAULT0_CLR_INT_ENA_R

Field FAULT0_CLR_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered when event_f0 clears.

FAULT0_CLR_INT_ENA_W

Field FAULT0_CLR_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered when event_f0 clears.

FAULT0_INT_ENA_R

Field FAULT0_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered when event_f0 starts.

FAULT0_INT_ENA_W

Field FAULT0_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered when event_f0 starts.

FAULT1_CLR_INT_ENA_R

Field FAULT1_CLR_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered when event_f1 clears.

FAULT1_CLR_INT_ENA_W

Field FAULT1_CLR_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered when event_f1 clears.

FAULT1_INT_ENA_R

Field FAULT1_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered when event_f1 starts.

FAULT1_INT_ENA_W

Field FAULT1_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered when event_f1 starts.

FAULT2_CLR_INT_ENA_R

Field FAULT2_CLR_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered when event_f2 clears.

FAULT2_CLR_INT_ENA_W

Field FAULT2_CLR_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered when event_f2 clears.

FAULT2_INT_ENA_R

Field FAULT2_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered when event_f2 starts.

FAULT2_INT_ENA_W

Field FAULT2_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered when event_f2 starts.

R

Register INT_ENA reader

TIMER0_STOP_INT_ENA_R

Field TIMER0_STOP_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered when the timer 0 stops.

TIMER0_STOP_INT_ENA_W

Field TIMER0_STOP_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered when the timer 0 stops.

TIMER0_TEP_INT_ENA_R

Field TIMER0_TEP_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a PWM timer 0 TEP event.

TIMER0_TEP_INT_ENA_W

Field TIMER0_TEP_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a PWM timer 0 TEP event.

TIMER0_TEZ_INT_ENA_R

Field TIMER0_TEZ_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a PWM timer 0 TEZ event.

TIMER0_TEZ_INT_ENA_W

Field TIMER0_TEZ_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a PWM timer 0 TEZ event.

TIMER1_STOP_INT_ENA_R

Field TIMER1_STOP_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered when the timer 1 stops.

TIMER1_STOP_INT_ENA_W

Field TIMER1_STOP_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered when the timer 1 stops.

TIMER1_TEP_INT_ENA_R

Field TIMER1_TEP_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a PWM timer 1 TEP event.

TIMER1_TEP_INT_ENA_W

Field TIMER1_TEP_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a PWM timer 1 TEP event.

TIMER1_TEZ_INT_ENA_R

Field TIMER1_TEZ_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a PWM timer 1 TEZ event.

TIMER1_TEZ_INT_ENA_W

Field TIMER1_TEZ_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a PWM timer 1 TEZ event.

TIMER2_STOP_INT_ENA_R

Field TIMER2_STOP_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered when the timer 2 stops.

TIMER2_STOP_INT_ENA_W

Field TIMER2_STOP_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered when the timer 2 stops.

TIMER2_TEP_INT_ENA_R

Field TIMER2_TEP_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a PWM timer 2 TEP event.

TIMER2_TEP_INT_ENA_W

Field TIMER2_TEP_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a PWM timer 2 TEP event.

TIMER2_TEZ_INT_ENA_R

Field TIMER2_TEZ_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a PWM timer 2 TEZ event.

TIMER2_TEZ_INT_ENA_W

Field TIMER2_TEZ_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a PWM timer 2 TEZ event.

TZ0_CBC_INT_ENA_R

Field TZ0_CBC_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a cycle-by-cycle mode action on PWM0.

TZ0_CBC_INT_ENA_W

Field TZ0_CBC_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a cycle-by-cycle mode action on PWM0.

TZ0_OST_INT_ENA_R

Field TZ0_OST_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a one-shot mode action on PWM0.

TZ0_OST_INT_ENA_W

Field TZ0_OST_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a one-shot mode action on PWM0.

TZ1_CBC_INT_ENA_R

Field TZ1_CBC_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a cycle-by-cycle mode action on PWM1.

TZ1_CBC_INT_ENA_W

Field TZ1_CBC_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a cycle-by-cycle mode action on PWM1.

TZ1_OST_INT_ENA_R

Field TZ1_OST_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a one-shot mode action on PWM1.

TZ1_OST_INT_ENA_W

Field TZ1_OST_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a one-shot mode action on PWM1.

TZ2_CBC_INT_ENA_R

Field TZ2_CBC_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a cycle-by-cycle mode action on PWM2.

TZ2_CBC_INT_ENA_W

Field TZ2_CBC_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a cycle-by-cycle mode action on PWM2.

TZ2_OST_INT_ENA_R

Field TZ2_OST_INT_ENA reader - Enable bit: Write 1 to enable the interrupt triggered by a one-shot mode action on PWM2.

TZ2_OST_INT_ENA_W

Field TZ2_OST_INT_ENA writer - Enable bit: Write 1 to enable the interrupt triggered by a one-shot mode action on PWM2.

W

Register INT_ENA writer