Type Alias esp32p4::mcpwm0::int_raw::R

pub type R = R<INT_RAW_SPEC>;

Register INT_RAW reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn timer0_stop_int_raw(&self) -> TIMER0_STOP_INT_RAW_R

Bit 0 - Raw status bit: The raw interrupt status of the interrupt triggered when the timer 0 stops.

pub fn timer1_stop_int_raw(&self) -> TIMER1_STOP_INT_RAW_R

Bit 1 - Raw status bit: The raw interrupt status of the interrupt triggered when the timer 1 stops.

pub fn timer2_stop_int_raw(&self) -> TIMER2_STOP_INT_RAW_R

Bit 2 - Raw status bit: The raw interrupt status of the interrupt triggered when the timer 2 stops.

pub fn timer0_tez_int_raw(&self) -> TIMER0_TEZ_INT_RAW_R

Bit 3 - Raw status bit: The raw interrupt status of the interrupt triggered by a PWM timer 0 TEZ event.

pub fn timer1_tez_int_raw(&self) -> TIMER1_TEZ_INT_RAW_R

Bit 4 - Raw status bit: The raw interrupt status of the interrupt triggered by a PWM timer 1 TEZ event.

pub fn timer2_tez_int_raw(&self) -> TIMER2_TEZ_INT_RAW_R

Bit 5 - Raw status bit: The raw interrupt status of the interrupt triggered by a PWM timer 2 TEZ event.

pub fn timer0_tep_int_raw(&self) -> TIMER0_TEP_INT_RAW_R

Bit 6 - Raw status bit: The raw interrupt status of the interrupt triggered by a PWM timer 0 TEP event.

pub fn timer1_tep_int_raw(&self) -> TIMER1_TEP_INT_RAW_R

Bit 7 - Raw status bit: The raw interrupt status of the interrupt triggered by a PWM timer 1 TEP event.

pub fn timer2_tep_int_raw(&self) -> TIMER2_TEP_INT_RAW_R

Bit 8 - Raw status bit: The raw interrupt status of the interrupt triggered by a PWM timer 2 TEP event.

pub fn fault0_int_raw(&self) -> FAULT0_INT_RAW_R

Bit 9 - Raw status bit: The raw interrupt status of the interrupt triggered when event_f0 starts.

pub fn fault1_int_raw(&self) -> FAULT1_INT_RAW_R

Bit 10 - Raw status bit: The raw interrupt status of the interrupt triggered when event_f1 starts.

pub fn fault2_int_raw(&self) -> FAULT2_INT_RAW_R

Bit 11 - Raw status bit: The raw interrupt status of the interrupt triggered when event_f2 starts.

pub fn fault0_clr_int_raw(&self) -> FAULT0_CLR_INT_RAW_R

Bit 12 - Raw status bit: The raw interrupt status of the interrupt triggered when event_f0 clears.

pub fn fault1_clr_int_raw(&self) -> FAULT1_CLR_INT_RAW_R

Bit 13 - Raw status bit: The raw interrupt status of the interrupt triggered when event_f1 clears.

pub fn fault2_clr_int_raw(&self) -> FAULT2_CLR_INT_RAW_R

Bit 14 - Raw status bit: The raw interrupt status of the interrupt triggered when event_f2 clears.

pub fn cmpr0_tea_int_raw(&self) -> CMPR0_TEA_INT_RAW_R

Bit 15 - Raw status bit: The raw interrupt status of the interrupt triggered by a PWM operator 0 TEA event

pub fn cmpr1_tea_int_raw(&self) -> CMPR1_TEA_INT_RAW_R

Bit 16 - Raw status bit: The raw interrupt status of the interrupt triggered by a PWM operator 1 TEA event

pub fn cmpr2_tea_int_raw(&self) -> CMPR2_TEA_INT_RAW_R

Bit 17 - Raw status bit: The raw interrupt status of the interrupt triggered by a PWM operator 2 TEA event

pub fn cmpr0_teb_int_raw(&self) -> CMPR0_TEB_INT_RAW_R

Bit 18 - Raw status bit: The raw interrupt status of the interrupt triggered by a PWM operator 0 TEB event

pub fn cmpr1_teb_int_raw(&self) -> CMPR1_TEB_INT_RAW_R

Bit 19 - Raw status bit: The raw interrupt status of the interrupt triggered by a PWM operator 1 TEB event

pub fn cmpr2_teb_int_raw(&self) -> CMPR2_TEB_INT_RAW_R

Bit 20 - Raw status bit: The raw interrupt status of the interrupt triggered by a PWM operator 2 TEB event

pub fn tz0_cbc_int_raw(&self) -> TZ0_CBC_INT_RAW_R

Bit 21 - Raw status bit: The raw interrupt status of the interrupt triggered by a cycle-by-cycle mode action on PWM0.

pub fn tz1_cbc_int_raw(&self) -> TZ1_CBC_INT_RAW_R

Bit 22 - Raw status bit: The raw interrupt status of the interrupt triggered by a cycle-by-cycle mode action on PWM1.

pub fn tz2_cbc_int_raw(&self) -> TZ2_CBC_INT_RAW_R

Bit 23 - Raw status bit: The raw interrupt status of the interrupt triggered by a cycle-by-cycle mode action on PWM2.

pub fn tz0_ost_int_raw(&self) -> TZ0_OST_INT_RAW_R

Bit 24 - Raw status bit: The raw interrupt status of the interrupt triggered by a one-shot mode action on PWM0.

pub fn tz1_ost_int_raw(&self) -> TZ1_OST_INT_RAW_R

Bit 25 - Raw status bit: The raw interrupt status of the interrupt triggered by a one-shot mode action on PWM1.

pub fn tz2_ost_int_raw(&self) -> TZ2_OST_INT_RAW_R

Bit 26 - Raw status bit: The raw interrupt status of the interrupt triggered by a one-shot mode action on PWM2.

pub fn cap0_int_raw(&self) -> CAP0_INT_RAW_R

Bit 27 - Raw status bit: The raw interrupt status of the interrupt triggered by capture on CAP0.

pub fn cap1_int_raw(&self) -> CAP1_INT_RAW_R

Bit 28 - Raw status bit: The raw interrupt status of the interrupt triggered by capture on CAP1.

pub fn cap2_int_raw(&self) -> CAP2_INT_RAW_R

Bit 29 - Raw status bit: The raw interrupt status of the interrupt triggered by capture on CAP2.