Type Alias esp32p4::pcnt::ctrl::W

pub type W = W<CTRL_SPEC>;

Register CTRL writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn pulse_cnt_rst_u0(&mut self) -> PULSE_CNT_RST_U0_W<'_, CTRL_SPEC>

Bit 0 - Set this bit to clear unit 0’s counter.

pub fn cnt_pause_u0(&mut self) -> CNT_PAUSE_U0_W<'_, CTRL_SPEC>

Bit 1 - Set this bit to freeze unit 0’s counter.

pub fn pulse_cnt_rst_u1(&mut self) -> PULSE_CNT_RST_U1_W<'_, CTRL_SPEC>

Bit 2 - Set this bit to clear unit 1’s counter.

pub fn cnt_pause_u1(&mut self) -> CNT_PAUSE_U1_W<'_, CTRL_SPEC>

Bit 3 - Set this bit to freeze unit 1’s counter.

pub fn pulse_cnt_rst_u2(&mut self) -> PULSE_CNT_RST_U2_W<'_, CTRL_SPEC>

Bit 4 - Set this bit to clear unit 2’s counter.

pub fn cnt_pause_u2(&mut self) -> CNT_PAUSE_U2_W<'_, CTRL_SPEC>

Bit 5 - Set this bit to freeze unit 2’s counter.

pub fn pulse_cnt_rst_u3(&mut self) -> PULSE_CNT_RST_U3_W<'_, CTRL_SPEC>

Bit 6 - Set this bit to clear unit 3’s counter.

pub fn cnt_pause_u3(&mut self) -> CNT_PAUSE_U3_W<'_, CTRL_SPEC>

Bit 7 - Set this bit to freeze unit 3’s counter.

pub fn dalta_change_en_u0(&mut self) -> DALTA_CHANGE_EN_U0_W<'_, CTRL_SPEC>

Bit 8 - Configures this bit to enable unit 0’s step comparator.

pub fn dalta_change_en_u1(&mut self) -> DALTA_CHANGE_EN_U1_W<'_, CTRL_SPEC>

Bit 9 - Configures this bit to enable unit 1’s step comparator.

pub fn dalta_change_en_u2(&mut self) -> DALTA_CHANGE_EN_U2_W<'_, CTRL_SPEC>

Bit 10 - Configures this bit to enable unit 2’s step comparator.

pub fn dalta_change_en_u3(&mut self) -> DALTA_CHANGE_EN_U3_W<'_, CTRL_SPEC>

Bit 11 - Configures this bit to enable unit 3’s step comparator.

pub fn clk_en(&mut self) -> CLK_EN_W<'_, CTRL_SPEC>

Bit 16 - The registers clock gate enable signal of PCNT module. 1: the registers can be read and written by application. 0: the registers can not be read or written by application

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual