Module esp32p4::ppa

PPA Peripheral

Modules

blend0_clut_data

CLUT sram data read/write register in background plane of blender

blend1_clut_data

CLUT sram data read/write register in foreground plane of blender

blend_byte_order

Blending engine byte order register

blend_color_mode

blending engine color mode register

blend_fix_alpha

Blending engine alpha override register

blend_fix_pixel

Blending engine fix pixel register

blend_rgb

RGB color register

blend_st

Blending engine status register

blend_trans_mode

Blending engine mode configure register

blend_tx_size

Fix pixel filling mode image size register

ck_bg_high

background color key higher threshold

ck_bg_low

background color key lower threshold

ck_default

default value when foreground and background both in color key range

ck_fg_high

foreground color key higher threshold

ck_fg_low

foreground color key lower threshold

clut_cnt

BLEND CLUT write counter register

clut_conf

CLUT configure register

date

PPA Version register

eco_cell_ctrl

Reserved.

eco_high

Reserved.

eco_low

Reserved.

int_clr

Interrupt clear bits

int_ena

Interrupt enable bits

int_raw

Raw status interrupt

int_st

Masked interrupt

reg_conf

Register clock enable register

sr_byte_order

Scaling and rotating engine byte order register

sr_color_mode

Scaling and rotating engine color mode register

sr_fix_alpha

Scaling and rotating engine alpha override register

sr_mem_pd

SR memory power done register

sr_param_err_st

Scaling and rotating coefficient error register

sr_scal_rotate

Scaling and rotating coefficient register

sr_status

SR FSM register

sram_ctrl

PPA SRAM Control Register

Structs

RegisterBlock

Register block

Type Aliases

BLEND0_CLUT_DATA

BLEND0_CLUT_DATA (rw) register accessor: CLUT sram data read/write register in background plane of blender

BLEND1_CLUT_DATA

BLEND1_CLUT_DATA (rw) register accessor: CLUT sram data read/write register in foreground plane of blender

BLEND_BYTE_ORDER

BLEND_BYTE_ORDER (rw) register accessor: Blending engine byte order register

BLEND_COLOR_MODE

BLEND_COLOR_MODE (rw) register accessor: blending engine color mode register

BLEND_FIX_ALPHA

BLEND_FIX_ALPHA (rw) register accessor: Blending engine alpha override register

BLEND_FIX_PIXEL

BLEND_FIX_PIXEL (rw) register accessor: Blending engine fix pixel register

BLEND_RGB

BLEND_RGB (rw) register accessor: RGB color register

BLEND_ST

BLEND_ST (r) register accessor: Blending engine status register

BLEND_TRANS_MODE

BLEND_TRANS_MODE (rw) register accessor: Blending engine mode configure register

BLEND_TX_SIZE

BLEND_TX_SIZE (rw) register accessor: Fix pixel filling mode image size register

CK_BG_HIGH

CK_BG_HIGH (rw) register accessor: background color key higher threshold

CK_BG_LOW

CK_BG_LOW (rw) register accessor: background color key lower threshold

CK_DEFAULT

CK_DEFAULT (rw) register accessor: default value when foreground and background both in color key range

CK_FG_HIGH

CK_FG_HIGH (rw) register accessor: foreground color key higher threshold

CK_FG_LOW

CK_FG_LOW (rw) register accessor: foreground color key lower threshold

CLUT_CNT

CLUT_CNT (r) register accessor: BLEND CLUT write counter register

CLUT_CONF

CLUT_CONF (rw) register accessor: CLUT configure register

DATE

DATE (rw) register accessor: PPA Version register

ECO_CELL_CTRL

ECO_CELL_CTRL (rw) register accessor: Reserved.

ECO_HIGH

ECO_HIGH (rw) register accessor: Reserved.

ECO_LOW

ECO_LOW (rw) register accessor: Reserved.

INT_CLR

INT_CLR (w) register accessor: Interrupt clear bits

INT_ENA

INT_ENA (rw) register accessor: Interrupt enable bits

INT_RAW

INT_RAW (rw) register accessor: Raw status interrupt

INT_ST

INT_ST (r) register accessor: Masked interrupt

REG_CONF

REG_CONF (rw) register accessor: Register clock enable register

SRAM_CTRL

SRAM_CTRL (rw) register accessor: PPA SRAM Control Register

SR_BYTE_ORDER

SR_BYTE_ORDER (rw) register accessor: Scaling and rotating engine byte order register

SR_COLOR_MODE

SR_COLOR_MODE (rw) register accessor: Scaling and rotating engine color mode register

SR_FIX_ALPHA

SR_FIX_ALPHA (rw) register accessor: Scaling and rotating engine alpha override register

SR_MEM_PD

SR_MEM_PD (rw) register accessor: SR memory power done register

SR_PARAM_ERR_ST

SR_PARAM_ERR_ST (r) register accessor: Scaling and rotating coefficient error register

SR_SCAL_ROTATE

SR_SCAL_ROTATE (rw) register accessor: Scaling and rotating coefficient register

SR_STATUS

SR_STATUS (r) register accessor: SR FSM register