Type Alias esp32p4::ppa::sr_color_mode::W

pub type W = W<SR_COLOR_MODE_SPEC>;

Register SR_COLOR_MODE writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn sr_rx_cm(&mut self) -> SR_RX_CM_W<'_, SR_COLOR_MODE_SPEC>

Bits 0:3 - The source image color mode for Scaling and Rotating engine Rx. 0: ARGB8888. 1: RGB888. 2: RGB565. 8: YUV420. others: Reserved.

pub fn sr_tx_cm(&mut self) -> SR_TX_CM_W<'_, SR_COLOR_MODE_SPEC>

Bits 4:7 - The destination image color mode for Scaling and Rotating engine Tx. 0: ARGB8888. 1: RGB888. 2: RGB565. 8: YUV420. others: Reserved.

pub fn yuv_rx_range(&mut self) -> YUV_RX_RANGE_W<'_, SR_COLOR_MODE_SPEC>

Bit 8 - YUV input range when reg_sr_rx_cm is 4’d8. 0: limit range. 1: full range

pub fn yuv_tx_range(&mut self) -> YUV_TX_RANGE_W<'_, SR_COLOR_MODE_SPEC>

Bit 9 - YUV output range when reg_sr_tx_cm is 4’d8. 0: limit range. 1: full range

pub fn yuv2rgb_protocal(&mut self) -> YUV2RGB_PROTOCAL_W<'_, SR_COLOR_MODE_SPEC>

Bit 10 - YUV to RGB protocal when reg_sr_rx_cm is 4’d8. 0: BT601. 1: BT709

pub fn rgb2yuv_protocal(&mut self) -> RGB2YUV_PROTOCAL_W<'_, SR_COLOR_MODE_SPEC>

Bit 11 - RGB to YUV protocal when reg_sr_tx_cm is 4’d8. 0: BT601. 1: BT709

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual