Type Alias esp32p4::ppa::sr_param_err_st::R

pub type R = R<SR_PARAM_ERR_ST_SPEC>;

Register SR_PARAM_ERR_ST reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn tx_dscr_vb_err_st(&self) -> TX_DSCR_VB_ERR_ST_R

Bit 0 - The error is that the scaled VB plus the offset of Y coordinate in 2DDMA receive descriptor is larger than VA in 2DDMA receive descriptor.

pub fn tx_dscr_hb_err_st(&self) -> TX_DSCR_HB_ERR_ST_R

Bit 1 - The error is that the scaled HB plus the offset of X coordinate in 2DDMA receive descriptor is larger than HA in 2DDMA receive descriptor.

pub fn y_rx_scal_equal_0_err_st(&self) -> Y_RX_SCAL_EQUAL_0_ERR_ST_R

Bit 2 - The error is that the PPA_SR_SCAL_Y_INT and PPA_SR_CAL_Y_FRAG both are 0.

pub fn rx_dscr_vb_err_st(&self) -> RX_DSCR_VB_ERR_ST_R

Bit 3 - The error is that VB in 2DDMA receive descriptor plus the offset of Y coordinate in 2DDMA transmit descriptor is larger than VA in 2DDMA transmit descriptor

pub fn ydst_len_too_samll_err_st(&self) -> YDST_LEN_TOO_SAMLL_ERR_ST_R

Bit 4 - The error is that the scaled image width is 0. For example. when source width is 14. scaled value is 1/16. and no rotate operation. then scaled width would be 0 as the result would be floored.

pub fn ydst_len_too_large_err_st(&self) -> YDST_LEN_TOO_LARGE_ERR_ST_R

Bit 5 - The error is that the scaled width is larger than (2^13 - 1).

pub fn x_rx_scal_equal_0_err_st(&self) -> X_RX_SCAL_EQUAL_0_ERR_ST_R

Bit 6 - The error is that the scaled image height is 0.

pub fn rx_dscr_hb_err_st(&self) -> RX_DSCR_HB_ERR_ST_R

Bit 7 - The error is that the HB in 2DDMA transmit descriptor plus the offset of X coordinate in 2DDMA transmit descriptor is larger than HA in 2DDMA transmit descriptor.

pub fn xdst_len_too_samll_err_st(&self) -> XDST_LEN_TOO_SAMLL_ERR_ST_R

Bit 8 - The error is that the scaled image height is 0. For example. when source height is 14. scaled value is 1/16. and no rotate operation. then scaled height would be 0 as the result would be floored.

pub fn xdst_len_too_large_err_st(&self) -> XDST_LEN_TOO_LARGE_ERR_ST_R

Bit 9 - The error is that the scaled image height is larger than (2^13 - 1).

pub fn x_yuv420_rx_scale_err_st(&self) -> X_YUV420_RX_SCALE_ERR_ST_R

Bit 10 - The error is that the ha/hb/x param in dma2d descriptor is an odd num when enable yuv420 rx

pub fn y_yuv420_rx_scale_err_st(&self) -> Y_YUV420_RX_SCALE_ERR_ST_R

Bit 11 - The error is that the va/vb/y param in dma2d descriptor is an odd num when enable yuv420 rx

pub fn x_yuv420_tx_scale_err_st(&self) -> X_YUV420_TX_SCALE_ERR_ST_R

Bit 12 - The error is that the ha/hb/x param in dma2d descriptor is an odd num when enable yuv420 tx

pub fn y_yuv420_tx_scale_err_st(&self) -> Y_YUV420_TX_SCALE_ERR_ST_R

Bit 13 - The error is that the va/vb/y param in dma2d descriptor is an odd num when enable yuv420 tx