Module esp32p4::ppa::sr_scal_rotate

Scaling and rotating coefficient register

Structs

SR_SCAL_ROTATE_SPEC

Scaling and rotating coefficient register

Type Aliases

R

Register SR_SCAL_ROTATE reader

SCAL_ROTATE_RST_R

Field SCAL_ROTATE_RST reader - Write 1 then write 0 to this bit to reset scaling and rotating engine.

SCAL_ROTATE_RST_W

Field SCAL_ROTATE_RST writer - Write 1 then write 0 to this bit to reset scaling and rotating engine.

SCAL_ROTATE_START_W

Field SCAL_ROTATE_START writer - Write 1 to enable scaling and rotating engine after parameter is configured.

SR_MIRROR_X_R

Field SR_MIRROR_X reader - Image mirror in X direction. 0: disable, 1: enable

SR_MIRROR_X_W

Field SR_MIRROR_X writer - Image mirror in X direction. 0: disable, 1: enable

SR_MIRROR_Y_R

Field SR_MIRROR_Y reader - Image mirror in Y direction. 0: disable, 1: enable

SR_MIRROR_Y_W

Field SR_MIRROR_Y writer - Image mirror in Y direction. 0: disable, 1: enable

SR_ROTATE_ANGLE_R

Field SR_ROTATE_ANGLE reader - The rotate angle. 0: 0 degree. 1: 90 degree. 2: 180 degree. 3: 270 degree.

SR_ROTATE_ANGLE_W

Field SR_ROTATE_ANGLE writer - The rotate angle. 0: 0 degree. 1: 90 degree. 2: 180 degree. 3: 270 degree.

SR_SCAL_X_FRAG_R

Field SR_SCAL_X_FRAG reader - The fragment part of scaling coefficient in X direction.

SR_SCAL_X_FRAG_W

Field SR_SCAL_X_FRAG writer - The fragment part of scaling coefficient in X direction.

SR_SCAL_X_INT_R

Field SR_SCAL_X_INT reader - The integrated part of scaling coefficient in X direction.

SR_SCAL_X_INT_W

Field SR_SCAL_X_INT writer - The integrated part of scaling coefficient in X direction.

SR_SCAL_Y_FRAG_R

Field SR_SCAL_Y_FRAG reader - The fragment part of scaling coefficient in Y direction.

SR_SCAL_Y_FRAG_W

Field SR_SCAL_Y_FRAG writer - The fragment part of scaling coefficient in Y direction.

SR_SCAL_Y_INT_R

Field SR_SCAL_Y_INT reader - The integrated part of scaling coefficient in Y direction.

SR_SCAL_Y_INT_W

Field SR_SCAL_Y_INT writer - The integrated part of scaling coefficient in Y direction.

W

Register SR_SCAL_ROTATE writer