Type Alias esp32p4::ppa::sr_scal_rotate::R

pub type R = R<SR_SCAL_ROTATE_SPEC>;

Register SR_SCAL_ROTATE reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn sr_scal_x_int(&self) -> SR_SCAL_X_INT_R

Bits 0:7 - The integrated part of scaling coefficient in X direction.

pub fn sr_scal_x_frag(&self) -> SR_SCAL_X_FRAG_R

Bits 8:11 - The fragment part of scaling coefficient in X direction.

pub fn sr_scal_y_int(&self) -> SR_SCAL_Y_INT_R

Bits 12:19 - The integrated part of scaling coefficient in Y direction.

pub fn sr_scal_y_frag(&self) -> SR_SCAL_Y_FRAG_R

Bits 20:23 - The fragment part of scaling coefficient in Y direction.

pub fn sr_rotate_angle(&self) -> SR_ROTATE_ANGLE_R

Bits 24:25 - The rotate angle. 0: 0 degree. 1: 90 degree. 2: 180 degree. 3: 270 degree.

pub fn scal_rotate_rst(&self) -> SCAL_ROTATE_RST_R

Bit 26 - Write 1 then write 0 to this bit to reset scaling and rotating engine.

pub fn sr_mirror_x(&self) -> SR_MIRROR_X_R

Bit 28 - Image mirror in X direction. 0: disable, 1: enable

pub fn sr_mirror_y(&self) -> SR_MIRROR_Y_R

Bit 29 - Image mirror in Y direction. 0: disable, 1: enable