Type Alias esp32p4::rmt::int_raw::W

pub type W = W<INT_RAW_SPEC>;

Register INT_RAW writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn ch0_tx_end_int_raw(&mut self) -> CH0_TX_END_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 0 - The interrupt raw bit for CHANNEL0. Triggered when transmission done.

pub fn ch1_tx_end_int_raw(&mut self) -> CH1_TX_END_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 1 - The interrupt raw bit for CHANNEL1. Triggered when transmission done.

pub fn ch2_tx_end_int_raw(&mut self) -> CH2_TX_END_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 2 - The interrupt raw bit for CHANNEL2. Triggered when transmission done.

pub fn ch3_tx_end_int_raw(&mut self) -> CH3_TX_END_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 3 - The interrupt raw bit for CHANNEL3. Triggered when transmission done.

pub fn tx_ch0_err_int_raw(&mut self) -> TX_CH0_ERR_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 4 - The interrupt raw bit for CHANNEL0. Triggered when error occurs.

pub fn tx_ch1_err_int_raw(&mut self) -> TX_CH1_ERR_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 5 - The interrupt raw bit for CHANNEL1. Triggered when error occurs.

pub fn tx_ch2_err_int_raw(&mut self) -> TX_CH2_ERR_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 6 - The interrupt raw bit for CHANNEL2. Triggered when error occurs.

pub fn tx_ch3_err_int_raw(&mut self) -> TX_CH3_ERR_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 7 - The interrupt raw bit for CHANNEL3. Triggered when error occurs.

pub fn ch0_tx_thr_event_int_raw( &mut self ) -> CH0_TX_THR_EVENT_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 8 - The interrupt raw bit for CHANNEL0. Triggered when transmitter sent more data than configured value.

pub fn ch1_tx_thr_event_int_raw( &mut self ) -> CH1_TX_THR_EVENT_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 9 - The interrupt raw bit for CHANNEL1. Triggered when transmitter sent more data than configured value.

pub fn ch2_tx_thr_event_int_raw( &mut self ) -> CH2_TX_THR_EVENT_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 10 - The interrupt raw bit for CHANNEL2. Triggered when transmitter sent more data than configured value.

pub fn ch3_tx_thr_event_int_raw( &mut self ) -> CH3_TX_THR_EVENT_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 11 - The interrupt raw bit for CHANNEL3. Triggered when transmitter sent more data than configured value.

pub fn ch0_tx_loop_int_raw(&mut self) -> CH0_TX_LOOP_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 12 - The interrupt raw bit for CHANNEL0. Triggered when the loop count reaches the configured threshold value.

pub fn ch1_tx_loop_int_raw(&mut self) -> CH1_TX_LOOP_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 13 - The interrupt raw bit for CHANNEL1. Triggered when the loop count reaches the configured threshold value.

pub fn ch2_tx_loop_int_raw(&mut self) -> CH2_TX_LOOP_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 14 - The interrupt raw bit for CHANNEL2. Triggered when the loop count reaches the configured threshold value.

pub fn ch3_tx_loop_int_raw(&mut self) -> CH3_TX_LOOP_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 15 - The interrupt raw bit for CHANNEL3. Triggered when the loop count reaches the configured threshold value.

pub fn ch4_rx_end_int_raw(&mut self) -> CH4_RX_END_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 16 - The interrupt raw bit for CHANNEL4. Triggered when reception done.

pub fn ch5_rx_end_int_raw(&mut self) -> CH5_RX_END_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 17 - The interrupt raw bit for CHANNEL5. Triggered when reception done.

pub fn ch6_rx_end_int_raw(&mut self) -> CH6_RX_END_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 18 - The interrupt raw bit for CHANNEL6. Triggered when reception done.

pub fn ch7_rx_end_int_raw(&mut self) -> CH7_RX_END_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 19 - The interrupt raw bit for CHANNEL7. Triggered when reception done.

pub fn rx_ch4_err_int_raw(&mut self) -> RX_CH4_ERR_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 20 - The interrupt raw bit for CHANNEL4. Triggered when error occurs.

pub fn rx_ch5_err_int_raw(&mut self) -> RX_CH5_ERR_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 21 - The interrupt raw bit for CHANNEL5. Triggered when error occurs.

pub fn rx_ch6_err_int_raw(&mut self) -> RX_CH6_ERR_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 22 - The interrupt raw bit for CHANNEL6. Triggered when error occurs.

pub fn rx_ch7_err_int_raw(&mut self) -> RX_CH7_ERR_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 23 - The interrupt raw bit for CHANNEL7. Triggered when error occurs.

pub fn ch4_rx_thr_event_int_raw( &mut self ) -> CH4_RX_THR_EVENT_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 24 - The interrupt raw bit for CHANNEL4. Triggered when receiver receive more data than configured value.

pub fn ch5_rx_thr_event_int_raw( &mut self ) -> CH5_RX_THR_EVENT_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 25 - The interrupt raw bit for CHANNEL5. Triggered when receiver receive more data than configured value.

pub fn ch6_rx_thr_event_int_raw( &mut self ) -> CH6_RX_THR_EVENT_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 26 - The interrupt raw bit for CHANNEL6. Triggered when receiver receive more data than configured value.

pub fn ch7_rx_thr_event_int_raw( &mut self ) -> CH7_RX_THR_EVENT_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 27 - The interrupt raw bit for CHANNEL7. Triggered when receiver receive more data than configured value.

pub fn tx_ch3_dma_access_fail_int_raw( &mut self ) -> TX_CH3_DMA_ACCESS_FAIL_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 28 - The interrupt raw bit for CHANNEL3. Triggered when dma accessing CHANNEL3 fails.

pub fn rx_ch7_dma_access_fail_int_raw( &mut self ) -> RX_CH7_DMA_ACCESS_FAIL_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 29 - The interrupt raw bit for CHANNEL7. Triggered when dma accessing CHANNEL7 fails.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual