Type Alias esp32p4::rmt::tx_chconf0::W

pub type W = W<TX_CHCONF0_SPEC>;

Register TX_CH%sCONF0 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn tx_start_ch0(&mut self) -> TX_START_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 0 - Set this bit to start sending data on CHANNEL%s.

pub fn mem_rd_rst_ch0(&mut self) -> MEM_RD_RST_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 1 - Set this bit to reset read ram address for CHANNEL%s by accessing transmitter.

pub fn apb_mem_rst_ch0(&mut self) -> APB_MEM_RST_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 2 - Set this bit to reset W/R ram address for CHANNEL%s by accessing apb fifo.

pub fn tx_conti_mode_ch0(&mut self) -> TX_CONTI_MODE_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 3 - Set this bit to restart transmission from the first data to the last data in CHANNEL%s.

pub fn mem_tx_wrap_en_ch0( &mut self ) -> MEM_TX_WRAP_EN_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 4 - This is the channel %s enable bit for wraparound mode: it will resume sending at the start when the data to be sent is more than its memory size.

pub fn idle_out_lv_ch0(&mut self) -> IDLE_OUT_LV_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 5 - This bit configures the level of output signal in CHANNEL%s when the latter is in IDLE state.

pub fn idle_out_en_ch0(&mut self) -> IDLE_OUT_EN_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 6 - This is the output enable-control bit for CHANNEL%s in IDLE state.

pub fn tx_stop_ch0(&mut self) -> TX_STOP_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 7 - Set this bit to stop the transmitter of CHANNEL%s sending data out.

pub fn div_cnt_ch0(&mut self) -> DIV_CNT_CH0_W<'_, TX_CHCONF0_SPEC>

Bits 8:15 - This register is used to configure the divider for clock of CHANNEL%s.

pub fn mem_size_ch0(&mut self) -> MEM_SIZE_CH0_W<'_, TX_CHCONF0_SPEC>

Bits 16:19 - This register is used to configure the maximum size of memory allocated to CHANNEL%s.

pub fn carrier_eff_en_ch0( &mut self ) -> CARRIER_EFF_EN_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 20 - 1: Add carrier modulation on the output signal only at the send data state for CHANNEL%s. 0: Add carrier modulation on the output signal at all state for CHANNEL%s. Only valid when RMT_CARRIER_EN_CH%s is 1.

pub fn carrier_en_ch0(&mut self) -> CARRIER_EN_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 21 - This is the carrier modulation enable-control bit for CHANNEL%s. 1: Add carrier modulation in the output signal. 0: No carrier modulation in sig_out.

pub fn carrier_out_lv_ch0( &mut self ) -> CARRIER_OUT_LV_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 22 - This bit is used to configure the position of carrier wave for CHANNEL%s.1’h0: add carrier wave on low level.1’h1: add carrier wave on high level.

pub fn afifo_rst_ch0(&mut self) -> AFIFO_RST_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 23 - Reserved

pub fn conf_update_ch0(&mut self) -> CONF_UPDATE_CH0_W<'_, TX_CHCONF0_SPEC>

Bit 24 - synchronization bit for CHANNEL%s

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual