Module esp32p4::sdhost::clk_edge_sel

SDIO control register.

Structs

CLK_EDGE_SEL_SPEC

SDIO control register.

Type Aliases

CCLKIN_EDGE_DRV_SEL_R

Field CCLKIN_EDGE_DRV_SEL reader - It’s used to select the clock phase of the output signal from phase 0, phase 90, phase 180, phase 270.

CCLKIN_EDGE_DRV_SEL_W

Field CCLKIN_EDGE_DRV_SEL writer - It’s used to select the clock phase of the output signal from phase 0, phase 90, phase 180, phase 270.

CCLKIN_EDGE_SAM_SEL_R

Field CCLKIN_EDGE_SAM_SEL reader - It’s used to select the clock phase of the input signal from phase 0, phase 90, phase 180, phase 270.

CCLKIN_EDGE_SAM_SEL_W

Field CCLKIN_EDGE_SAM_SEL writer - It’s used to select the clock phase of the input signal from phase 0, phase 90, phase 180, phase 270.

CCLKIN_EDGE_SLF_SEL_R

Field CCLKIN_EDGE_SLF_SEL reader - It’s used to select the clock phase of the internal signal from phase 0, phase 90, phase 180, phase 270.

CCLKIN_EDGE_SLF_SEL_W

Field CCLKIN_EDGE_SLF_SEL writer - It’s used to select the clock phase of the internal signal from phase 0, phase 90, phase 180, phase 270.

CCLK_EN_R

Field CCLK_EN reader - Sdio clock enable.

CCLK_EN_W

Field CCLK_EN writer - Sdio clock enable.

CCLLKIN_EDGE_H_R

Field CCLLKIN_EDGE_H reader - The high level of the divider clock. The value should be smaller than CCLKIN_EDGE_L.

CCLLKIN_EDGE_H_W

Field CCLLKIN_EDGE_H writer - The high level of the divider clock. The value should be smaller than CCLKIN_EDGE_L.

CCLLKIN_EDGE_L_R

Field CCLLKIN_EDGE_L reader - The low level of the divider clock. The value should be larger than CCLKIN_EDGE_H.

CCLLKIN_EDGE_L_W

Field CCLLKIN_EDGE_L writer - The low level of the divider clock. The value should be larger than CCLKIN_EDGE_H.

CCLLKIN_EDGE_N_R

Field CCLLKIN_EDGE_N reader - The clock division of cclk_in.

CCLLKIN_EDGE_N_W

Field CCLLKIN_EDGE_N writer - The clock division of cclk_in.

ESDIO_MODE_R

Field ESDIO_MODE reader - Enable esdio mode.

ESDIO_MODE_W

Field ESDIO_MODE writer - Enable esdio mode.

ESD_MODE_R

Field ESD_MODE reader - Enable esd mode.

ESD_MODE_W

Field ESD_MODE writer - Enable esd mode.

R

Register CLK_EDGE_SEL reader

ULTRA_HIGH_SPEED_MODE_R

Field ULTRA_HIGH_SPEED_MODE reader - Enable ultra high speed mode, use dll to generate clk.

ULTRA_HIGH_SPEED_MODE_W

Field ULTRA_HIGH_SPEED_MODE writer - Enable ultra high speed mode, use dll to generate clk.

W

Register CLK_EDGE_SEL writer