Type Alias esp32p4::sdhost::clk_edge_sel::W

pub type W = W<CLK_EDGE_SEL_SPEC>;

Register CLK_EDGE_SEL writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn cclkin_edge_drv_sel( &mut self ) -> CCLKIN_EDGE_DRV_SEL_W<'_, CLK_EDGE_SEL_SPEC>

Bits 0:2 - It’s used to select the clock phase of the output signal from phase 0, phase 90, phase 180, phase 270.

pub fn cclkin_edge_sam_sel( &mut self ) -> CCLKIN_EDGE_SAM_SEL_W<'_, CLK_EDGE_SEL_SPEC>

Bits 3:5 - It’s used to select the clock phase of the input signal from phase 0, phase 90, phase 180, phase 270.

pub fn cclkin_edge_slf_sel( &mut self ) -> CCLKIN_EDGE_SLF_SEL_W<'_, CLK_EDGE_SEL_SPEC>

Bits 6:8 - It’s used to select the clock phase of the internal signal from phase 0, phase 90, phase 180, phase 270.

pub fn ccllkin_edge_h(&mut self) -> CCLLKIN_EDGE_H_W<'_, CLK_EDGE_SEL_SPEC>

Bits 9:12 - The high level of the divider clock. The value should be smaller than CCLKIN_EDGE_L.

pub fn ccllkin_edge_l(&mut self) -> CCLLKIN_EDGE_L_W<'_, CLK_EDGE_SEL_SPEC>

Bits 13:16 - The low level of the divider clock. The value should be larger than CCLKIN_EDGE_H.

pub fn ccllkin_edge_n(&mut self) -> CCLLKIN_EDGE_N_W<'_, CLK_EDGE_SEL_SPEC>

Bits 17:20 - The clock division of cclk_in.

pub fn esdio_mode(&mut self) -> ESDIO_MODE_W<'_, CLK_EDGE_SEL_SPEC>

Bit 21 - Enable esdio mode.

pub fn esd_mode(&mut self) -> ESD_MODE_W<'_, CLK_EDGE_SEL_SPEC>

Bit 22 - Enable esd mode.

pub fn cclk_en(&mut self) -> CCLK_EN_W<'_, CLK_EDGE_SEL_SPEC>

Bit 23 - Sdio clock enable.

pub fn ultra_high_speed_mode( &mut self ) -> ULTRA_HIGH_SPEED_MODE_W<'_, CLK_EDGE_SEL_SPEC>

Bit 24 - Enable ultra high speed mode, use dll to generate clk.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual