Module esp32p4::spi0

SPI (Serial Peripheral Interface) Controller 0

Modules

spi_fmem_pms_addr

SPI1 flash PMS section %s start address register

spi_fmem_pms_attr

MSPI flash PMS section %s attribute register

spi_fmem_pms_size

SPI1 flash PMS section %s start address register

spi_mem_axi_err_addr

SPI0 AXI request error address.

spi_mem_axi_err_resp_en

SPI0 AXI error response enable register

spi_mem_cache_fctrl

SPI0 bit mode control register.

spi_mem_cache_sctrl

SPI0 external RAM control register

spi_mem_clock

SPI clock division control register.

spi_mem_clock_gate

SPI0 clock gate register

spi_mem_cmd

SPI0 FSM status register

spi_mem_ctrl

SPI0 control register.

spi_mem_ctrl1

SPI0 control1 register.

spi_mem_ctrl2

SPI0 control2 register.

spi_mem_date

SPI0 version control register

spi_mem_ddr

SPI0 flash DDR mode control register

spi_mem_din_mode

MSPI flash input timing delay mode control register

spi_mem_din_num

MSPI flash input timing delay number control register

spi_mem_dout_mode

MSPI flash output timing adjustment control register

spi_mem_dpa_ctrl

SPI memory cryption DPA register

spi_mem_ecc_ctrl

MSPI ECC control register

spi_mem_ecc_err_addr

MSPI ECC error address register

spi_mem_fsm

SPI0 FSM status register

spi_mem_int_clr

SPI0 interrupt clear register

spi_mem_int_ena

SPI0 interrupt enable register

spi_mem_int_raw

SPI0 interrupt raw register

spi_mem_int_st

SPI0 interrupt status register

spi_mem_misc

SPI0 misc register

spi_mem_mmu_item_content

MSPI-MMU item content register

spi_mem_mmu_item_index

MSPI-MMU item index register

spi_mem_mmu_power_ctrl

MSPI MMU power control register

spi_mem_pms_reject

SPI1 access reject register

spi_mem_rd_status

SPI0 read control register.

spi_mem_registerrnd_eco_high

MSPI ECO high register

spi_mem_registerrnd_eco_low

MSPI ECO low register

spi_mem_sram_clk

SPI0 external RAM clock control register

spi_mem_sram_cmd

SPI0 external RAM mode control register

spi_mem_sram_drd_cmd

SPI0 external RAM DDR read command control register

spi_mem_sram_dwr_cmd

SPI0 external RAM DDR write command control register

spi_mem_timing_cali

SPI0 flash timing calibration register

spi_mem_user

SPI0 user register.

spi_mem_user1

SPI0 user1 register.

spi_mem_user2

SPI0 user2 register.

spi_mem_xts_date

Manual Encryption version register

spi_mem_xts_destination

Manual Encryption destination register

spi_mem_xts_destroy

Manual Encryption physical address register

spi_mem_xts_linesize

Manual Encryption Line-Size register

spi_mem_xts_physical_address

Manual Encryption physical address register

spi_mem_xts_plain_base

The base address of the memory that stores plaintext in Manual Encryption

spi_mem_xts_release

Manual Encryption physical address register

spi_mem_xts_state

Manual Encryption physical address register

spi_mem_xts_trigger

Manual Encryption physical address register

spi_smem_ac

MSPI external RAM ECC and SPI CS timing control register

spi_smem_axi_addr_ctrl

SPI0 AXI address control register

spi_smem_ddr

SPI0 external RAM DDR mode control register

spi_smem_din_hex_mode

MSPI 16x external RAM input timing delay mode control register

spi_smem_din_hex_num

MSPI 16x external RAM input timing delay number control register

spi_smem_din_mode

MSPI external RAM input timing delay mode control register

spi_smem_din_num

MSPI external RAM input timing delay number control register

spi_smem_dout_hex_mode

MSPI 16x external RAM output timing adjustment control register

spi_smem_dout_mode

MSPI external RAM output timing adjustment control register

spi_smem_ecc_ctrl

MSPI ECC control register

spi_smem_pms_addr

SPI1 external RAM PMS section %s start address register

spi_smem_pms_attr

SPI1 flash PMS section %s start address register

spi_smem_pms_size

SPI1 external RAM PMS section %s start address register

spi_smem_timing_cali

MSPI external RAM timing calibration register

Structs

RegisterBlock

Register block

Type Aliases

SPI_FMEM_PMS_ADDR

SPI_FMEM_PMS_ADDR (rw) register accessor: SPI1 flash PMS section %s start address register

SPI_FMEM_PMS_ATTR

SPI_FMEM_PMS_ATTR (rw) register accessor: MSPI flash PMS section %s attribute register

SPI_FMEM_PMS_SIZE

SPI_FMEM_PMS_SIZE (rw) register accessor: SPI1 flash PMS section %s start address register

SPI_MEM_AXI_ERR_ADDR

SPI_MEM_AXI_ERR_ADDR (r) register accessor: SPI0 AXI request error address.

SPI_MEM_AXI_ERR_RESP_EN

SPI_MEM_AXI_ERR_RESP_EN (rw) register accessor: SPI0 AXI error response enable register

SPI_MEM_CACHE_FCTRL

SPI_MEM_CACHE_FCTRL (rw) register accessor: SPI0 bit mode control register.

SPI_MEM_CACHE_SCTRL

SPI_MEM_CACHE_SCTRL (rw) register accessor: SPI0 external RAM control register

SPI_MEM_CLOCK

SPI_MEM_CLOCK (rw) register accessor: SPI clock division control register.

SPI_MEM_CLOCK_GATE

SPI_MEM_CLOCK_GATE (rw) register accessor: SPI0 clock gate register

SPI_MEM_CMD

SPI_MEM_CMD (r) register accessor: SPI0 FSM status register

SPI_MEM_CTRL

SPI_MEM_CTRL (rw) register accessor: SPI0 control register.

SPI_MEM_CTRL1

SPI_MEM_CTRL1 (rw) register accessor: SPI0 control1 register.

SPI_MEM_CTRL2

SPI_MEM_CTRL2 (rw) register accessor: SPI0 control2 register.

SPI_MEM_DATE

SPI_MEM_DATE (rw) register accessor: SPI0 version control register

SPI_MEM_DDR

SPI_MEM_DDR (rw) register accessor: SPI0 flash DDR mode control register

SPI_MEM_DIN_MODE

SPI_MEM_DIN_MODE (rw) register accessor: MSPI flash input timing delay mode control register

SPI_MEM_DIN_NUM

SPI_MEM_DIN_NUM (rw) register accessor: MSPI flash input timing delay number control register

SPI_MEM_DOUT_MODE

SPI_MEM_DOUT_MODE (rw) register accessor: MSPI flash output timing adjustment control register

SPI_MEM_DPA_CTRL

SPI_MEM_DPA_CTRL (rw) register accessor: SPI memory cryption DPA register

SPI_MEM_ECC_CTRL

SPI_MEM_ECC_CTRL (rw) register accessor: MSPI ECC control register

SPI_MEM_ECC_ERR_ADDR

SPI_MEM_ECC_ERR_ADDR (r) register accessor: MSPI ECC error address register

SPI_MEM_FSM

SPI_MEM_FSM (rw) register accessor: SPI0 FSM status register

SPI_MEM_INT_CLR

SPI_MEM_INT_CLR (w) register accessor: SPI0 interrupt clear register

SPI_MEM_INT_ENA

SPI_MEM_INT_ENA (rw) register accessor: SPI0 interrupt enable register

SPI_MEM_INT_RAW

SPI_MEM_INT_RAW (rw) register accessor: SPI0 interrupt raw register

SPI_MEM_INT_ST

SPI_MEM_INT_ST (r) register accessor: SPI0 interrupt status register

SPI_MEM_MISC

SPI_MEM_MISC (rw) register accessor: SPI0 misc register

SPI_MEM_MMU_ITEM_CONTENT

SPI_MEM_MMU_ITEM_CONTENT (rw) register accessor: MSPI-MMU item content register

SPI_MEM_MMU_ITEM_INDEX

SPI_MEM_MMU_ITEM_INDEX (rw) register accessor: MSPI-MMU item index register

SPI_MEM_MMU_POWER_CTRL

SPI_MEM_MMU_POWER_CTRL (rw) register accessor: MSPI MMU power control register

SPI_MEM_PMS_REJECT

SPI_MEM_PMS_REJECT (rw) register accessor: SPI1 access reject register

SPI_MEM_RD_STATUS

SPI_MEM_RD_STATUS (rw) register accessor: SPI0 read control register.

SPI_MEM_REGISTERRND_ECO_HIGH

SPI_MEM_REGISTERRND_ECO_HIGH (rw) register accessor: MSPI ECO high register

SPI_MEM_REGISTERRND_ECO_LOW

SPI_MEM_REGISTERRND_ECO_LOW (rw) register accessor: MSPI ECO low register

SPI_MEM_SRAM_CLK

SPI_MEM_SRAM_CLK (rw) register accessor: SPI0 external RAM clock control register

SPI_MEM_SRAM_CMD

SPI_MEM_SRAM_CMD (rw) register accessor: SPI0 external RAM mode control register

SPI_MEM_SRAM_DRD_CMD

SPI_MEM_SRAM_DRD_CMD (rw) register accessor: SPI0 external RAM DDR read command control register

SPI_MEM_SRAM_DWR_CMD

SPI_MEM_SRAM_DWR_CMD (rw) register accessor: SPI0 external RAM DDR write command control register

SPI_MEM_TIMING_CALI

SPI_MEM_TIMING_CALI (rw) register accessor: SPI0 flash timing calibration register

SPI_MEM_USER

SPI_MEM_USER (rw) register accessor: SPI0 user register.

SPI_MEM_USER1

SPI_MEM_USER1 (rw) register accessor: SPI0 user1 register.

SPI_MEM_USER2

SPI_MEM_USER2 (rw) register accessor: SPI0 user2 register.

SPI_MEM_XTS_DATE

SPI_MEM_XTS_DATE (rw) register accessor: Manual Encryption version register

SPI_MEM_XTS_DESTINATION

SPI_MEM_XTS_DESTINATION (rw) register accessor: Manual Encryption destination register

SPI_MEM_XTS_DESTROY

SPI_MEM_XTS_DESTROY (w) register accessor: Manual Encryption physical address register

SPI_MEM_XTS_LINESIZE

SPI_MEM_XTS_LINESIZE (rw) register accessor: Manual Encryption Line-Size register

SPI_MEM_XTS_PHYSICAL_ADDRESS

SPI_MEM_XTS_PHYSICAL_ADDRESS (rw) register accessor: Manual Encryption physical address register

SPI_MEM_XTS_PLAIN_BASE

SPI_MEM_XTS_PLAIN_BASE (rw) register accessor: The base address of the memory that stores plaintext in Manual Encryption

SPI_MEM_XTS_RELEASE

SPI_MEM_XTS_RELEASE (w) register accessor: Manual Encryption physical address register

SPI_MEM_XTS_STATE

SPI_MEM_XTS_STATE (r) register accessor: Manual Encryption physical address register

SPI_MEM_XTS_TRIGGER

SPI_MEM_XTS_TRIGGER (w) register accessor: Manual Encryption physical address register

SPI_SMEM_AC

SPI_SMEM_AC (rw) register accessor: MSPI external RAM ECC and SPI CS timing control register

SPI_SMEM_AXI_ADDR_CTRL

SPI_SMEM_AXI_ADDR_CTRL (r) register accessor: SPI0 AXI address control register

SPI_SMEM_DDR

SPI_SMEM_DDR (rw) register accessor: SPI0 external RAM DDR mode control register

SPI_SMEM_DIN_HEX_MODE

SPI_SMEM_DIN_HEX_MODE (rw) register accessor: MSPI 16x external RAM input timing delay mode control register

SPI_SMEM_DIN_HEX_NUM

SPI_SMEM_DIN_HEX_NUM (rw) register accessor: MSPI 16x external RAM input timing delay number control register

SPI_SMEM_DIN_MODE

SPI_SMEM_DIN_MODE (rw) register accessor: MSPI external RAM input timing delay mode control register

SPI_SMEM_DIN_NUM

SPI_SMEM_DIN_NUM (rw) register accessor: MSPI external RAM input timing delay number control register

SPI_SMEM_DOUT_HEX_MODE

SPI_SMEM_DOUT_HEX_MODE (rw) register accessor: MSPI 16x external RAM output timing adjustment control register

SPI_SMEM_DOUT_MODE

SPI_SMEM_DOUT_MODE (rw) register accessor: MSPI external RAM output timing adjustment control register

SPI_SMEM_ECC_CTRL

SPI_SMEM_ECC_CTRL (rw) register accessor: MSPI ECC control register

SPI_SMEM_PMS_ADDR

SPI_SMEM_PMS_ADDR (rw) register accessor: SPI1 external RAM PMS section %s start address register

SPI_SMEM_PMS_ATTR

SPI_SMEM_PMS_ATTR (rw) register accessor: SPI1 flash PMS section %s start address register

SPI_SMEM_PMS_SIZE

SPI_SMEM_PMS_SIZE (rw) register accessor: SPI1 external RAM PMS section %s start address register

SPI_SMEM_TIMING_CALI

SPI_SMEM_TIMING_CALI (rw) register accessor: MSPI external RAM timing calibration register