Type Alias esp32p4::spi0::spi_mem_axi_err_resp_en::W

pub type W = W<SPI_MEM_AXI_ERR_RESP_EN_SPEC>;

Register SPI_MEM_AXI_ERR_RESP_EN writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn spi_mem_aw_resp_en_mmu_vld( &mut self ) -> SPI_MEM_AW_RESP_EN_MMU_VLD_W<'_, SPI_MEM_AXI_ERR_RESP_EN_SPEC>

Bit 0 - Set this bit to enable AXI response function for mmu valid err in axi write trans.

pub fn spi_mem_aw_resp_en_mmu_gid( &mut self ) -> SPI_MEM_AW_RESP_EN_MMU_GID_W<'_, SPI_MEM_AXI_ERR_RESP_EN_SPEC>

Bit 1 - Set this bit to enable AXI response function for mmu gid err in axi write trans.

pub fn spi_mem_aw_resp_en_axi_size( &mut self ) -> SPI_MEM_AW_RESP_EN_AXI_SIZE_W<'_, SPI_MEM_AXI_ERR_RESP_EN_SPEC>

Bit 2 - Set this bit to enable AXI response function for axi size err in axi write trans.

pub fn spi_mem_aw_resp_en_axi_flash( &mut self ) -> SPI_MEM_AW_RESP_EN_AXI_FLASH_W<'_, SPI_MEM_AXI_ERR_RESP_EN_SPEC>

Bit 3 - Set this bit to enable AXI response function for axi flash err in axi write trans.

pub fn spi_mem_aw_resp_en_mmu_ecc( &mut self ) -> SPI_MEM_AW_RESP_EN_MMU_ECC_W<'_, SPI_MEM_AXI_ERR_RESP_EN_SPEC>

Bit 4 - Set this bit to enable AXI response function for mmu ecc err in axi write trans.

pub fn spi_mem_aw_resp_en_mmu_sens( &mut self ) -> SPI_MEM_AW_RESP_EN_MMU_SENS_W<'_, SPI_MEM_AXI_ERR_RESP_EN_SPEC>

Bit 5 - Set this bit to enable AXI response function for mmu sens in err axi write trans.

pub fn spi_mem_aw_resp_en_axi_wstrb( &mut self ) -> SPI_MEM_AW_RESP_EN_AXI_WSTRB_W<'_, SPI_MEM_AXI_ERR_RESP_EN_SPEC>

Bit 6 - Set this bit to enable AXI response function for axi wstrb err in axi write trans.

pub fn spi_mem_ar_resp_en_mmu_vld( &mut self ) -> SPI_MEM_AR_RESP_EN_MMU_VLD_W<'_, SPI_MEM_AXI_ERR_RESP_EN_SPEC>

Bit 7 - Set this bit to enable AXI response function for mmu valid err in axi read trans.

pub fn spi_mem_ar_resp_en_mmu_gid( &mut self ) -> SPI_MEM_AR_RESP_EN_MMU_GID_W<'_, SPI_MEM_AXI_ERR_RESP_EN_SPEC>

Bit 8 - Set this bit to enable AXI response function for mmu gid err in axi read trans.

pub fn spi_mem_ar_resp_en_mmu_ecc( &mut self ) -> SPI_MEM_AR_RESP_EN_MMU_ECC_W<'_, SPI_MEM_AXI_ERR_RESP_EN_SPEC>

Bit 9 - Set this bit to enable AXI response function for mmu ecc err in axi read trans.

pub fn spi_mem_ar_resp_en_mmu_sens( &mut self ) -> SPI_MEM_AR_RESP_EN_MMU_SENS_W<'_, SPI_MEM_AXI_ERR_RESP_EN_SPEC>

Bit 10 - Set this bit to enable AXI response function for mmu sensitive err in axi read trans.

pub fn spi_mem_ar_resp_en_axi_size( &mut self ) -> SPI_MEM_AR_RESP_EN_AXI_SIZE_W<'_, SPI_MEM_AXI_ERR_RESP_EN_SPEC>

Bit 11 - Set this bit to enable AXI response function for axi size err in axi read trans.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual