Type Alias esp32p4::spi0::spi_mem_cache_sctrl::W

pub type W = W<SPI_MEM_CACHE_SCTRL_SPEC>;

Register SPI_MEM_CACHE_SCTRL writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn spi_mem_cache_usr_saddr_4byte( &mut self ) -> SPI_MEM_CACHE_USR_SADDR_4BYTE_W<'_, SPI_MEM_CACHE_SCTRL_SPEC>

Bit 0 - For SPI0, In the external RAM mode, cache read flash with 4 bytes command, 1: enable, 0:disable.

pub fn spi_mem_usr_sram_dio( &mut self ) -> SPI_MEM_USR_SRAM_DIO_W<'_, SPI_MEM_CACHE_SCTRL_SPEC>

Bit 1 - For SPI0, In the external RAM mode, spi dual I/O mode enable, 1: enable, 0:disable

pub fn spi_mem_usr_sram_qio( &mut self ) -> SPI_MEM_USR_SRAM_QIO_W<'_, SPI_MEM_CACHE_SCTRL_SPEC>

Bit 2 - For SPI0, In the external RAM mode, spi quad I/O mode enable, 1: enable, 0:disable

pub fn spi_mem_usr_wr_sram_dummy( &mut self ) -> SPI_MEM_USR_WR_SRAM_DUMMY_W<'_, SPI_MEM_CACHE_SCTRL_SPEC>

Bit 3 - For SPI0, In the external RAM mode, it is the enable bit of dummy phase for write operations.

pub fn spi_mem_usr_rd_sram_dummy( &mut self ) -> SPI_MEM_USR_RD_SRAM_DUMMY_W<'_, SPI_MEM_CACHE_SCTRL_SPEC>

Bit 4 - For SPI0, In the external RAM mode, it is the enable bit of dummy phase for read operations.

pub fn spi_mem_cache_sram_usr_rcmd( &mut self ) -> SPI_MEM_CACHE_SRAM_USR_RCMD_W<'_, SPI_MEM_CACHE_SCTRL_SPEC>

Bit 5 - For SPI0, In the external RAM mode cache read external RAM for user define command.

pub fn spi_mem_sram_rdummy_cyclelen( &mut self ) -> SPI_MEM_SRAM_RDUMMY_CYCLELEN_W<'_, SPI_MEM_CACHE_SCTRL_SPEC>

Bits 6:11 - For SPI0, In the external RAM mode, it is the length in bits of read dummy phase. The register value shall be (bit_num-1).

pub fn spi_mem_sram_addr_bitlen( &mut self ) -> SPI_MEM_SRAM_ADDR_BITLEN_W<'_, SPI_MEM_CACHE_SCTRL_SPEC>

Bits 14:19 - For SPI0, In the external RAM mode, it is the length in bits of address phase. The register value shall be (bit_num-1).

pub fn spi_mem_cache_sram_usr_wcmd( &mut self ) -> SPI_MEM_CACHE_SRAM_USR_WCMD_W<'_, SPI_MEM_CACHE_SCTRL_SPEC>

Bit 20 - For SPI0, In the external RAM mode cache write sram for user define command

pub fn spi_mem_sram_oct( &mut self ) -> SPI_MEM_SRAM_OCT_W<'_, SPI_MEM_CACHE_SCTRL_SPEC>

Bit 21 - reserved

pub fn spi_mem_sram_wdummy_cyclelen( &mut self ) -> SPI_MEM_SRAM_WDUMMY_CYCLELEN_W<'_, SPI_MEM_CACHE_SCTRL_SPEC>

Bits 22:27 - For SPI0, In the external RAM mode, it is the length in bits of write dummy phase. The register value shall be (bit_num-1).

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual