Module esp32p4::spi0::spi_mem_ecc_ctrl

MSPI ECC control register

Structs

SPI_MEM_ECC_CTRL_SPEC

MSPI ECC control register

Type Aliases

R

Register SPI_MEM_ECC_CTRL reader

SPI_FMEM_ECC_ADDR_EN_R

Field SPI_FMEM_ECC_ADDR_EN reader - Set this bit to enable MSPI ECC address conversion, no matter MSPI accesses to the ECC region or non-ECC region of flash. If there is no ECC region in flash, this bit should be 0. Otherwise, this bit should be 1.

SPI_FMEM_ECC_ADDR_EN_W

Field SPI_FMEM_ECC_ADDR_EN writer - Set this bit to enable MSPI ECC address conversion, no matter MSPI accesses to the ECC region or non-ECC region of flash. If there is no ECC region in flash, this bit should be 0. Otherwise, this bit should be 1.

SPI_FMEM_ECC_ERR_INT_EN_R

Field SPI_FMEM_ECC_ERR_INT_EN reader - Set this bit to calculate the error times of MSPI ECC read when accesses to flash.

SPI_FMEM_ECC_ERR_INT_EN_W

Field SPI_FMEM_ECC_ERR_INT_EN writer - Set this bit to calculate the error times of MSPI ECC read when accesses to flash.

SPI_FMEM_ECC_ERR_INT_NUM_R

Field SPI_FMEM_ECC_ERR_INT_NUM reader - Set the error times of MSPI ECC read to generate MSPI SPI_MEM_ECC_ERR_INT interrupt.

SPI_FMEM_ECC_ERR_INT_NUM_W

Field SPI_FMEM_ECC_ERR_INT_NUM writer - Set the error times of MSPI ECC read to generate MSPI SPI_MEM_ECC_ERR_INT interrupt.

SPI_FMEM_PAGE_SIZE_R

Field SPI_FMEM_PAGE_SIZE reader - Set the page size of the flash accessed by MSPI. 0: 256 bytes. 1: 512 bytes. 2: 1024 bytes. 3: 2048 bytes.

SPI_FMEM_PAGE_SIZE_W

Field SPI_FMEM_PAGE_SIZE writer - Set the page size of the flash accessed by MSPI. 0: 256 bytes. 1: 512 bytes. 2: 1024 bytes. 3: 2048 bytes.

SPI_MEM_ECC_CONTINUE_RECORD_ERR_EN_R

Field SPI_MEM_ECC_CONTINUE_RECORD_ERR_EN reader - 1: The error information in SPI_MEM_ECC_ERR_BITS and SPI_MEM_ECC_ERR_ADDR is updated when there is an ECC error. 0: SPI_MEM_ECC_ERR_BITS and SPI_MEM_ECC_ERR_ADDR record the first ECC error information.

SPI_MEM_ECC_CONTINUE_RECORD_ERR_EN_W

Field SPI_MEM_ECC_CONTINUE_RECORD_ERR_EN writer - 1: The error information in SPI_MEM_ECC_ERR_BITS and SPI_MEM_ECC_ERR_ADDR is updated when there is an ECC error. 0: SPI_MEM_ECC_ERR_BITS and SPI_MEM_ECC_ERR_ADDR record the first ECC error information.

SPI_MEM_ECC_ERR_BITS_R

Field SPI_MEM_ECC_ERR_BITS reader - Records the first ECC error bit number in the 16 bytes(From 0~127, corresponding to byte 0 bit 0 to byte 15 bit 7)

SPI_MEM_ECC_ERR_CNT_R

Field SPI_MEM_ECC_ERR_CNT reader - This bits show the error times of MSPI ECC read. It is cleared by when SPI_MEM_ECC_ERR_INT_CLR bit is set.

SPI_MEM_USR_ECC_ADDR_EN_R

Field SPI_MEM_USR_ECC_ADDR_EN reader - Set this bit to enable ECC address convert in SPI0/1 USR_CMD transfer.

SPI_MEM_USR_ECC_ADDR_EN_W

Field SPI_MEM_USR_ECC_ADDR_EN writer - Set this bit to enable ECC address convert in SPI0/1 USR_CMD transfer.

W

Register SPI_MEM_ECC_CTRL writer