Type Alias esp32p4::spi0::spi_smem_ac::W

pub type W = W<SPI_SMEM_AC_SPEC>;

Register SPI_SMEM_AC writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn spi_smem_cs_setup(&mut self) -> SPI_SMEM_CS_SETUP_W<'_, SPI_SMEM_AC_SPEC>

Bit 0 - For SPI0 and SPI1, spi cs is enable when spi is in prepare phase. 1: enable 0: disable.

pub fn spi_smem_cs_hold(&mut self) -> SPI_SMEM_CS_HOLD_W<'_, SPI_SMEM_AC_SPEC>

Bit 1 - For SPI0 and SPI1, spi cs keep low when spi is in done phase. 1: enable 0: disable.

pub fn spi_smem_cs_setup_time( &mut self ) -> SPI_SMEM_CS_SETUP_TIME_W<'_, SPI_SMEM_AC_SPEC>

Bits 2:6 - For spi0, (cycles-1) of prepare phase by spi clock this bits are combined with spi_mem_cs_setup bit.

pub fn spi_smem_cs_hold_time( &mut self ) -> SPI_SMEM_CS_HOLD_TIME_W<'_, SPI_SMEM_AC_SPEC>

Bits 7:11 - For SPI0 and SPI1, spi cs signal is delayed to inactive by spi clock this bits are combined with spi_mem_cs_hold bit.

pub fn spi_smem_ecc_cs_hold_time( &mut self ) -> SPI_SMEM_ECC_CS_HOLD_TIME_W<'_, SPI_SMEM_AC_SPEC>

Bits 12:14 - SPI_SMEM_CS_HOLD_TIME + SPI_SMEM_ECC_CS_HOLD_TIME is the SPI0 and SPI1 CS hold cycles in ECC mode when accessed external RAM.

pub fn spi_smem_ecc_skip_page_corner( &mut self ) -> SPI_SMEM_ECC_SKIP_PAGE_CORNER_W<'_, SPI_SMEM_AC_SPEC>

Bit 15 - 1: SPI0 skips page corner when accesses external RAM. 0: Not skip page corner when accesses external RAM.

pub fn spi_smem_ecc_16to18_byte_en( &mut self ) -> SPI_SMEM_ECC_16TO18_BYTE_EN_W<'_, SPI_SMEM_AC_SPEC>

Bit 16 - Set this bit to enable SPI0 and SPI1 ECC 16 bytes data with 2 ECC bytes mode when accesses external RAM.

pub fn spi_smem_cs_hold_delay( &mut self ) -> SPI_SMEM_CS_HOLD_DELAY_W<'_, SPI_SMEM_AC_SPEC>

Bits 25:30 - These bits are used to set the minimum CS high time tSHSL between SPI burst transfer when accesses to external RAM. tSHSL is (SPI_SMEM_CS_HOLD_DELAY[5:0] + 1) MSPI core clock cycles.

pub fn spi_smem_split_trans_en( &mut self ) -> SPI_SMEM_SPLIT_TRANS_EN_W<'_, SPI_SMEM_AC_SPEC>

Bit 31 - Set this bit to enable SPI0 split one AXI accesses EXT_RAM transfer into two SPI transfers when one transfer will cross flash/EXT_RAM page corner, valid no matter whether there is an ECC region or not.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual