Type Alias esp32p4::spi0::spi_smem_dout_mode::W

pub type W = W<SPI_SMEM_DOUT_MODE_SPEC>;

Register SPI_SMEM_DOUT_MODE writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn spi_smem_dout0_mode( &mut self ) -> SPI_SMEM_DOUT0_MODE_W<'_, SPI_SMEM_DOUT_MODE_SPEC>

Bit 0 - the output signals are delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the posedge of clk_160,4 output with the negedge of clk_160,5: output with the spi_clk high edge ,6: output with the spi_clk low edge

pub fn spi_smem_dout1_mode( &mut self ) -> SPI_SMEM_DOUT1_MODE_W<'_, SPI_SMEM_DOUT_MODE_SPEC>

Bit 1 - the output signals are delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the posedge of clk_160,4 output with the negedge of clk_160,5: output with the spi_clk high edge ,6: output with the spi_clk low edge

pub fn spi_smem_dout2_mode( &mut self ) -> SPI_SMEM_DOUT2_MODE_W<'_, SPI_SMEM_DOUT_MODE_SPEC>

Bit 2 - the output signals are delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the posedge of clk_160,4 output with the negedge of clk_160,5: output with the spi_clk high edge ,6: output with the spi_clk low edge

pub fn spi_smem_dout3_mode( &mut self ) -> SPI_SMEM_DOUT3_MODE_W<'_, SPI_SMEM_DOUT_MODE_SPEC>

Bit 3 - the output signals are delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the posedge of clk_160,4 output with the negedge of clk_160,5: output with the spi_clk high edge ,6: output with the spi_clk low edge

pub fn spi_smem_dout4_mode( &mut self ) -> SPI_SMEM_DOUT4_MODE_W<'_, SPI_SMEM_DOUT_MODE_SPEC>

Bit 4 - the output signals are delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the posedge of clk_160,4 output with the negedge of clk_160,5: output with the spi_clk high edge ,6: output with the spi_clk low edge

pub fn spi_smem_dout5_mode( &mut self ) -> SPI_SMEM_DOUT5_MODE_W<'_, SPI_SMEM_DOUT_MODE_SPEC>

Bit 5 - the output signals are delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the posedge of clk_160,4 output with the negedge of clk_160,5: output with the spi_clk high edge ,6: output with the spi_clk low edge

pub fn spi_smem_dout6_mode( &mut self ) -> SPI_SMEM_DOUT6_MODE_W<'_, SPI_SMEM_DOUT_MODE_SPEC>

Bit 6 - the output signals are delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the posedge of clk_160,4 output with the negedge of clk_160,5: output with the spi_clk high edge ,6: output with the spi_clk low edge

pub fn spi_smem_dout7_mode( &mut self ) -> SPI_SMEM_DOUT7_MODE_W<'_, SPI_SMEM_DOUT_MODE_SPEC>

Bit 7 - the output signals are delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the posedge of clk_160,4 output with the negedge of clk_160,5: output with the spi_clk high edge ,6: output with the spi_clk low edge

pub fn spi_smem_douts_mode( &mut self ) -> SPI_SMEM_DOUTS_MODE_W<'_, SPI_SMEM_DOUT_MODE_SPEC>

Bit 8 - the output signals are delayed by system clock cycles, 0: output without delayed, 1: output with the posedge of clk_apb,2 output with the negedge of clk_apb, 3: output with the posedge of clk_160,4 output with the negedge of clk_160,5: output with the spi_clk high edge ,6: output with the spi_clk low edge

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual