Module esp32p4::spi1::spi_mem_sus_status

SPI1 flash suspend status register

Structs

SPI_MEM_SUS_STATUS_SPEC

SPI1 flash suspend status register

Type Aliases

R

Register SPI_MEM_SUS_STATUS reader

SPI_MEM_FLASH_DP_DLY_128_R

Field SPI_MEM_FLASH_DP_DLY_128 reader - 1: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 128) SPI_CLK cycles after DP command is sent. 0: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 4) SPI_CLK cycles after DP command is sent.

SPI_MEM_FLASH_DP_DLY_128_W

Field SPI_MEM_FLASH_DP_DLY_128 writer - 1: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 128) SPI_CLK cycles after DP command is sent. 0: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 4) SPI_CLK cycles after DP command is sent.

SPI_MEM_FLASH_HPM_DLY_128_R

Field SPI_MEM_FLASH_HPM_DLY_128 reader - 1: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 128) SPI_CLK cycles after HPM command is sent. 0: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 4) SPI_CLK cycles after HPM command is sent.

SPI_MEM_FLASH_HPM_DLY_128_W

Field SPI_MEM_FLASH_HPM_DLY_128 writer - 1: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 128) SPI_CLK cycles after HPM command is sent. 0: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 4) SPI_CLK cycles after HPM command is sent.

SPI_MEM_FLASH_PER_COMMAND_R

Field SPI_MEM_FLASH_PER_COMMAND reader - Program/Erase resume command.

SPI_MEM_FLASH_PER_COMMAND_W

Field SPI_MEM_FLASH_PER_COMMAND writer - Program/Erase resume command.

SPI_MEM_FLASH_PER_DLY_128_R

Field SPI_MEM_FLASH_PER_DLY_128 reader - Valid when SPI_MEM_FLASH_PER_WAIT_EN is 1. 1: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 128) SPI_CLK cycles after PER command is sent. 0: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 4) SPI_CLK cycles after PER command is sent.

SPI_MEM_FLASH_PER_DLY_128_W

Field SPI_MEM_FLASH_PER_DLY_128 writer - Valid when SPI_MEM_FLASH_PER_WAIT_EN is 1. 1: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 128) SPI_CLK cycles after PER command is sent. 0: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 4) SPI_CLK cycles after PER command is sent.

SPI_MEM_FLASH_PESR_CMD_2B_R

Field SPI_MEM_FLASH_PESR_CMD_2B reader - 1: The bit length of Program/Erase Suspend/Resume command is 16. 0: The bit length of Program/Erase Suspend/Resume command is 8.

SPI_MEM_FLASH_PESR_CMD_2B_W

Field SPI_MEM_FLASH_PESR_CMD_2B writer - 1: The bit length of Program/Erase Suspend/Resume command is 16. 0: The bit length of Program/Erase Suspend/Resume command is 8.

SPI_MEM_FLASH_PES_DLY_128_R

Field SPI_MEM_FLASH_PES_DLY_128 reader - Valid when SPI_MEM_FLASH_PES_WAIT_EN is 1. 1: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 128) SPI_CLK cycles after PES command is sent. 0: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 4) SPI_CLK cycles after PES command is sent.

SPI_MEM_FLASH_PES_DLY_128_W

Field SPI_MEM_FLASH_PES_DLY_128 writer - Valid when SPI_MEM_FLASH_PES_WAIT_EN is 1. 1: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 128) SPI_CLK cycles after PES command is sent. 0: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 4) SPI_CLK cycles after PES command is sent.

SPI_MEM_FLASH_RES_DLY_128_R

Field SPI_MEM_FLASH_RES_DLY_128 reader - 1: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 128) SPI_CLK cycles after RES command is sent. 0: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 4) SPI_CLK cycles after RES command is sent.

SPI_MEM_FLASH_RES_DLY_128_W

Field SPI_MEM_FLASH_RES_DLY_128 writer - 1: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 128) SPI_CLK cycles after RES command is sent. 0: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 4) SPI_CLK cycles after RES command is sent.

SPI_MEM_FLASH_SUS_R

Field SPI_MEM_FLASH_SUS reader - The status of flash suspend, only used in SPI1.

SPI_MEM_FLASH_SUS_W

Field SPI_MEM_FLASH_SUS writer - The status of flash suspend, only used in SPI1.

SPI_MEM_SPI0_LOCK_EN_R

Field SPI_MEM_SPI0_LOCK_EN reader - 1: Enable SPI0 lock SPI0/1 arbiter option. 0: Disable it.

SPI_MEM_SPI0_LOCK_EN_W

Field SPI_MEM_SPI0_LOCK_EN writer - 1: Enable SPI0 lock SPI0/1 arbiter option. 0: Disable it.

SPI_MEM_WAIT_PESR_CMD_2B_R

Field SPI_MEM_WAIT_PESR_CMD_2B reader - 1: SPI1 sends out SPI_MEM_WAIT_PESR_COMMAND[15:0] to check SUS/SUS1/SUS2 bit. 0: SPI1 sends out SPI_MEM_WAIT_PESR_COMMAND[7:0] to check SUS/SUS1/SUS2 bit.

SPI_MEM_WAIT_PESR_CMD_2B_W

Field SPI_MEM_WAIT_PESR_CMD_2B writer - 1: SPI1 sends out SPI_MEM_WAIT_PESR_COMMAND[15:0] to check SUS/SUS1/SUS2 bit. 0: SPI1 sends out SPI_MEM_WAIT_PESR_COMMAND[7:0] to check SUS/SUS1/SUS2 bit.

W

Register SPI_MEM_SUS_STATUS writer