Module esp32p4::spi2::spi_clock

SPI clock control register

Structs

SPI_CLOCK_SPEC

SPI clock control register

Type Aliases

R

Register SPI_CLOCK reader

SPI_CLKCNT_H_R

Field SPI_CLKCNT_H reader - In the master mode it must be floor((spi_clkcnt_N+1)/2-1). In the slave mode it must be 0. Can be configured in CONF state.

SPI_CLKCNT_H_W

Field SPI_CLKCNT_H writer - In the master mode it must be floor((spi_clkcnt_N+1)/2-1). In the slave mode it must be 0. Can be configured in CONF state.

SPI_CLKCNT_L_R

Field SPI_CLKCNT_L reader - In the master mode it must be equal to spi_clkcnt_N. In the slave mode it must be 0. Can be configured in CONF state.

SPI_CLKCNT_L_W

Field SPI_CLKCNT_L writer - In the master mode it must be equal to spi_clkcnt_N. In the slave mode it must be 0. Can be configured in CONF state.

SPI_CLKCNT_N_R

Field SPI_CLKCNT_N reader - In the master mode it is the divider of spi_clk. So spi_clk frequency is system/(spi_clkdiv_pre+1)/(spi_clkcnt_N+1). Can be configured in CONF state.

SPI_CLKCNT_N_W

Field SPI_CLKCNT_N writer - In the master mode it is the divider of spi_clk. So spi_clk frequency is system/(spi_clkdiv_pre+1)/(spi_clkcnt_N+1). Can be configured in CONF state.

SPI_CLKDIV_PRE_R

Field SPI_CLKDIV_PRE reader - In the master mode it is pre-divider of spi_clk. Can be configured in CONF state.

SPI_CLKDIV_PRE_W

Field SPI_CLKDIV_PRE writer - In the master mode it is pre-divider of spi_clk. Can be configured in CONF state.

SPI_CLK_EQU_SYSCLK_R

Field SPI_CLK_EQU_SYSCLK reader - In the master mode 1: spi_clk is eqaul to system 0: spi_clk is divided from system clock. Can be configured in CONF state.

SPI_CLK_EQU_SYSCLK_W

Field SPI_CLK_EQU_SYSCLK writer - In the master mode 1: spi_clk is eqaul to system 0: spi_clk is divided from system clock. Can be configured in CONF state.

W

Register SPI_CLOCK writer