Type Alias esp32p4::spi2::spi_misc::R

pub type R = R<SPI_MISC_SPEC>;

Register SPI_MISC reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn spi_cs0_dis(&self) -> SPI_CS0_DIS_R

Bit 0 - SPI CS0 pin enable, 1: disable CS0, 0: spi_cs0 signal is from/to CS0 pin. Can be configured in CONF state.

pub fn spi_cs1_dis(&self) -> SPI_CS1_DIS_R

Bit 1 - SPI CS1 pin enable, 1: disable CS1, 0: spi_cs1 signal is from/to CS1 pin. Can be configured in CONF state.

pub fn spi_cs2_dis(&self) -> SPI_CS2_DIS_R

Bit 2 - SPI CS2 pin enable, 1: disable CS2, 0: spi_cs2 signal is from/to CS2 pin. Can be configured in CONF state.

pub fn spi_cs3_dis(&self) -> SPI_CS3_DIS_R

Bit 3 - SPI CS3 pin enable, 1: disable CS3, 0: spi_cs3 signal is from/to CS3 pin. Can be configured in CONF state.

pub fn spi_cs4_dis(&self) -> SPI_CS4_DIS_R

Bit 4 - SPI CS4 pin enable, 1: disable CS4, 0: spi_cs4 signal is from/to CS4 pin. Can be configured in CONF state.

pub fn spi_cs5_dis(&self) -> SPI_CS5_DIS_R

Bit 5 - SPI CS5 pin enable, 1: disable CS5, 0: spi_cs5 signal is from/to CS5 pin. Can be configured in CONF state.

pub fn spi_ck_dis(&self) -> SPI_CK_DIS_R

Bit 6 - 1: spi clk out disable, 0: spi clk out enable. Can be configured in CONF state.

pub fn spi_master_cs_pol(&self) -> SPI_MASTER_CS_POL_R

Bits 7:12 - In the master mode the bits are the polarity of spi cs line, the value is equivalent to spi_cs ^ spi_master_cs_pol. Can be configured in CONF state.

pub fn spi_clk_data_dtr_en(&self) -> SPI_CLK_DATA_DTR_EN_R

Bit 16 - 1: SPI master DTR mode is applied to SPI clk, data and spi_dqs. 0: SPI master DTR mode is only applied to spi_dqs. This bit should be used with bit 17/18/19.

pub fn spi_data_dtr_en(&self) -> SPI_DATA_DTR_EN_R

Bit 17 - 1: SPI clk and data of SPI_DOUT and SPI_DIN state are in DTR mode, including master 1/2/4/8-bm. 0: SPI clk and data of SPI_DOUT and SPI_DIN state are in STR mode. Can be configured in CONF state.

pub fn spi_addr_dtr_en(&self) -> SPI_ADDR_DTR_EN_R

Bit 18 - 1: SPI clk and data of SPI_SEND_ADDR state are in DTR mode, including master 1/2/4/8-bm. 0: SPI clk and data of SPI_SEND_ADDR state are in STR mode. Can be configured in CONF state.

pub fn spi_cmd_dtr_en(&self) -> SPI_CMD_DTR_EN_R

Bit 19 - 1: SPI clk and data of SPI_SEND_CMD state are in DTR mode, including master 1/2/4/8-bm. 0: SPI clk and data of SPI_SEND_CMD state are in STR mode. Can be configured in CONF state.

pub fn spi_slave_cs_pol(&self) -> SPI_SLAVE_CS_POL_R

Bit 23 - spi slave input cs polarity select. 1: inv 0: not change. Can be configured in CONF state.

pub fn spi_dqs_idle_edge(&self) -> SPI_DQS_IDLE_EDGE_R

Bit 24 - The default value of spi_dqs. Can be configured in CONF state.

pub fn spi_ck_idle_edge(&self) -> SPI_CK_IDLE_EDGE_R

Bit 29 - 1: spi clk line is high when idle 0: spi clk line is low when idle. Can be configured in CONF state.

pub fn spi_cs_keep_active(&self) -> SPI_CS_KEEP_ACTIVE_R

Bit 30 - spi cs line keep low when the bit is set. Can be configured in CONF state.

pub fn spi_quad_din_pin_swap(&self) -> SPI_QUAD_DIN_PIN_SWAP_R

Bit 31 - 1: SPI quad input swap enable, swap FSPID with FSPIQ, swap FSPIWP with FSPIHD. 0: spi quad input swap disable. Can be configured in CONF state.