Type Alias esp32p4::spi2::spi_slave::W

pub type W = W<SPI_SLAVE_SPEC>;

Register SPI_SLAVE writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn spi_clk_mode(&mut self) -> SPI_CLK_MODE_W<'_, SPI_SLAVE_SPEC>

Bits 0:1 - SPI clock mode bits. 0: SPI clock is off when CS inactive 1: SPI clock is delayed one cycle after CS inactive 2: SPI clock is delayed two cycles after CS inactive 3: SPI clock is alwasy on. Can be configured in CONF state.

pub fn spi_clk_mode_13(&mut self) -> SPI_CLK_MODE_13_W<'_, SPI_SLAVE_SPEC>

Bit 2 - {CPOL, CPHA},1: support spi clk mode 1 and 3, first edge output data B[0]/B[7]. 0: support spi clk mode 0 and 2, first edge output data B[1]/B[6].

pub fn spi_rsck_data_out(&mut self) -> SPI_RSCK_DATA_OUT_W<'_, SPI_SLAVE_SPEC>

Bit 3 - It saves half a cycle when tsck is the same as rsck. 1: output data at rsck posedge 0: output data at tsck posedge

pub fn spi_slv_rddma_bitlen_en( &mut self ) -> SPI_SLV_RDDMA_BITLEN_EN_W<'_, SPI_SLAVE_SPEC>

Bit 8 - 1: SPI_SLV_DATA_BITLEN stores data bit length of master-read-slave data length in DMA controlled mode(Rd_DMA). 0: others

pub fn spi_slv_wrdma_bitlen_en( &mut self ) -> SPI_SLV_WRDMA_BITLEN_EN_W<'_, SPI_SLAVE_SPEC>

Bit 9 - 1: SPI_SLV_DATA_BITLEN stores data bit length of master-write-to-slave data length in DMA controlled mode(Wr_DMA). 0: others

pub fn spi_slv_rdbuf_bitlen_en( &mut self ) -> SPI_SLV_RDBUF_BITLEN_EN_W<'_, SPI_SLAVE_SPEC>

Bit 10 - 1: SPI_SLV_DATA_BITLEN stores data bit length of master-read-slave data length in CPU controlled mode(Rd_BUF). 0: others

pub fn spi_slv_wrbuf_bitlen_en( &mut self ) -> SPI_SLV_WRBUF_BITLEN_EN_W<'_, SPI_SLAVE_SPEC>

Bit 11 - 1: SPI_SLV_DATA_BITLEN stores data bit length of master-write-to-slave data length in CPU controlled mode(Wr_BUF). 0: others

pub fn spi_dma_seg_magic_value( &mut self ) -> SPI_DMA_SEG_MAGIC_VALUE_W<'_, SPI_SLAVE_SPEC>

Bits 22:25 - The magic value of BM table in master DMA seg-trans.

pub fn mode(&mut self) -> MODE_W<'_, SPI_SLAVE_SPEC>

Bit 26 - Set SPI work mode. 1: slave mode 0: master mode.

pub fn spi_soft_reset(&mut self) -> SPI_SOFT_RESET_W<'_, SPI_SLAVE_SPEC>

Bit 27 - Software reset enable, reset the spi clock line cs line and data lines. Can be configured in CONF state.

pub fn spi_usr_conf(&mut self) -> SPI_USR_CONF_W<'_, SPI_SLAVE_SPEC>

Bit 28 - 1: Enable the DMA CONF phase of current seg-trans operation, which means seg-trans will start. 0: This is not seg-trans mode.

pub fn spi_mst_fd_wait_dma_tx_data( &mut self ) -> SPI_MST_FD_WAIT_DMA_TX_DATA_W<'_, SPI_SLAVE_SPEC>

Bit 29 - In master full-duplex mode, 1: GP-SPI will wait DMA TX data is ready before starting SPI transfer. 0: GP-SPI does not wait DMA TX data before starting SPI transfer.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual