Module esp32p4::spi3::spi_dma_conf

SPI DMA control register

Structs

SPI_DMA_CONF_SPEC

SPI DMA control register

Type Aliases

R

Register SPI_DMA_CONF reader

SPI_BUF_AFIFO_RST_W

Field SPI_BUF_AFIFO_RST writer - Set this bit to reset BUF TX AFIFO, which is used send data out in SPI slave CPU controlled mode transfer and master mode transfer.

SPI_DMA_AFIFO_RST_W

Field SPI_DMA_AFIFO_RST writer - Set this bit to reset DMA TX AFIFO, which is used to send data out in SPI slave DMA controlled mode transfer.

SPI_DMA_INFIFO_FULL_R

Field SPI_DMA_INFIFO_FULL reader - Records the status of DMA RX FIFO. 1: DMA RX FIFO is not ready for receiving data. 0: DMA RX FIFO is ready for receiving data.

SPI_DMA_OUTFIFO_EMPTY_R

Field SPI_DMA_OUTFIFO_EMPTY reader - Records the status of DMA TX FIFO. 1: DMA TX FIFO is not ready for sending data. 0: DMA TX FIFO is ready for sending data.

SPI_DMA_RX_ENA_R

Field SPI_DMA_RX_ENA reader - Set this bit to enable SPI DMA controlled receive data mode.

SPI_DMA_RX_ENA_W

Field SPI_DMA_RX_ENA writer - Set this bit to enable SPI DMA controlled receive data mode.

SPI_DMA_SLV_SEG_TRANS_EN_R

Field SPI_DMA_SLV_SEG_TRANS_EN reader - Enable dma segment transfer in spi dma half slave mode. 1: enable. 0: disable.

SPI_DMA_SLV_SEG_TRANS_EN_W

Field SPI_DMA_SLV_SEG_TRANS_EN writer - Enable dma segment transfer in spi dma half slave mode. 1: enable. 0: disable.

SPI_DMA_TX_ENA_R

Field SPI_DMA_TX_ENA reader - Set this bit to enable SPI DMA controlled send data mode.

SPI_DMA_TX_ENA_W

Field SPI_DMA_TX_ENA writer - Set this bit to enable SPI DMA controlled send data mode.

SPI_RX_AFIFO_RST_W

Field SPI_RX_AFIFO_RST writer - Set this bit to reset RX AFIFO, which is used to receive data in SPI master and slave mode transfer.

SPI_RX_EOF_EN_R

Field SPI_RX_EOF_EN reader - 1: spi_dma_inlink_eof is set when the number of dma pushed data bytes is equal to the value of spi_slv/mst_dma_rd_bytelen[19:0] in spi dma transition. 0: spi_dma_inlink_eof is set by spi_trans_done in non-seg-trans or spi_dma_seg_trans_done in seg-trans.

SPI_RX_EOF_EN_W

Field SPI_RX_EOF_EN writer - 1: spi_dma_inlink_eof is set when the number of dma pushed data bytes is equal to the value of spi_slv/mst_dma_rd_bytelen[19:0] in spi dma transition. 0: spi_dma_inlink_eof is set by spi_trans_done in non-seg-trans or spi_dma_seg_trans_done in seg-trans.

SPI_SLV_RX_SEG_TRANS_CLR_EN_R

Field SPI_SLV_RX_SEG_TRANS_CLR_EN reader - 1: spi_dma_infifo_full_vld is cleared by spi slave cmd 5. 0: spi_dma_infifo_full_vld is cleared by spi_trans_done.

SPI_SLV_RX_SEG_TRANS_CLR_EN_W

Field SPI_SLV_RX_SEG_TRANS_CLR_EN writer - 1: spi_dma_infifo_full_vld is cleared by spi slave cmd 5. 0: spi_dma_infifo_full_vld is cleared by spi_trans_done.

SPI_SLV_TX_SEG_TRANS_CLR_EN_R

Field SPI_SLV_TX_SEG_TRANS_CLR_EN reader - 1: spi_dma_outfifo_empty_vld is cleared by spi slave cmd 6. 0: spi_dma_outfifo_empty_vld is cleared by spi_trans_done.

SPI_SLV_TX_SEG_TRANS_CLR_EN_W

Field SPI_SLV_TX_SEG_TRANS_CLR_EN writer - 1: spi_dma_outfifo_empty_vld is cleared by spi slave cmd 6. 0: spi_dma_outfifo_empty_vld is cleared by spi_trans_done.

W

Register SPI_DMA_CONF writer