Module esp32p4::spi3::spi_misc

SPI misc register

Structs

SPI_MISC_SPEC

SPI misc register

Type Aliases

R

Register SPI_MISC reader

SPI_CK_DIS_R

Field SPI_CK_DIS reader - 1: spi clk out disable, 0: spi clk out enable. Can be configured in CONF state.

SPI_CK_DIS_W

Field SPI_CK_DIS writer - 1: spi clk out disable, 0: spi clk out enable. Can be configured in CONF state.

SPI_CK_IDLE_EDGE_R

Field SPI_CK_IDLE_EDGE reader - 1: spi clk line is high when idle 0: spi clk line is low when idle. Can be configured in CONF state.

SPI_CK_IDLE_EDGE_W

Field SPI_CK_IDLE_EDGE writer - 1: spi clk line is high when idle 0: spi clk line is low when idle. Can be configured in CONF state.

SPI_CS0_DIS_R

Field SPI_CS0_DIS reader - SPI CS0 pin enable, 1: disable CS0, 0: spi_cs0 signal is from/to CS0 pin. Can be configured in CONF state.

SPI_CS0_DIS_W

Field SPI_CS0_DIS writer - SPI CS0 pin enable, 1: disable CS0, 0: spi_cs0 signal is from/to CS0 pin. Can be configured in CONF state.

SPI_CS1_DIS_R

Field SPI_CS1_DIS reader - SPI CS1 pin enable, 1: disable CS1, 0: spi_cs1 signal is from/to CS1 pin. Can be configured in CONF state.

SPI_CS1_DIS_W

Field SPI_CS1_DIS writer - SPI CS1 pin enable, 1: disable CS1, 0: spi_cs1 signal is from/to CS1 pin. Can be configured in CONF state.

SPI_CS2_DIS_R

Field SPI_CS2_DIS reader - SPI CS2 pin enable, 1: disable CS2, 0: spi_cs2 signal is from/to CS2 pin. Can be configured in CONF state.

SPI_CS2_DIS_W

Field SPI_CS2_DIS writer - SPI CS2 pin enable, 1: disable CS2, 0: spi_cs2 signal is from/to CS2 pin. Can be configured in CONF state.

SPI_CS_KEEP_ACTIVE_R

Field SPI_CS_KEEP_ACTIVE reader - spi cs line keep low when the bit is set. Can be configured in CONF state.

SPI_CS_KEEP_ACTIVE_W

Field SPI_CS_KEEP_ACTIVE writer - spi cs line keep low when the bit is set. Can be configured in CONF state.

SPI_MASTER_CS_POL_R

Field SPI_MASTER_CS_POL reader - In the master mode the bits are the polarity of spi cs line, the value is equivalent to spi_cs ^ spi_master_cs_pol. Can be configured in CONF state.

SPI_MASTER_CS_POL_W

Field SPI_MASTER_CS_POL writer - In the master mode the bits are the polarity of spi cs line, the value is equivalent to spi_cs ^ spi_master_cs_pol. Can be configured in CONF state.

SPI_QUAD_DIN_PIN_SWAP_R

Field SPI_QUAD_DIN_PIN_SWAP reader - 1: spi quad input swap enable 0: spi quad input swap disable. Can be configured in CONF state.

SPI_QUAD_DIN_PIN_SWAP_W

Field SPI_QUAD_DIN_PIN_SWAP writer - 1: spi quad input swap enable 0: spi quad input swap disable. Can be configured in CONF state.

SPI_SLAVE_CS_POL_R

Field SPI_SLAVE_CS_POL reader - spi slave input cs polarity select. 1: inv 0: not change. Can be configured in CONF state.

SPI_SLAVE_CS_POL_W

Field SPI_SLAVE_CS_POL writer - spi slave input cs polarity select. 1: inv 0: not change. Can be configured in CONF state.

W

Register SPI_MISC writer