Type Alias esp32p4::spi3::spi_misc::W

pub type W = W<SPI_MISC_SPEC>;

Register SPI_MISC writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn spi_cs0_dis(&mut self) -> SPI_CS0_DIS_W<'_, SPI_MISC_SPEC>

Bit 0 - SPI CS0 pin enable, 1: disable CS0, 0: spi_cs0 signal is from/to CS0 pin. Can be configured in CONF state.

pub fn spi_cs1_dis(&mut self) -> SPI_CS1_DIS_W<'_, SPI_MISC_SPEC>

Bit 1 - SPI CS1 pin enable, 1: disable CS1, 0: spi_cs1 signal is from/to CS1 pin. Can be configured in CONF state.

pub fn spi_cs2_dis(&mut self) -> SPI_CS2_DIS_W<'_, SPI_MISC_SPEC>

Bit 2 - SPI CS2 pin enable, 1: disable CS2, 0: spi_cs2 signal is from/to CS2 pin. Can be configured in CONF state.

pub fn spi_ck_dis(&mut self) -> SPI_CK_DIS_W<'_, SPI_MISC_SPEC>

Bit 6 - 1: spi clk out disable, 0: spi clk out enable. Can be configured in CONF state.

pub fn spi_master_cs_pol(&mut self) -> SPI_MASTER_CS_POL_W<'_, SPI_MISC_SPEC>

Bits 7:9 - In the master mode the bits are the polarity of spi cs line, the value is equivalent to spi_cs ^ spi_master_cs_pol. Can be configured in CONF state.

pub fn spi_slave_cs_pol(&mut self) -> SPI_SLAVE_CS_POL_W<'_, SPI_MISC_SPEC>

Bit 23 - spi slave input cs polarity select. 1: inv 0: not change. Can be configured in CONF state.

pub fn spi_ck_idle_edge(&mut self) -> SPI_CK_IDLE_EDGE_W<'_, SPI_MISC_SPEC>

Bit 29 - 1: spi clk line is high when idle 0: spi clk line is low when idle. Can be configured in CONF state.

pub fn spi_cs_keep_active(&mut self) -> SPI_CS_KEEP_ACTIVE_W<'_, SPI_MISC_SPEC>

Bit 30 - spi cs line keep low when the bit is set. Can be configured in CONF state.

pub fn spi_quad_din_pin_swap( &mut self ) -> SPI_QUAD_DIN_PIN_SWAP_W<'_, SPI_MISC_SPEC>

Bit 31 - 1: spi quad input swap enable 0: spi quad input swap disable. Can be configured in CONF state.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual