Type Alias esp32p4::spi3::spi_user::W

pub type W = W<SPI_USER_SPEC>;

Register SPI_USER writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn spi_doutdin(&mut self) -> SPI_DOUTDIN_W<'_, SPI_USER_SPEC>

Bit 0 - Set the bit to enable full duplex communication. 1: enable 0: disable. Can be configured in CONF state.

pub fn spi_qpi_mode(&mut self) -> SPI_QPI_MODE_W<'_, SPI_USER_SPEC>

Bit 3 - Both for master mode and slave mode. 1: spi controller is in QPI mode. 0: others. Can be configured in CONF state.

pub fn spi_tsck_i_edge(&mut self) -> SPI_TSCK_I_EDGE_W<'_, SPI_USER_SPEC>

Bit 5 - In the slave mode, this bit can be used to change the polarity of tsck. 0: tsck = spi_ck_i. 1:tsck = !spi_ck_i.

pub fn spi_cs_hold(&mut self) -> SPI_CS_HOLD_W<'_, SPI_USER_SPEC>

Bit 6 - spi cs keep low when spi is in done phase. 1: enable 0: disable. Can be configured in CONF state.

pub fn spi_cs_setup(&mut self) -> SPI_CS_SETUP_W<'_, SPI_USER_SPEC>

Bit 7 - spi cs is enable when spi is in prepare phase. 1: enable 0: disable. Can be configured in CONF state.

pub fn spi_rsck_i_edge(&mut self) -> SPI_RSCK_I_EDGE_W<'_, SPI_USER_SPEC>

Bit 8 - In the slave mode, this bit can be used to change the polarity of rsck. 0: rsck = !spi_ck_i. 1:rsck = spi_ck_i.

pub fn spi_ck_out_edge(&mut self) -> SPI_CK_OUT_EDGE_W<'_, SPI_USER_SPEC>

Bit 9 - the bit combined with spi_mosi_delay_mode bits to set mosi signal delay mode. Can be configured in CONF state.

pub fn spi_fwrite_dual(&mut self) -> SPI_FWRITE_DUAL_W<'_, SPI_USER_SPEC>

Bit 12 - In the write operations read-data phase apply 2 signals. Can be configured in CONF state.

pub fn spi_fwrite_quad(&mut self) -> SPI_FWRITE_QUAD_W<'_, SPI_USER_SPEC>

Bit 13 - In the write operations read-data phase apply 4 signals. Can be configured in CONF state.

pub fn spi_sio(&mut self) -> SPI_SIO_W<'_, SPI_USER_SPEC>

Bit 17 - Set the bit to enable 3-line half duplex communication mosi and miso signals share the same pin. 1: enable 0: disable. Can be configured in CONF state.

pub fn spi_usr_miso_highpart( &mut self ) -> SPI_USR_MISO_HIGHPART_W<'_, SPI_USER_SPEC>

Bit 24 - read-data phase only access to high-part of the buffer spi_w8~spi_w15. 1: enable 0: disable. Can be configured in CONF state.

pub fn spi_usr_mosi_highpart( &mut self ) -> SPI_USR_MOSI_HIGHPART_W<'_, SPI_USER_SPEC>

Bit 25 - write-data phase only access to high-part of the buffer spi_w8~spi_w15. 1: enable 0: disable. Can be configured in CONF state.

pub fn spi_usr_dummy_idle(&mut self) -> SPI_USR_DUMMY_IDLE_W<'_, SPI_USER_SPEC>

Bit 26 - spi clock is disable in dummy phase when the bit is enable. Can be configured in CONF state.

pub fn spi_usr_mosi(&mut self) -> SPI_USR_MOSI_W<'_, SPI_USER_SPEC>

Bit 27 - This bit enable the write-data phase of an operation. Can be configured in CONF state.

pub fn spi_usr_miso(&mut self) -> SPI_USR_MISO_W<'_, SPI_USER_SPEC>

Bit 28 - This bit enable the read-data phase of an operation. Can be configured in CONF state.

pub fn spi_usr_dummy(&mut self) -> SPI_USR_DUMMY_W<'_, SPI_USER_SPEC>

Bit 29 - This bit enable the dummy phase of an operation. Can be configured in CONF state.

pub fn spi_usr_addr(&mut self) -> SPI_USR_ADDR_W<'_, SPI_USER_SPEC>

Bit 30 - This bit enable the address phase of an operation. Can be configured in CONF state.

pub fn spi_usr_command(&mut self) -> SPI_USR_COMMAND_W<'_, SPI_USER_SPEC>

Bit 31 - This bit enable the command phase of an operation. Can be configured in CONF state.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual