Struct esp32p4::LP_AON_CLKRST

pub struct LP_AON_CLKRST { /* private fields */ }

LP_AON_CLKRST Peripheral

Implementations

impl LP_AON_CLKRST

pub const PTR: *const RegisterBlock = {0x50111000 as *const lp_aon_clkrst::RegisterBlock}

Pointer to the register block

pub const fn ptr() -> *const RegisterBlock

Return the pointer to the register block

pub unsafe fn steal() -> Self

Steal an instance of this peripheral

Safety

Ensure that the new instance of the peripheral cannot be used in a way that may race with any existing instances, for example by only accessing read-only or write-only registers, or by consuming the original peripheral and using critical sections to coordinate access between multiple new instances.

Additionally, other software such as HALs may rely on only one peripheral instance existing to ensure memory safety; ensure no stolen instances are passed to such software.