Struct esp32p4::LP_I2S0

pub struct LP_I2S0 { /* private fields */ }

Low-power I2S (Inter-IC Sound) Controller 0

Implementations

impl LP_I2S0

pub const PTR: *const RegisterBlock = {0x50125000 as *const lp_i2s0::RegisterBlock}

Pointer to the register block

pub const fn ptr() -> *const RegisterBlock

Return the pointer to the register block

pub unsafe fn steal() -> Self

Steal an instance of this peripheral

Safety

Ensure that the new instance of the peripheral cannot be used in a way that may race with any existing instances, for example by only accessing read-only or write-only registers, or by consuming the original peripheral and using critical sections to coordinate access between multiple new instances.

Additionally, other software such as HALs may rely on only one peripheral instance existing to ensure memory safety; ensure no stolen instances are passed to such software.