Module esp32p4::trace0::config

trace configuration register

Structs

CONFIG_SPEC

trace configuration register

Type Aliases

DM_TRIGGER_ENA_R

Field DM_TRIGGER_ENA reader - Configure whether or not enable cpu trigger action.\1: enable\0:disable\

DM_TRIGGER_ENA_W

Field DM_TRIGGER_ENA writer - Configure whether or not enable cpu trigger action.\1: enable\0:disable\

FULL_ADDRESS_R

Field FULL_ADDRESS reader - Configure whether or not enable full-address mode.\1: full address mode.\0: delta address mode\

FULL_ADDRESS_W

Field FULL_ADDRESS writer - Configure whether or not enable full-address mode.\1: full address mode.\0: delta address mode\

HALT_ENA_R

Field HALT_ENA reader - Configure whether or not enable trace cpu is halted, when enabeld, if the cpu halted, the encoder will output a packet to report the address of the last instruction, and upon halted deassertion, the encoder start again.When disabled, encoder will not report the last address before halted and first address after halted, cpu halted information will not be tracked. \1: enabeld\0: disabled\

HALT_ENA_W

Field HALT_ENA writer - Configure whether or not enable trace cpu is halted, when enabeld, if the cpu halted, the encoder will output a packet to report the address of the last instruction, and upon halted deassertion, the encoder start again.When disabled, encoder will not report the last address before halted and first address after halted, cpu halted information will not be tracked. \1: enabeld\0: disabled\

IMPLICIT_EXCEPT_R

Field IMPLICIT_EXCEPT reader - Configure whether or not enabel implicit exception mode. When enabled,, do not sent exception address, only exception cause in exception packets.\1: enabled\0: disabled\

IMPLICIT_EXCEPT_W

Field IMPLICIT_EXCEPT writer - Configure whether or not enabel implicit exception mode. When enabled,, do not sent exception address, only exception cause in exception packets.\1: enabled\0: disabled\

R

Register CONFIG reader

RESET_ENA_R

Field RESET_ENA reader - Configure whether or not enable trace cpu haverest, when enabeld, if cpu have reset, the encoder will output a packet to report the address of the last instruction, and upon reset deassertion, the encoder start again.\1: enabeld\0: disabled\

RESET_ENA_W

Field RESET_ENA writer - Configure whether or not enable trace cpu haverest, when enabeld, if cpu have reset, the encoder will output a packet to report the address of the last instruction, and upon reset deassertion, the encoder start again.\1: enabeld\0: disabled\

STALL_ENA_R

Field STALL_ENA reader - Configure whether or not enable stall cpu. When enabled, when the fifo almost full, the cpu will be stalled until the packets is able to write to fifo.\1: enabled.\0: disabled\

STALL_ENA_W

Field STALL_ENA writer - Configure whether or not enable stall cpu. When enabled, when the fifo almost full, the cpu will be stalled until the packets is able to write to fifo.\1: enabled.\0: disabled\

W

Register CONFIG writer