Module esp32p4::trace0::filter_comparator_control

filter comparator match control register

Structs

FILTER_COMPARATOR_CONTROL_SPEC

filter comparator match control register

Type Aliases

MATCH_MODE_R

Field MATCH_MODE reader - 0: only primary matches, \1: primary and secondary comparator both matches(P&amp;&amp;S),\ 2:either primary or secondary comparator matches !(P&amp;&amp;S), \3: set when primary matches and continue to match until after secondary comparator matches

MATCH_MODE_W

Field MATCH_MODE writer - 0: only primary matches, \1: primary and secondary comparator both matches(P&amp;&amp;S),\ 2:either primary or secondary comparator matches !(P&amp;&amp;S), \3: set when primary matches and continue to match until after secondary comparator matches

P_FUNCTION_R

Field P_FUNCTION reader - Select the primary comparator function. \0: equal, \1: not equal, \2: less than, \3: less than or equal, \4: greater than, \5: greater than or equal, \other: always match

P_FUNCTION_W

Field P_FUNCTION writer - Select the primary comparator function. \0: equal, \1: not equal, \2: less than, \3: less than or equal, \4: greater than, \5: greater than or equal, \other: always match

P_INPUT_R

Field P_INPUT reader - Determines which input to compare against the primary comparator, \0: iaddr, \1: tval.

P_INPUT_W

Field P_INPUT writer - Determines which input to compare against the primary comparator, \0: iaddr, \1: tval.

P_NOTIFY_R

Field P_NOTIFY reader - Generate a trace packet explicitly reporting the address that cause the primary match

P_NOTIFY_W

Field P_NOTIFY writer - Generate a trace packet explicitly reporting the address that cause the primary match

R

Register FILTER_COMPARATOR_CONTROL reader

S_FUNCTION_R

Field S_FUNCTION reader - Select the secondary comparator function. \0: equal, \1: not equal, \2: less than, \3: less than or equal, \4: greater than, \5: greater than or equal, \other: always match

S_FUNCTION_W

Field S_FUNCTION writer - Select the secondary comparator function. \0: equal, \1: not equal, \2: less than, \3: less than or equal, \4: greater than, \5: greater than or equal, \other: always match

S_INPUT_R

Field S_INPUT reader - Determines which input to compare against the secondary comparator, \0: iaddr, \1: tval.

S_INPUT_W

Field S_INPUT writer - Determines which input to compare against the secondary comparator, \0: iaddr, \1: tval.

S_NOTIFY_R

Field S_NOTIFY reader - Generate a trace packet explicitly reporting the address that cause the secondary match

S_NOTIFY_W

Field S_NOTIFY writer - Generate a trace packet explicitly reporting the address that cause the secondary match

W

Register FILTER_COMPARATOR_CONTROL writer