Type Alias esp32p4::trace0::filter_comparator_control::R

pub type R = R<FILTER_COMPARATOR_CONTROL_SPEC>;

Register FILTER_COMPARATOR_CONTROL reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn p_input(&self) -> P_INPUT_R

Bit 0 - Determines which input to compare against the primary comparator, \0: iaddr, \1: tval.

pub fn p_function(&self) -> P_FUNCTION_R

Bits 2:4 - Select the primary comparator function. \0: equal, \1: not equal, \2: less than, \3: less than or equal, \4: greater than, \5: greater than or equal, \other: always match

pub fn p_notify(&self) -> P_NOTIFY_R

Bit 5 - Generate a trace packet explicitly reporting the address that cause the primary match

pub fn s_input(&self) -> S_INPUT_R

Bit 8 - Determines which input to compare against the secondary comparator, \0: iaddr, \1: tval.

pub fn s_function(&self) -> S_FUNCTION_R

Bits 10:12 - Select the secondary comparator function. \0: equal, \1: not equal, \2: less than, \3: less than or equal, \4: greater than, \5: greater than or equal, \other: always match

pub fn s_notify(&self) -> S_NOTIFY_R

Bit 13 - Generate a trace packet explicitly reporting the address that cause the secondary match

pub fn match_mode(&self) -> MATCH_MODE_R

Bits 16:17 - 0: only primary matches, \1: primary and secondary comparator both matches(P&amp;&amp;S),\ 2:either primary or secondary comparator matches !(P&amp;&amp;S), \3: set when primary matches and continue to match until after secondary comparator matches