Type Alias esp32p4::trace0::filter_comparator_control::W

pub type W = W<FILTER_COMPARATOR_CONTROL_SPEC>;

Register FILTER_COMPARATOR_CONTROL writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn p_input(&mut self) -> P_INPUT_W<'_, FILTER_COMPARATOR_CONTROL_SPEC>

Bit 0 - Determines which input to compare against the primary comparator, \0: iaddr, \1: tval.

pub fn p_function(&mut self) -> P_FUNCTION_W<'_, FILTER_COMPARATOR_CONTROL_SPEC>

Bits 2:4 - Select the primary comparator function. \0: equal, \1: not equal, \2: less than, \3: less than or equal, \4: greater than, \5: greater than or equal, \other: always match

pub fn p_notify(&mut self) -> P_NOTIFY_W<'_, FILTER_COMPARATOR_CONTROL_SPEC>

Bit 5 - Generate a trace packet explicitly reporting the address that cause the primary match

pub fn s_input(&mut self) -> S_INPUT_W<'_, FILTER_COMPARATOR_CONTROL_SPEC>

Bit 8 - Determines which input to compare against the secondary comparator, \0: iaddr, \1: tval.

pub fn s_function(&mut self) -> S_FUNCTION_W<'_, FILTER_COMPARATOR_CONTROL_SPEC>

Bits 10:12 - Select the secondary comparator function. \0: equal, \1: not equal, \2: less than, \3: less than or equal, \4: greater than, \5: greater than or equal, \other: always match

pub fn s_notify(&mut self) -> S_NOTIFY_W<'_, FILTER_COMPARATOR_CONTROL_SPEC>

Bit 13 - Generate a trace packet explicitly reporting the address that cause the secondary match

pub fn match_mode(&mut self) -> MATCH_MODE_W<'_, FILTER_COMPARATOR_CONTROL_SPEC>

Bits 16:17 - 0: only primary matches, \1: primary and secondary comparator both matches(P&amp;&amp;S),\ 2:either primary or secondary comparator matches !(P&amp;&amp;S), \3: set when primary matches and continue to match until after secondary comparator matches

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual