Module esp32p4::trace0::filter_control

filter control register

Structs

FILTER_CONTROL_SPEC

filter control register

Type Aliases

FILTER_EN_R

Field FILTER_EN reader - Configure whether or not enable filter unit. \1: enable filter.\ 0: always match

FILTER_EN_W

Field FILTER_EN writer - Configure whether or not enable filter unit. \1: enable filter.\ 0: always match

MATCH_COMP_R

Field MATCH_COMP reader - when set, the comparator must be high in order for the filter to match

MATCH_COMP_W

Field MATCH_COMP writer - when set, the comparator must be high in order for the filter to match

MATCH_ECAUSE_R

Field MATCH_ECAUSE reader - when set, start matching from exception cause codes specified by \hyperref[fielddesc:TRACEMATCHCHOICEECAUSE]{TRACE_MATCH_CHOICE_ECAUSE}, and stop matching upon return from the 1st matching exception.

MATCH_ECAUSE_W

Field MATCH_ECAUSE writer - when set, start matching from exception cause codes specified by \hyperref[fielddesc:TRACEMATCHCHOICEECAUSE]{TRACE_MATCH_CHOICE_ECAUSE}, and stop matching upon return from the 1st matching exception.

MATCH_INTERRUPT_R

Field MATCH_INTERRUPT reader - when set, start matching from a trap with the interrupt level codes specified by \hyperref[fielddesc:TRACEMATCHVALUEINTERRUPT]{TRACE_MATCH_VALUE_INTERRUPT}, and stop matching upon return from the 1st matching trap.

MATCH_INTERRUPT_W

Field MATCH_INTERRUPT writer - when set, start matching from a trap with the interrupt level codes specified by \hyperref[fielddesc:TRACEMATCHVALUEINTERRUPT]{TRACE_MATCH_VALUE_INTERRUPT}, and stop matching upon return from the 1st matching trap.

MATCH_PRIVILEGE_R

Field MATCH_PRIVILEGE reader - when set, match privilege levels specified by \hyperref[fielddesc:TRACEMATCHCHOICEPRIVILEGE]{TRACE_MATCH_CHOICE_PRIVILEGE}.

MATCH_PRIVILEGE_W

Field MATCH_PRIVILEGE writer - when set, match privilege levels specified by \hyperref[fielddesc:TRACEMATCHCHOICEPRIVILEGE]{TRACE_MATCH_CHOICE_PRIVILEGE}.

R

Register FILTER_CONTROL reader

W

Register FILTER_CONTROL writer