Type Alias esp32p4::trace0::filter_control::W

pub type W = W<FILTER_CONTROL_SPEC>;

Register FILTER_CONTROL writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn filter_en(&mut self) -> FILTER_EN_W<'_, FILTER_CONTROL_SPEC>

Bit 0 - Configure whether or not enable filter unit. \1: enable filter.\ 0: always match

pub fn match_comp(&mut self) -> MATCH_COMP_W<'_, FILTER_CONTROL_SPEC>

Bit 1 - when set, the comparator must be high in order for the filter to match

pub fn match_privilege(&mut self) -> MATCH_PRIVILEGE_W<'_, FILTER_CONTROL_SPEC>

Bit 2 - when set, match privilege levels specified by \hyperref[fielddesc:TRACEMATCHCHOICEPRIVILEGE]{TRACE_MATCH_CHOICE_PRIVILEGE}.

pub fn match_ecause(&mut self) -> MATCH_ECAUSE_W<'_, FILTER_CONTROL_SPEC>

Bit 3 - when set, start matching from exception cause codes specified by \hyperref[fielddesc:TRACEMATCHCHOICEECAUSE]{TRACE_MATCH_CHOICE_ECAUSE}, and stop matching upon return from the 1st matching exception.

pub fn match_interrupt(&mut self) -> MATCH_INTERRUPT_W<'_, FILTER_CONTROL_SPEC>

Bit 4 - when set, start matching from a trap with the interrupt level codes specified by \hyperref[fielddesc:TRACEMATCHVALUEINTERRUPT]{TRACE_MATCH_VALUE_INTERRUPT}, and stop matching upon return from the 1st matching trap.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual