Module esp32p4::twai0::mode

TWAI mode register.

Structs

MODE_SPEC

TWAI mode register.

Type Aliases

ACCEPTANCE_FILTER_MODE_R

Field ACCEPTANCE_FILTER_MODE reader - 1:single, the single acceptance filter option is enabled (one filter with the length of 32 bit is active). 0:dual, the dual acceptance filter option is enabled (two filters, each with the length of 16 bit are active).

ACCEPTANCE_FILTER_MODE_W

Field ACCEPTANCE_FILTER_MODE writer - 1:single, the single acceptance filter option is enabled (one filter with the length of 32 bit is active). 0:dual, the dual acceptance filter option is enabled (two filters, each with the length of 16 bit are active).

LISTEN_ONLY_MODE_R

Field LISTEN_ONLY_MODE reader - 1: listen only, in this mode the TWAI controller would give no acknowledge to the TWAI-bus, even if a message is received successfully. The error counters are stopped at the current value. 0: normal.

LISTEN_ONLY_MODE_W

Field LISTEN_ONLY_MODE writer - 1: listen only, in this mode the TWAI controller would give no acknowledge to the TWAI-bus, even if a message is received successfully. The error counters are stopped at the current value. 0: normal.

R

Register MODE reader

RESET_MODE_R

Field RESET_MODE reader - 1: reset, detection of a set reset mode bit results in aborting the current transmission/reception of a message and entering the reset mode. 0: normal, on the ‘1-to-0’ transition of the reset mode bit, the TWAI controller returns to the operating mode.

RESET_MODE_W

Field RESET_MODE writer - 1: reset, detection of a set reset mode bit results in aborting the current transmission/reception of a message and entering the reset mode. 0: normal, on the ‘1-to-0’ transition of the reset mode bit, the TWAI controller returns to the operating mode.

SELF_TEST_MODE_R

Field SELF_TEST_MODE reader - 1: self test, in this mode a full node test is possible without any other active node on the bus using the self reception request command. The TWAI controller will perform a successful transmission, even if there is no acknowledge received. 0: normal, an acknowledge is required for successful transmission.

SELF_TEST_MODE_W

Field SELF_TEST_MODE writer - 1: self test, in this mode a full node test is possible without any other active node on the bus using the self reception request command. The TWAI controller will perform a successful transmission, even if there is no acknowledge received. 0: normal, an acknowledge is required for successful transmission.

W

Register MODE writer