Type Alias esp32p4::uhci0::escape_conf::W

pub type W = W<ESCAPE_CONF_SPEC>;

Register ESCAPE_CONF writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn tx_c0_esc_en(&mut self) -> TX_C0_ESC_EN_W<'_, ESCAPE_CONF_SPEC>

Bit 0 - Set this bit to enable resolve char 0xC0 when DMA receiving data.

pub fn tx_db_esc_en(&mut self) -> TX_DB_ESC_EN_W<'_, ESCAPE_CONF_SPEC>

Bit 1 - Set this bit to enable resolve char 0xDB when DMA receiving data.

pub fn tx_11_esc_en(&mut self) -> TX_11_ESC_EN_W<'_, ESCAPE_CONF_SPEC>

Bit 2 - Set this bit to enable resolve flow control char 0x11 when DMA receiving data.

pub fn tx_13_esc_en(&mut self) -> TX_13_ESC_EN_W<'_, ESCAPE_CONF_SPEC>

Bit 3 - Set this bit to enable resolve flow control char 0x13 when DMA receiving data.

pub fn rx_c0_esc_en(&mut self) -> RX_C0_ESC_EN_W<'_, ESCAPE_CONF_SPEC>

Bit 4 - Set this bit to enable replacing 0xC0 with special char when DMA receiving data.

pub fn rx_db_esc_en(&mut self) -> RX_DB_ESC_EN_W<'_, ESCAPE_CONF_SPEC>

Bit 5 - Set this bit to enable replacing 0xDB with special char when DMA receiving data.

pub fn rx_11_esc_en(&mut self) -> RX_11_ESC_EN_W<'_, ESCAPE_CONF_SPEC>

Bit 6 - Set this bit to enable replacing 0x11 with special char when DMA receiving data.

pub fn rx_13_esc_en(&mut self) -> RX_13_ESC_EN_W<'_, ESCAPE_CONF_SPEC>

Bit 7 - Set this bit to enable replacing 0x13 with special char when DMA receiving data.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual