Type Alias esp32p4::bitscrambler::rx_ctrl::R

pub type R = R<RX_CTRL_SPEC>;

Register RX_CTRL reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn rx_ena(&self) -> RX_ENA_R

Bit 0 - write this bit to enable the bitscrambler rx

pub fn rx_pause(&self) -> RX_PAUSE_R

Bit 1 - write this bit to pause the bitscrambler rx core

pub fn rx_halt(&self) -> RX_HALT_R

Bit 2 - write this bit to halt the bitscrambler rx core

pub fn rx_eof_mode(&self) -> RX_EOF_MODE_R

Bit 3 - write this bit to ser the bitscrambler rx core EOF signal generating mode which is combined with reg_bitscrambler_rx_tailing_bits, 0: counter by read peripheral buffer, 0 counter by write dma fifo

pub fn rx_cond_mode(&self) -> RX_COND_MODE_R

Bit 4 - write this bit to specify the LOOP instruction condition mode of bitscrambler rx core, 0: use the little than operator to get the condition, 1: use not equal operator to get the condition

pub fn rx_fetch_mode(&self) -> RX_FETCH_MODE_R

Bit 5 - write this bit to set the bitscrambler rx core fetch instruction mode, 0: prefetch by reset, 1: fetch by instrutions

pub fn rx_halt_mode(&self) -> RX_HALT_MODE_R

Bit 6 - write this bit to set the bitscrambler rx core halt mode when rx_halt is set, 0: wait write data back done, , 1: ignore write data back

pub fn rx_rd_dummy(&self) -> RX_RD_DUMMY_R

Bit 7 - write this bit to set the bitscrambler rx core read data mode when EOF received.0: wait read data, 1: ignore read data