Module esp32p4::bitscrambler::tx_ctrl

Control and configuration registers

Structs

TX_CTRL_SPEC

Control and configuration registers

Type Aliases

R

Register TX_CTRL reader

TX_COND_MODE_R

Field TX_COND_MODE reader - write this bit to specify the LOOP instruction condition mode of bitscrambler tx core, 0: use the little than operator to get the condition, 1: use not equal operator to get the condition

TX_COND_MODE_W

Field TX_COND_MODE writer - write this bit to specify the LOOP instruction condition mode of bitscrambler tx core, 0: use the little than operator to get the condition, 1: use not equal operator to get the condition

TX_ENA_R

Field TX_ENA reader - write this bit to enable the bitscrambler tx

TX_ENA_W

Field TX_ENA writer - write this bit to enable the bitscrambler tx

TX_EOF_MODE_R

Field TX_EOF_MODE reader - write this bit to ser the bitscrambler tx core EOF signal generating mode which is combined with reg_bitscrambler_tx_tailing_bits, 0: counter by read dma fifo, 0 counter by write peripheral buffer

TX_EOF_MODE_W

Field TX_EOF_MODE writer - write this bit to ser the bitscrambler tx core EOF signal generating mode which is combined with reg_bitscrambler_tx_tailing_bits, 0: counter by read dma fifo, 0 counter by write peripheral buffer

TX_FETCH_MODE_R

Field TX_FETCH_MODE reader - write this bit to set the bitscrambler tx core fetch instruction mode, 0: prefetch by reset, 1: fetch by instrutions

TX_FETCH_MODE_W

Field TX_FETCH_MODE writer - write this bit to set the bitscrambler tx core fetch instruction mode, 0: prefetch by reset, 1: fetch by instrutions

TX_FIFO_RST_W

Field TX_FIFO_RST writer - write this bit to reset the bitscrambler tx fifo

TX_HALT_MODE_R

Field TX_HALT_MODE reader - write this bit to set the bitscrambler tx core halt mode when tx_halt is set, 0: wait write data back done, , 1: ignore write data back

TX_HALT_MODE_W

Field TX_HALT_MODE writer - write this bit to set the bitscrambler tx core halt mode when tx_halt is set, 0: wait write data back done, , 1: ignore write data back

TX_HALT_R

Field TX_HALT reader - write this bit to halt the bitscrambler tx core

TX_HALT_W

Field TX_HALT writer - write this bit to halt the bitscrambler tx core

TX_PAUSE_R

Field TX_PAUSE reader - write this bit to pause the bitscrambler tx core

TX_PAUSE_W

Field TX_PAUSE writer - write this bit to pause the bitscrambler tx core

TX_RD_DUMMY_R

Field TX_RD_DUMMY reader - write this bit to set the bitscrambler tx core read data mode when EOF received.0: wait read data, 1: ignore read data

TX_RD_DUMMY_W

Field TX_RD_DUMMY writer - write this bit to set the bitscrambler tx core read data mode when EOF received.0: wait read data, 1: ignore read data

W

Register TX_CTRL writer