Struct esp32p4::i3c_mst::RegisterBlock

#[repr(C)]
pub struct RegisterBlock { /* private fields */ }

Register block

Implementations

impl RegisterBlock

pub const fn device_ctrl(&self) -> &DEVICE_CTRL

0x00 - DEVICE_CTRL register controls the transfer properties and disposition of controllers capabilities.

pub const fn buffer_thld_ctrl(&self) -> &BUFFER_THLD_CTRL

0x1c - In-Band Interrupt Status Threshold Value . Every In Band Interrupt received by I3C controller generates an IBI status. This field controls the number of IBI status entries in the IBI buffer that trigger the IBI_STATUS_THLD_STAT interrupt.

pub const fn data_buffer_thld_ctrl(&self) -> &DATA_BUFFER_THLD_CTRL

0x20 - NA

pub const fn ibi_notify_ctrl(&self) -> &IBI_NOTIFY_CTRL

0x24 - NA

pub const fn ibi_sir_req_payload(&self) -> &IBI_SIR_REQ_PAYLOAD

0x28 - NA

pub const fn ibi_sir_req_reject(&self) -> &IBI_SIR_REQ_REJECT

0x2c - NA

pub const fn int_clr(&self) -> &INT_CLR

0x30 - NA

pub const fn int_raw(&self) -> &INT_RAW

0x34 - NA

pub const fn int_st(&self) -> &INT_ST

0x38 - NA

pub const fn int_st_ena(&self) -> &INT_ST_ENA

0x3c - The Interrupt status will be updated in INTR_STATUS register if corresponding Status Enable bit set.

pub const fn reset_ctrl(&self) -> &RESET_CTRL

0x44 - NA

pub const fn buffer_status_level(&self) -> &BUFFER_STATUS_LEVEL

0x48 - BUFFER_STATUS_LEVEL reflects the status level of Buffers in the controller.

pub const fn data_buffer_status_level(&self) -> &DATA_BUFFER_STATUS_LEVEL

0x4c - DATA_BUFFER_STATUS_LEVEL reflects the status level of the Buffers in the controller.

pub const fn present_state0(&self) -> &PRESENT_STATE0

0x50 - NA

pub const fn present_state1(&self) -> &PRESENT_STATE1

0x54 - NA

pub const fn device_table(&self) -> &DEVICE_TABLE

0x58 - Pointer for Device Address Table

pub const fn time_out_value(&self) -> &TIME_OUT_VALUE

0x5c - NA

pub const fn scl_i3c_mst_od_time(&self) -> &SCL_I3C_MST_OD_TIME

0x60 - NA

pub const fn scl_i3c_mst_pp_time(&self) -> &SCL_I3C_MST_PP_TIME

0x64 - NA

pub const fn scl_i2c_fm_time(&self) -> &SCL_I2C_FM_TIME

0x68 - NA

pub const fn scl_i2c_fmp_time(&self) -> &SCL_I2C_FMP_TIME

0x6c - NA

pub const fn scl_ext_low_time(&self) -> &SCL_EXT_LOW_TIME

0x70 - NA

pub const fn sda_sample_time(&self) -> &SDA_SAMPLE_TIME

0x74 - NA

pub const fn sda_hold_time(&self) -> &SDA_HOLD_TIME

0x78 - NA

pub const fn scl_start_hold(&self) -> &SCL_START_HOLD

0x7c - NA

pub const fn scl_rstart_setup(&self) -> &SCL_RSTART_SETUP

0x80 - NA

pub const fn scl_stop_hold(&self) -> &SCL_STOP_HOLD

0x84 - NA

pub const fn scl_stop_setup(&self) -> &SCL_STOP_SETUP

0x88 - NA

pub const fn bus_free_time(&self) -> &BUS_FREE_TIME

0x90 - NA

pub const fn scl_termn_t_ext_low_time(&self) -> &SCL_TERMN_T_EXT_LOW_TIME

0x94 - NA

pub const fn ver_id(&self) -> &VER_ID

0xa0 - NA

pub const fn ver_type(&self) -> &VER_TYPE

0xa4 - NA

pub const fn fpga_debug_probe(&self) -> &FPGA_DEBUG_PROBE

0xac - NA

pub const fn rnd_eco_cs(&self) -> &RND_ECO_CS

0xb0 - NA

pub const fn rnd_eco_low(&self) -> &RND_ECO_LOW

0xb4 - NA

pub const fn rnd_eco_high(&self) -> &RND_ECO_HIGH

0xb8 - NA